Move solver to separate file
