+++ src/data/levelSeeds.ts
@@ -3,5 +3,5 @@
  */
 export type SeedMap = Record<string, [seed: number, moves: number][]>;
 export const levelSeeds: SeedMap = JSON.parse(
-  '{"31976334":[[517311749,19],[359901870,22],[905403767,17],[899221565,19],[502420529,18],[865228431,21],[449626174,22],[632158867,20],[727952816,18],[60402820,20],[92967415,20],[474064842,21],[555354324,19],[518441848,22],[807000469,18],[943221663,20],[394416111,20],[148813978,21],[930002483,21],[220584264,18],[132060070,20],[806602223,20],[217421779,19],[555140744,22],[432897287,20],[167467368,20],[704725278,21],[812651450,19],[413999644,17],[888271664,21],[169253812,22],[489809536,20],[919745632,23],[828988207,23],[874527827,18],[269540869,22],[55691246,21],[720733021,21],[367860449,19],[481485622,22],[618809582,19],[864587310,20],[852005523,21],[166244857,21],[431171753,17],[373866676,19],[644642175,23],[801569483,24],[17552663,18],[599427656,21],[468175326,20],[556807265,16],[66790086,20],[328732305,18],[751363202,17],[849600546,20],[320090598,21],[33882977,16],[317073340,19],[183084739,21],[41191705,20],[785452374,20],[873347848,20],[702193759,20],[706620350,24],[777122960,22],[718392276,21],[489561862,22],[183090954,21],[112790229,21],[240856267,21],[456612460,17],[467510959,20],[331145315,21],[303148030,20],[221906716,19],[527271607,19],[979683981,21],[747439287,23],[306975097,18],[100686610,21],[107044363,17],[943165002,21],[53471920,21],[567308900,19],[421749944,20],[776865233,22],[736255507,19],[990486276,20],[234540976,22],[963453212,22],[937442019,18],[862293920,23],[326322243,19],[890488284,20],[471040617,24],[742661439,22],[116439555,21],[61006409,19],[936383014,19]],"65462911":[[779553376,43],[359901870,46],[905403774,49],[899221566,46],[632158868,47],[60402821,43],[474064843,45],[862301166,49],[353377833,48],[555354327,42],[394416114,45],[148813979,47],[220584269,47],[217421786,48],[614435426,48],[167467372,41],[812651456,49],[828988211,47],[420587419,47],[55691246,46],[720733021,45],[481485625,46],[852005523,42],[166244865,47],[431171753,46],[373866683,44],[644642180,43],[17552665,45],[896510679,47],[751363202,43],[41191714,45],[785452377,45],[183090958,50],[456612462,45],[427782219,43],[331145315,46],[747439292,47],[100686611,47],[53471922,46],[493355776,45],[286655128,45],[956556884,45],[375794656,49],[845724001,45],[206308936,48],[359255220,47],[705709741,47],[895521904,47],[274893941,41],[935386165,44],[467635272,46],[271118085,43],[175432469,46],[212243109,50],[876589606,45],[692375533,47],[35744365,46],[471116723,47],[866599604,51],[747978519,44],[971348278,44],[65010629,45],[848417029,45],[2799698,39],[303593325,46],[712191755,40],[29220768,48],[669952966,46],[471046174,47],[33070403,43],[230436279,48],[628673029,48],[549635989,44],[393212983,46],[660179143,45],[890898956,40],[428069860,45],[81432315,48],[12491695,45],[417977601,44],[593386838,46],[726982392,48],[330154711,43],[564052589,48],[891709515,48],[437491342,47],[752815986,48],[680377114,49],[840609601,46],[279479088,49],[647584689,48],[287892493,47],[344020955,51],[176087773,47],[227631814,44],[677951094,48],[30051300,44],[775941680,47],[525176575,41],[395208169,47]],"91451568":[[779553371,12],[359901871,13],[865228431,11],[449626175,11],[632158866,15],[727952819,14],[60402820,14],[92967416,15],[474064842,9],[862301169,14],[353377834,11],[555354326,13],[518441848,10],[807000472,16],[943221663,14],[394416114,14],[148813978,14],[930002487,9],[957535220,12],[220584267,12],[132060075,12],[217421781,14],[164924694,14],[614435430,14],[432897289,12],[167467375,13],[704725279,10],[812651451,12],[413999641,14],[986801717,10],[383963550,12],[489809536,19],[919745630,9],[828988207,14],[874527829,12],[420587419,13],[269540869,14],[406991365,14],[720733021,15],[367860448,13],[481485622,13],[852005524,11],[166244859,8],[431171754,13],[644642174,12],[599427658,11],[468175326,16],[556807265,12],[896510683,14],[66790087,12],[849600547,10],[320090595,12],[33882977,15],[530779521,16],[317073340,14],[183084739,12],[135483700,14],[873347852,15],[702193759,10],[706620349,12],[777122962,10],[489561863,14],[183090953,12],[112790231,9],[240856267,12],[849489344,10],[467510963,11],[331145321,10],[221906714,13],[979683980,16],[747439288,12],[306975097,14],[100686610,15],[107044363,13],[943165002,11],[53471920,11],[969121356,15],[666781849,12],[710346279,12],[862259528,16],[658907162,10],[909148578,10],[196278083,13],[265947629,11],[588949489,14],[860228488,9],[79582321,12],[23167076,12],[619367357,14],[31163027,21],[738834808,14],[688234239,13],[269555220,11],[441841729,14],[82193470,17],[324834154,11],[729801146,10],[890053904,11],[792191692,13],[641379344,13]],"219749905":[[517311754,25],[779553370,23],[359901876,24],[905403770,22],[899221565,22],[865228433,18],[632158868,24],[727952817,23],[60402823,24],[92967417,19],[555354324,23],[518441849,22],[807000469,23],[394416115,18],[220584264,19],[164924694,22],[555140748,22],[614435426,24],[167467371,24],[704725278,18],[812651450,20],[169253813,21],[919745627,21],[406991364,20],[720733021,21],[481485623,26],[864587312,19],[166244858,19],[431171756,20],[373866677,27],[644642174,24],[17552666,18],[599427656,20],[468175328,23],[556807265,21],[896510692,22],[66790087,21],[328732322,23],[751363202,22],[33882980,24],[530779525,23],[317073340,23],[183084741,23],[785452373,25],[135483701,22],[702193763,24],[706620350,24],[777122961,21],[718392283,22],[849489355,21],[427782220,25],[303148032,22],[527271606,23],[747439288,20],[306975105,19],[943165002,24],[805973626,26],[559900113,25],[82829059,20],[371051852,23],[884931362,22],[125884505,19],[203334952,20],[533219758,24],[5714782,25],[19929148,22],[818644476,22],[650902999,19],[944496963,22],[325037154,21],[458128230,23],[336891429,22],[202173528,25],[156844913,25],[548727181,24],[173793340,24],[421344995,22],[465111487,20],[465706432,21],[555708526,26],[962147098,22],[520589567,23],[187940970,16],[423611055,22],[266149082,22],[347138641,24],[657214793,23],[746467382,24],[885504965,22],[668115596,27],[937305649,20],[806024690,21],[970870576,18],[408129797,19],[207868292,25],[311209102,20],[100881709,22],[749051105,25],[213475316,23],[858171746,21]],"222816842":[[905403769,55],[862301169,48],[351041161,49],[217421779,54],[555140757,57],[704725281,52],[369922196,52],[383963553,50],[55691264,54],[367860455,53],[166244862,54],[801569481,51],[17552666,54],[599427673,55],[556807267,51],[328732306,55],[849600553,55],[135483705,50],[702193764,51],[784546905,54],[303148039,54],[527271623,50],[100686629,53],[943165021,55],[594915256,50],[998120312,55],[215807321,56],[926211065,52],[191531400,50],[548255473,53],[871075993,54],[905346971,52],[710346285,53],[306975113,53],[210712379,59],[59720321,55],[800796213,54],[9356700,46],[628271029,53],[417569857,54],[397637729,53],[671303126,52],[558978558,55],[515182601,50],[777662259,50],[965484389,53],[669224446,55],[290392761,53],[368153383,54],[884931365,49],[921867444,55],[37022994,50],[3307457,52],[119542847,53],[639338148,49],[965976001,58],[719478961,52],[506709653,49],[535454388,56],[135036039,48],[322460838,57],[262869511,55],[659581665,52],[665871535,55],[173452184,51],[932525110,53],[895964651,52],[600407087,55],[939063725,53],[892361979,48],[325957035,55],[748960869,55],[764921864,54],[311682411,54],[985534857,50],[186667289,55],[821738694,55],[788111866,51],[22637663,56],[767692425,56],[300285042,53],[786262767,52],[836380026,55],[394039211,47],[49475103,53],[984211073,52],[656375824,53],[175445352,54],[844663296,55],[407603386,58],[208310235,55],[474693338,54],[162393345,50],[337723847,54],[385626491,53],[436406994,51],[715509072,53],[522437084,56],[283756623,52],[476944440,58]],"226958187":[[517311749,13],[779553368,14],[359901870,14],[905403767,15],[899221565,14],[502420527,13],[865228431,15],[449626174,16],[632158866,15],[727952816,13],[60402820,13],[92967414,17],[474064842,14],[862301166,16],[353377832,12],[555354324,16],[518441848,11],[807000468,13],[943221661,14],[394416111,16],[148813978,14],[930002483,13],[957535220,15],[220584264,11],[132060070,15],[806602223,14],[217421779,17],[164924693,15],[555140744,18],[614435426,13],[432897288,13],[167467368,15],[704725277,14],[812651449,12],[413999641,12],[986801716,16],[888271664,15],[383963550,14],[169253812,15],[489809536,17],[919745627,15],[828988207,16],[874527827,13],[420587416,14],[269540869,16],[406991364,14],[55691246,16],[720733021,13],[367860448,15],[481485622,12],[618809582,15],[864587310,14],[852005523,15],[166244857,16],[431171752,15],[373866675,15],[644642173,16],[801569480,12],[17552661,15],[599427656,15],[468175326,12],[556807265,17],[896510679,16],[66790086,15],[328732304,16],[751363202,13],[849600546,13],[320090595,14],[33882977,13],[530779519,16],[317073340,16],[183084738,15],[41191704,16],[785452373,15],[135483698,11],[873347847,14],[702193759,15],[706620349,16],[777122960,14],[718392276,13],[489561862,15],[183090953,13],[112790229,13],[240856267,15],[849489344,15],[456612460,13],[467510959,14],[427782219,14],[331145315,14],[303148030,17],[221906714,13],[527271606,14],[979683980,15],[747439287,11],[306975095,16],[100686610,15],[107044364,14],[943165002,13],[53471920,15],[335179851,14]],"239814608":[[517311750,6],[779553369,8],[359901872,11],[905403768,9],[899221566,6],[502420538,7],[865228433,7],[449626174,11],[632158870,10],[727952828,5],[60402820,8],[92967419,8],[474064855,11],[862301166,10],[353377837,7],[555354325,6],[518441853,9],[807000471,10],[943221662,8],[394416113,8],[148813978,6],[930002486,7],[957535224,6],[220584268,6],[132060071,5],[806602223,9],[217421783,7],[164924696,9],[555140758,10],[614435427,8],[432897307,11],[167467371,7],[704725285,10],[812651450,10],[413999643,6],[986801719,6],[888271665,6],[169253814,5],[489809539,11],[919745631,8],[828988213,7],[420587420,7],[406991364,11],[55691248,7],[720733026,7],[367860448,7],[481485623,10],[618809582,4],[864587320,6],[852005529,7],[166244871,6],[431171754,9],[373866679,5],[644642179,8],[801569480,6],[17552663,5],[599427657,10],[468175326,8],[556807266,6],[896510679,10],[66790086,8],[789548694,8],[328732314,9],[751363205,5],[849600546,9],[320090596,9],[33882977,7],[530779519,5],[317073344,8],[183084740,10],[41191704,8],[785452375,10],[135483699,7],[702193772,8],[706620350,7],[777122970,6],[718392280,8],[489561866,8],[183090968,9],[112790235,8],[240856271,9],[849489346,8],[456612473,8],[467510963,7],[427782219,8],[331145316,7],[303148035,10],[221906715,10],[527271612,6],[979683984,6],[747439290,5],[306975102,9],[100686611,9],[107044376,7],[943165007,7],[53471925,8],[112005979,9],[775446497,7],[809567040,6],[508431958,8]],"240865250":[[517311750,24],[359901870,23],[905403768,26],[899221566,24],[865228432,26],[449626174,25],[632158866,23],[727952816,24],[60402826,23],[92967414,23],[474064844,26],[862301166,25],[353377833,24],[555354324,25],[518441850,23],[943221663,23],[394416111,25],[148813978,27],[930002483,26],[957535220,23],[220584264,26],[132060071,26],[806602224,25],[217421779,25],[555140745,24],[614435426,26],[167467368,26],[812651449,22],[413999641,26],[888271664,24],[383963551,24],[828988207,25],[269540870,25],[406991365,25],[720733021,24],[481485623,28],[864587310,24],[852005527,23],[166244857,24],[373866675,23],[801569481,24],[17552666,24],[599427656,25],[556807265,25],[896510681,23],[66790086,25],[789548691,25],[751363202,24],[849600547,23],[33882977,23],[530779519,26],[183084738,23],[785452374,23],[135483698,23],[702193761,24],[718392277,21],[489561864,26],[183090955,26],[456612460,29],[467510960,22],[427782220,25],[331145317,27],[221906714,26],[747439290,28],[100686618,20],[943165002,25],[53471921,24],[301674070,25],[706640133,27],[866599604,25],[95896677,28],[359950547,27],[223080981,25],[79582322,24],[49216690,23],[366245939,25],[458128229,27],[242487866,27],[34234637,26],[899069763,25],[938108730,25],[706384425,24],[948226112,27],[511965087,25],[454768124,26],[912469055,22],[705321270,26],[125905687,24],[600538391,25],[614085913,25],[551732586,26],[101372962,25],[321946448,25],[625206644,27],[491418665,26],[734251823,25],[445093203,21],[409813391,23],[549973349,23],[321824715,24]],"304412621":[[517311749,17],[779553368,17],[359901871,15],[905403767,19],[899221565,20],[502420527,20],[865228431,16],[632158866,20],[727952816,17],[60402821,17],[92967414,19],[474064842,19],[862301166,17],[353377831,15],[555354324,18],[518441848,17],[807000469,20],[943221661,17],[148813978,18],[930002483,16],[957535220,20],[220584265,19],[806602223,18],[217421781,18],[614435426,18],[432897287,18],[167467369,20],[704725277,17],[413999641,18],[986801716,19],[888271664,18],[383963550,18],[169253813,19],[489809536,17],[919745627,18],[828988210,14],[874527827,17],[420587416,16],[269540869,20],[406991364,19],[55691246,15],[720733023,18],[367860448,18],[481485622,18],[618809583,20],[864587310,21],[852005523,21],[166244857,19],[431171752,20],[373866675,19],[644642173,18],[801569480,21],[17552662,16],[599427658,19],[468175328,20],[556807266,17],[66790087,19],[789548690,16],[328732305,18],[751363202,18],[849600546,20],[320090595,18],[33882979,19],[530779519,19],[317073340,20],[183084738,18],[41191704,17],[785452373,17],[135483697,17],[873347847,17],[777122960,20],[489561862,22],[183090954,19],[112790229,18],[240856267,18],[849489345,17],[456612460,17],[427782221,19],[331145315,17],[303148030,17],[221906714,19],[527271606,19],[979683981,17],[747439287,19],[306975095,19],[100686610,17],[107044363,17],[53471920,20],[643747234,19],[500060559,17],[991124157,16],[42658434,18],[779954357,20],[305269964,18],[735246590,18],[243288428,21],[95432311,17],[866599603,18],[654402732,20],[20367593,21]],"305556867":[[517311749,17],[779553368,19],[359901870,19],[905403767,20],[899221565,18],[502420527,21],[865228431,18],[632158866,21],[727952816,18],[60402820,19],[92967414,20],[474064842,20],[862301166,21],[353377831,19],[555354324,21],[518441848,16],[807000467,20],[943221662,18],[394416111,19],[148813978,18],[957535220,21],[220584264,20],[132060070,21],[806602223,18],[217421779,21],[164924693,19],[555140744,21],[614435426,18],[432897287,16],[167467368,20],[704725277,22],[812651449,21],[413999641,19],[986801716,22],[888271664,17],[383963550,18],[169253812,20],[489809536,20],[919745627,16],[828988209,17],[874527827,18],[420587416,18],[269540869,19],[406991364,19],[55691246,15],[720733021,19],[367860448,22],[481485622,19],[618809582,19],[864587311,19],[852005523,19],[431171752,21],[373866675,18],[644642173,18],[801569480,22],[17552661,17],[599427656,21],[468175326,19],[556807265,18],[896510680,19],[66790086,20],[789548689,20],[328732305,17],[751363203,20],[849600546,21],[320090595,17],[33882977,16],[317073340,24],[183084738,25],[41191704,19],[785452373,20],[135483697,19],[873347848,22],[702193759,21],[706620349,22],[777122960,21],[718392276,19],[489561862,21],[183090953,20],[240856267,20],[849489344,21],[456612460,19],[467510959,21],[427782219,21],[331145315,18],[303148030,21],[221906714,20],[527271606,17],[979683980,20],[747439287,20],[306975095,18],[100686610,19],[107044363,16],[943165002,19],[53471920,19],[405886192,17],[208063620,19],[30519807,20],[594078541,20],[103481204,18]],"363887855":[[517311750,6],[779553369,8],[359901871,10],[905403767,7],[865228431,6],[449626174,11],[632158869,11],[727952816,9],[60402820,7],[92967414,7],[474064842,10],[862301166,10],[353377833,9],[555354325,8],[518441848,7],[807000469,9],[943221661,9],[394416112,10],[148813978,7],[930002483,8],[957535220,9],[220584264,11],[132060070,10],[806602223,9],[217421782,9],[164924693,8],[555140744,12],[614435426,11],[432897287,8],[167467369,8],[704725280,8],[812651449,9],[413999641,9],[986801717,7],[888271664,9],[169253814,5],[489809536,8],[828988209,8],[874527828,6],[420587416,7],[406991364,11],[55691246,7],[720733021,9],[367860448,9],[481485622,7],[618809582,4],[864587310,8],[166244857,8],[431171752,8],[373866675,6],[644642173,8],[801569480,6],[17552662,10],[599427657,10],[468175326,8],[556807266,7],[896510679,10],[66790086,8],[328732305,10],[751363202,6],[320090596,9],[33882977,7],[530779519,5],[317073341,11],[41191704,7],[785452373,11],[135483697,8],[873347847,9],[702193759,7],[706620349,13],[777122960,14],[489561862,9],[183090956,7],[112790230,9],[456612461,11],[467510960,13],[427782222,6],[331145316,7],[303148035,10],[221906715,10],[527271612,7],[979683980,11],[747439289,7],[306975095,8],[100686610,11],[107044364,7],[943165002,10],[53471921,11],[71107439,8],[916161507,11],[21945869,8],[376231936,10],[721063485,8],[35595057,8],[416755066,10],[340900290,10],[800796210,11],[543808335,9],[610470729,7],[888893344,9]],"374972231":[[779553377,63],[905403774,65],[458128243,59],[738545686,65],[469250498,64],[120329970,59],[353377836,59],[555354336,60],[807000481,65],[962310698,62],[895395975,60],[251007671,62],[209086746,62],[703212775,60],[986801729,63],[812886551,62],[655650051,59],[441244713,62],[874527831,64],[420587428,62],[269540874,62],[685400728,61],[720733022,60],[481485638,62],[815581696,61],[644642181,64],[53980371,61],[666080563,63],[468175335,60],[556807267,66],[815709796,66],[435327457,62],[789548692,62],[46412911,60],[320090613,65],[33882996,61],[530779522,60],[183084744,59],[542813250,60],[873347847,58],[706620354,59],[255990100,58],[183090960,57],[456612463,65],[332067092,61],[221906725,66],[225923201,63],[979683991,58],[799372887,64],[226033155,67],[107044364,54],[241199285,58],[447369006,61],[54169439,62],[943972963,59],[604636360,63],[223421024,58],[503224317,62],[838387769,59],[473213756,61],[895082804,61],[541916032,60],[760623388,65],[892048661,61],[280423076,63],[478034690,65],[642243325,66],[379514965,61],[974657845,59],[500352402,63],[819009261,59],[42293115,62],[726392925,62],[608953624,64],[362904866,60],[864605982,62],[970934049,63],[66727715,62],[285878051,60],[151519476,64],[157753813,64],[284587699,63],[961836643,63],[858063318,61],[46745836,60],[327335749,62],[91099226,59],[555641019,63],[966780947,63],[845221767,65],[536791548,61],[239841998,58],[639672258,66],[861482978,58],[32288721,61],[608843598,62],[638307819,63],[460971467,57],[539373476,62],[212369408,62]],"494781908":[[779553371,63],[359901879,90],[899221565,100],[865228434,74],[449626182,101],[727952819,66],[60402824,77],[474064856,79],[148813985,74],[957535225,94],[627853239,141],[217421786,63],[986801720,76],[919745628,83],[420587420,58],[55691256,72],[720733022,82],[618809583,114],[852005534,69],[801569480,83],[17552661,76],[328732308,138],[751363210,85],[317073343,69],[183084743,107],[873347851,60],[489561872,79],[456612470,94],[303148031,94],[306975098,109],[100686610,58],[53471920,64],[615813431,80],[823363678,74],[880626252,60],[267930983,78],[827675738,69],[840745950,99],[5194965,77],[339517716,70],[449849673,94],[79619503,73],[890592470,110],[268365465,104],[680538360,76],[390042101,79],[481918044,81],[966394410,95],[179470473,91],[562489761,107],[288056135,130],[25085993,90],[768464201,93],[203474104,137],[2605514,98],[262728522,70],[939465647,83],[585690726,106],[999786411,116],[936489495,90],[739025986,59],[714375259,103],[353745231,76],[127921611,90],[848172387,120],[724927200,95],[965492895,90],[428069863,68],[59824355,87],[300177011,88],[606767488,82],[138826313,79],[133009131,75],[496895529,67],[183161573,86],[717986924,79],[643386304,84],[549572716,108],[753567094,77],[217117950,106],[580192953,82],[525255461,85],[16398980,67],[82736248,96],[971534157,107],[386476465,82],[337534327,87],[934027695,73],[678024427,64],[837628145,79],[412187125,112],[504736587,105],[676692723,75],[784239055,73],[752112484,79],[160120994,103],[432183331,77],[915042488,84],[591926650,65],[830714662,129]],"512250895":[[517311749,4],[779553368,5],[359901872,4],[905403769,5],[899221565,5],[502420528,4],[865228434,7],[449626177,4],[632158867,5],[60402821,3],[92967414,4],[474064842,7],[862301166,6],[353377837,4],[555354324,4],[518441848,4],[807000467,7],[943221661,5],[394416112,4],[148813978,4],[930002484,3],[957535223,3],[220584264,5],[132060071,4],[806602223,4],[217421779,6],[164924693,6],[555140744,5],[614435426,3],[432897288,3],[167467369,7],[704725279,5],[812651449,4],[413999645,4],[986801717,3],[888271664,5],[383963554,3],[169253813,4],[489809536,4],[919745627,5],[828988207,4],[874527827,3],[420587417,4],[269540871,5],[406991364,3],[55691246,4],[720733022,5],[367860448,3],[481485624,4],[618809582,4],[852005523,7],[166244863,4],[431171753,4],[373866675,3],[644642173,4],[801569481,6],[17552663,4],[599427658,3],[468175328,3],[896510679,7],[66790087,7],[789548689,3],[328732305,4],[751363203,4],[849600546,6],[320090595,7],[33882977,5],[530779520,3],[317073340,6],[183084739,4],[41191707,4],[785452376,5],[135483697,3],[873347849,4],[702193759,7],[706620350,5],[777122960,5],[718392279,6],[489561862,6],[183090954,3],[112790229,4],[240856269,5],[849489344,4],[456612460,3],[467510959,5],[427782219,4],[331145315,4],[303148030,7],[221906715,4],[527271608,5],[979683980,4],[747439287,4],[306975095,4],[100686610,3],[107044363,5],[943165003,7],[53471920,3],[993223207,3],[401741359,4],[842437860,4]],"532916128":[[779553369,34],[905403767,31],[502420528,36],[449626174,35],[727952816,34],[60402820,36],[92967414,32],[474064844,34],[353377831,34],[555354325,34],[807000468,36],[943221661,36],[394416111,37],[930002483,35],[957535222,34],[220584265,36],[132060070,37],[806602225,34],[217421779,34],[164924693,34],[432897290,35],[704725278,35],[812651449,33],[413999641,35],[986801720,33],[888271666,33],[383963553,33],[169253813,36],[919745629,33],[828988210,33],[874527827,34],[269540869,34],[406991365,35],[55691248,33],[720733021,35],[367860448,35],[618809583,36],[864587311,35],[852005524,37],[166244857,35],[373866676,31],[801569480,38],[17552661,28],[599427656,35],[468175326,34],[556807265,30],[789548689,33],[328732305,34],[751363202,33],[849600546,39],[320090596,33],[33882978,31],[530779519,32],[317073340,36],[183084738,38],[41191705,34],[785452373,33],[873347847,36],[777122960,36],[718392277,32],[489561862,34],[112790231,35],[240856267,33],[456612461,36],[467510960,34],[221906718,31],[100686611,36],[107044364,34],[53471920,36],[76236421,33],[969614397,36],[627585821,34],[77741398,35],[547726857,35],[877979090,35],[789450155,35],[66727700,34],[526876597,32],[258718442,35],[88551611,33],[153211493,32],[325227388,34],[13254489,35],[482070066,35],[278093623,34],[266784548,33],[720788590,35],[494195624,35],[409631100,33],[775446497,35],[283072971,32],[876956447,35],[112451665,33],[696945059,33],[166591232,34],[930551513,37],[926539590,32],[817389292,35],[62676673,34],[482759852,30]],"598143099":[[517311751,18],[779553368,15],[359901871,16],[905403767,18],[899221566,18],[502420527,21],[865228433,19],[449626175,18],[632158866,17],[727952817,22],[60402820,19],[92967414,18],[474064842,16],[862301167,15],[353377831,18],[555354324,17],[518441848,19],[807000467,18],[943221661,18],[394416111,20],[148813979,18],[930002483,14],[957535220,20],[220584265,18],[132060070,22],[806602223,18],[217421779,22],[164924693,20],[555140744,19],[614435427,14],[432897287,18],[167467368,16],[704725277,17],[812651453,20],[413999641,20],[986801717,17],[888271665,20],[383963550,20],[169253812,20],[489809536,19],[919745627,16],[828988207,17],[874527828,18],[420587416,17],[269540869,20],[406991364,21],[55691246,21],[720733022,16],[367860448,19],[481485622,15],[618809582,21],[864587310,17],[852005523,20],[166244857,20],[431171752,18],[373866675,16],[644642173,19],[801569480,21],[17552661,19],[599427656,21],[468175327,16],[556807266,16],[896510679,16],[66790086,19],[789548689,15],[328732304,18],[751363202,18],[849600546,19],[320090597,18],[33882979,15],[530779519,22],[317073340,20],[183084738,18],[41191704,14],[785452374,15],[135483698,16],[873347847,17],[702193760,20],[706620349,21],[777122961,17],[718392276,16],[489561862,22],[183090955,19],[112790229,20],[240856267,18],[849489344,17],[456612460,21],[467510959,20],[427782219,16],[331145315,17],[303148031,19],[221906714,17],[527271606,17],[979683981,14],[747439288,17],[306975096,18],[100686610,19],[107044363,19],[943165002,16],[53471920,18]],"636324142":[[517311749,4],[779553368,5],[359901870,5],[905403767,4],[899221565,6],[502420527,4],[865228431,6],[449626174,4],[632158867,5],[727952816,4],[60402820,7],[92967414,4],[474064842,7],[862301166,6],[353377832,5],[555354324,4],[518441848,4],[807000467,7],[943221661,5],[394416112,4],[148813978,4],[930002484,3],[957535220,4],[220584264,7],[132060071,4],[806602223,4],[217421779,7],[164924693,6],[555140744,5],[614435426,3],[432897287,6],[167467368,4],[704725277,5],[812651449,4],[413999641,6],[986801716,6],[888271664,5],[383963551,5],[169253812,4],[489809536,4],[919745627,5],[828988207,4],[874527827,3],[269540869,6],[406991364,3],[55691246,4],[720733021,8],[367860448,3],[481485622,4],[618809582,4],[864587310,6],[852005523,7],[166244857,6],[431171752,5],[373866675,3],[644642173,4],[801569480,4],[17552663,4],[599427657,9],[468175327,7],[556807266,6],[896510679,7],[66790086,4],[789548689,3],[328732304,4],[751363202,6],[849600546,6],[320090595,6],[33882977,6],[530779519,4],[317073340,7],[183084739,4],[41191704,6],[785452373,9],[135483697,3],[873347847,6],[702193759,6],[706620349,7],[777122960,5],[718392276,4],[489561862,6],[183090954,3],[112790229,4],[240856267,7],[849489344,4],[456612460,3],[467510959,5],[427782219,4],[331145315,4],[303148030,7],[221906714,6],[527271606,6],[979683980,4],[747439287,4],[306975095,4],[100686610,3],[107044363,5],[943165002,6],[53471920,3],[644859537,5]],"661504400":[[779553382,20],[359901876,24],[905403773,24],[899221577,24],[449626177,25],[632158867,23],[727952817,23],[521573372,22],[353377831,23],[555354339,23],[518441854,23],[807000477,26],[148813980,21],[957535224,21],[806602238,20],[299800229,24],[555140759,23],[614435427,20],[812651451,19],[413999646,22],[986801727,21],[383963552,21],[169253813,21],[489809536,21],[493112077,26],[828988219,23],[367860459,20],[618809594,23],[864587321,23],[801569483,21],[17552680,18],[599427656,20],[468175329,23],[556807265,19],[789548706,17],[751363202,23],[849600550,27],[320090601,26],[33882981,22],[702193763,24],[706620352,19],[718392285,22],[754031611,21],[427782238,19],[121801053,20],[303148035,23],[953023070,22],[882592124,22],[719812831,22],[622482417,24],[200618604,25],[555425550,21],[602354185,22],[336068240,22],[410514377,21],[899867591,22],[437491350,19],[581338027,25],[130861286,24],[604636354,24],[171928595,25],[454901889,26],[944823866,19],[120329985,20],[613612098,24],[140073126,26],[705797191,26],[317000445,24],[395624470,20],[787598935,23],[541795468,24],[739025987,24],[340691029,15],[522483102,24],[217008244,22],[588456911,23],[887779083,23],[982940617,23],[267482588,22],[792191694,20],[198914252,24],[662979154,24],[919322780,24],[194182701,22],[563965677,23],[585757025,14],[942995714,22],[340287741,23],[867093745,26],[784369148,21],[116240515,21],[57270765,24],[825513622,19],[712958440,22],[173421454,20],[896688588,20],[724927229,19],[639338153,26],[885504981,22],[151703714,19]],"684698259":[[215107856,30],[874890925,40],[15144563,27],[321449692,29],[767931295,27],[285204707,36],[569127029,42],[369070372,28],[928309510,30],[299269846,32],[864097863,32],[211354790,29],[409475649,25],[308976570,26],[747884705,23],[117120741,26],[594682794,33],[986801732,32],[11212415,31],[441244721,27],[455902176,38],[541338576,29],[406991374,37],[506709660,21],[417677359,32],[206210373,35],[442175641,34],[931790505,31],[422316083,34],[672077511,34],[612352458,38],[391931204,31],[777122969,26],[489561871,32],[494051496,42],[361737511,26],[208076200,29],[888122969,29],[769028152,25],[943165006,36],[127921611,35],[280869691,25],[721765,35],[170629443,30],[159420443,37],[93332416,27],[429786675,29],[901439199,27],[656043063,30],[305160306,34],[155767384,47],[230370925,44],[663053482,35],[268265801,30],[922679552,36],[214052793,36],[356475707,34],[857820464,40],[893897231,33],[39507825,29],[124187392,30],[648466910,37],[275788789,29],[674301997,24],[367939803,36],[395034521,37],[369165992,34],[515282470,31],[708763207,29],[586393264,33],[721304344,26],[300354990,29],[343127308,30],[71906975,31],[42658443,25],[953502943,32],[886921011,27],[442416900,36],[352254189,31],[988740343,31],[598632749,26],[466462693,31],[464489827,40],[125934226,40],[548122984,31],[931770622,36],[718963563,27],[111232209,35],[654618582,34],[461875420,32],[644789906,28],[560678964,28],[191185890,25],[937411389,27],[23783162,31],[741001013,28],[222401890,36],[744730609,36],[52019592,30],[471046187,28]],"718288442":[[517311749,17],[779553368,19],[359901870,18],[905403767,21],[899221565,18],[502420527,21],[865228431,18],[449626174,19],[632158866,21],[727952816,18],[60402820,19],[92967415,18],[474064842,19],[862301166,20],[353377831,17],[555354324,20],[518441848,14],[807000467,20],[943221662,18],[394416111,19],[148813978,17],[930002483,20],[957535220,21],[220584264,19],[132060071,20],[806602223,17],[217421779,19],[164924693,19],[555140744,18],[614435426,18],[432897287,16],[167467368,19],[704725277,20],[812651449,20],[413999641,18],[986801716,22],[888271664,16],[383963550,18],[169253812,20],[489809536,18],[919745627,16],[828988209,16],[874527827,17],[420587416,18],[269540869,19],[406991364,18],[55691246,15],[720733021,18],[367860448,22],[481485622,17],[618809582,19],[864587311,19],[852005523,17],[166244857,19],[431171752,19],[373866675,17],[644642173,18],[801569480,21],[17552661,15],[599427656,21],[468175326,16],[556807265,18],[896510680,18],[66790086,21],[789548689,19],[328732305,17],[751363203,20],[849600546,21],[320090595,16],[33882977,15],[530779519,21],[317073340,24],[183084738,25],[41191704,19],[785452373,20],[135483697,19],[873347847,19],[702193759,21],[706620349,22],[777122960,21],[718392276,17],[489561862,21],[183090953,20],[112790229,21],[240856267,20],[849489344,18],[456612460,19],[467510959,18],[427782219,19],[331145315,18],[303148030,21],[221906714,18],[527271606,16],[979683980,20],[747439287,17],[306975095,18],[100686610,18],[107044363,16],[943165002,19],[53471920,19]],"784481713":[[444588742,32],[426848623,38],[717672653,44],[370072895,40],[754652487,37],[733973323,38],[407474129,41],[973084493,36],[251731131,40],[452036023,45],[489809537,37],[269540869,32],[720733021,36],[367860449,31],[864587320,42],[852005533,36],[431171752,36],[66790086,40],[41191707,33],[873347850,39],[702193762,35],[718392277,34],[112790234,39],[240856270,35],[849489351,37],[527271608,34],[979683980,46],[306975096,37],[100686615,42],[151703706,36],[569521155,37],[866045478,39],[854498838,42],[330302319,35],[883892330,43],[383650849,49],[844663282,40],[271063516,39],[556271136,34],[788105733,33],[678681868,38],[316093570,29],[476558766,40],[442024482,41],[316188663,36],[529873334,44],[108163818,39],[44336560,51],[52925288,38],[424749135,40],[600047391,37],[910887089,45],[872783118,46],[567776389,35],[604152515,30],[793498911,44],[722431459,33],[126070018,41],[845201,38],[688798942,35],[294557189,35],[519361997,48],[198633822,39],[576938488,32],[86184707,32],[113973705,44],[502169118,38],[506252183,40],[667558086,35],[38178224,39],[194990487,38],[515175282,35],[263487889,40],[653460192,37],[81022692,36],[398431606,37],[515282457,35],[982893169,38],[332716030,32],[808107531,41],[667788774,39],[750827263,39],[126092915,38],[867541608,40],[556748464,47],[469636567,36],[371963542,44],[839020552,37],[431376531,38],[224683297,50],[611315822,38],[995417590,56],[535496757,35],[525255474,32],[787365318,42],[18937237,43],[590067929,41],[881036790,36],[534119098,34],[229205302,39]],"849132023":[[517311751,65],[359901870,72],[865228433,96],[632158867,71],[862301168,88],[943221661,77],[148813983,65],[164924694,64],[986801721,63],[406991364,57],[55691246,61],[618809582,72],[864587311,81],[852005529,54],[431171755,77],[373866675,58],[644642176,55],[328732305,75],[530779523,64],[183084738,71],[41191706,87],[135483700,52],[873347847,64],[702193762,55],[706620349,66],[777122960,65],[718392282,62],[489561862,63],[183090954,77],[112790231,67],[240856268,67],[456612461,81],[427782224,74],[331145317,62],[303148031,58],[979683982,85],[107044368,73],[992782212,65],[102084982,49],[920632345,64],[464884290,83],[963096032,88],[756235789,66],[868627370,74],[243890960,71],[758576727,74],[928515491,68],[764572418,66],[829504763,92],[331363144,72],[340900292,53],[299875116,70],[708331027,69],[889662352,79],[557969577,58],[10150482,58],[860032800,67],[754031610,69],[562737898,71],[657705590,64],[428096734,71],[3679607,90],[47584521,68],[295120906,75],[392147427,67],[762296235,64],[382223588,74],[886642647,83],[873405291,73],[401137995,66],[368332158,75],[317856290,92],[669251627,61],[217063481,74],[328043781,78],[857570388,76],[179197658,68],[375189226,63],[599129177,59],[844919951,76],[848172385,63],[702242421,79],[966822911,67],[126492333,54],[135175563,59],[128684669,79],[896931745,70],[946160958,53],[889391931,67],[869355856,59],[204067060,75],[907438053,65],[569915788,80],[331851285,81],[807510859,69],[913145571,72],[384863429,61],[347445332,69],[378145294,70],[182249841,69]],"878487799":[[517311750,6],[779553368,8],[359901870,6],[905403767,7],[899221565,9],[502420527,8],[865228431,5],[449626174,11],[632158866,8],[727952816,8],[60402820,8],[92967414,7],[474064842,9],[862301166,11],[353377831,9],[555354324,7],[518441848,6],[807000467,10],[943221661,9],[394416111,7],[148813978,7],[930002483,8],[957535220,9],[220584264,10],[132060070,9],[806602223,10],[217421779,8],[164924693,7],[555140744,11],[614435426,11],[432897287,7],[167467368,6],[704725277,10],[812651449,9],[413999641,9],[986801716,9],[888271664,10],[383963550,9],[169253812,10],[489809536,7],[919745627,10],[828988207,9],[874527828,7],[420587416,7],[269540869,11],[406991364,11],[55691246,7],[720733021,7],[367860448,8],[481485622,7],[618809582,4],[864587310,7],[852005523,9],[166244857,8],[431171752,8],[373866675,6],[644642173,8],[801569480,7],[17552662,10],[599427656,9],[468175326,8],[556807265,8],[896510679,11],[66790086,9],[789548689,10],[328732305,10],[751363202,5],[849600546,10],[320090595,10],[33882977,7],[530779519,5],[317073340,10],[183084738,10],[41191704,9],[785452373,11],[135483697,8],[873347847,7],[702193759,7],[706620349,10],[777122960,11],[718392276,9],[489561862,8],[183090953,9],[112790229,7],[240856267,11],[849489344,8],[456612460,8],[467510959,8],[427782219,8],[331145315,8],[303148030,10],[221906714,6],[527271606,12],[979683980,9],[747439287,8],[306975095,6],[100686610,10],[107044363,7],[943165002,9],[53471920,6]],"880938594":[[67514109,57],[779553377,63],[738545686,64],[469250498,64],[727952818,60],[788920159,60],[521573371,58],[353377836,58],[163364822,60],[807000481,65],[674990297,57],[167763230,62],[597209688,59],[251007667,58],[564310258,58],[209086746,62],[810302254,56],[151841915,60],[493112082,60],[595642275,62],[269540874,62],[685400728,59],[720733022,60],[175440176,60],[166244875,60],[431376529,64],[644642181,61],[468175335,60],[68469099,56],[462220619,57],[33882996,59],[985521853,58],[183084744,58],[41191706,63],[612352447,63],[768394376,62],[134829885,61],[456612463,65],[332067092,61],[583261546,60],[979683991,58],[747439287,61],[107044364,54],[53471934,62],[85473184,64],[923016651,61],[389434688,58],[15204743,61],[534730807,61],[701874348,59],[411913944,64],[93319944,59],[534872969,63],[581653012,66],[590815931,58],[894693233,62],[579038083,60],[416427426,63],[298652071,64],[509977847,64],[979905934,59],[696099136,61],[610695565,60],[877341477,59],[256558690,63],[244481742,56],[697558197,61],[391252651,63],[678279918,62],[287757023,60],[483767088,59],[240623837,62],[627794408,57],[215849667,60],[890117834,61],[406926795,65],[188757651,64],[362047812,62],[67668854,61],[613909967,61],[23400685,63],[267482571,61],[995156146,59],[395954211,58],[514782063,59],[819360392,60],[537355934,62],[889567427,65],[425668750,60],[961910530,61],[919501962,64],[19133214,64],[801385684,56],[980885838,61],[680766829,58],[20226450,66],[830480173,59],[343288763,59],[239841998,58],[816286873,61]],"908554960":[[517311751,38],[779553369,39],[359901870,48],[905403769,41],[899221567,38],[502420527,41],[449626174,39],[727952816,40],[92967414,33],[862301166,38],[555354324,37],[518441848,40],[807000469,36],[394416112,42],[148813986,35],[930002484,43],[220584265,37],[806602225,37],[217421784,34],[164924693,49],[555140746,41],[614435428,43],[432897287,40],[986801718,41],[489809536,41],[828988210,37],[269540869,32],[406991365,43],[55691246,38],[367860449,32],[852005525,40],[166244858,45],[801569480,31],[599427656,38],[468175326,35],[556807265,30],[896510679,37],[66790086,39],[789548689,40],[751363209,40],[320090595,34],[33882977,39],[183084739,47],[135483697,39],[702193759,40],[706620349,36],[718392277,34],[183090958,38],[112790233,40],[456612461,44],[427782219,38],[331145317,37],[303148031,41],[221906720,36],[527271607,33],[100686612,44],[107044363,49],[53471921,45],[330554621,39],[223036990,40],[123989685,41],[153724245,33],[929624958,42],[130910677,47],[454424584,45],[169760354,42],[20231231,38],[71733674,38],[124840228,41],[294123364,41],[391689748,41],[99637606,36],[699492136,35],[472532522,43],[785890281,38],[86922177,36],[412010851,39],[906075849,41],[842168691,41],[115511722,45],[65010628,38],[479926502,32],[598466087,40],[259330184,40],[930397852,40],[730574185,35],[2094708,38],[858569096,39],[10120574,44],[174541558,37],[71009783,44],[728636604,38],[102581850,43],[904188920,38],[82063573,47],[225128701,33],[486206376,41],[219512397,40],[739979742,37],[772566115,38]],"932844753":[[517311750,41],[449626177,36],[632158873,32],[60402822,33],[862301166,30],[555354324,32],[518441850,26],[943221666,32],[957535228,35],[220584266,32],[132060079,32],[164924696,34],[555140746,39],[614435434,34],[253945709,38],[888271677,29],[383963552,34],[828988211,35],[420587422,30],[406991369,41],[55691255,30],[367860448,31],[373866689,29],[17552661,29],[599427657,29],[789548691,38],[328732313,37],[33882978,34],[41191706,31],[702193762,35],[112790230,32],[331145315,33],[53471920,40],[901439194,40],[350093853,32],[901244076,35],[552306138,35],[582582270,29],[324834155,30],[554778654,32],[65059018,36],[934706545,37],[776723768,41],[820703947,36],[951358825,32],[614283731,39],[942359317,30],[950071294,33],[482716293,45],[308430142,33],[410514377,38],[936701038,32],[250859386,30],[899330584,34],[100382286,34],[669712197,37],[653956768,35],[618382938,31],[566928648,31],[821778829,33],[273748598,39],[543682292,32],[912252567,41],[303958748,33],[238614954,31],[655469339,37],[355148646,32],[170199693,34],[882471282,38],[287710957,33],[199683207,37],[787598932,36],[962383938,33],[880093775,35],[904400869,34],[35908040,38],[111893172,36],[638091725,36],[190328742,34],[586505687,27],[545813098,37],[447559489,34],[82412661,34],[525107442,38],[852479623,34],[983778531,32],[528476806,35],[498654575,36],[215980399,32],[174322226,31],[954353400,32],[981741263,36],[57077704,38],[722768665,29],[119795169,31],[862064961,37],[955300332,29],[425546342,30],[796988416,30],[639672250,35]],"1010042162":[[517311749,5],[779553368,5],[359901870,6],[905403767,4],[899221565,5],[502420527,4],[865228432,6],[449626174,4],[632158866,5],[727952816,5],[60402820,6],[92967414,4],[474064842,7],[862301166,6],[353377831,5],[555354324,5],[518441848,5],[807000467,7],[943221661,5],[394416112,5],[148813978,4],[930002484,5],[957535220,5],[220584264,6],[132060071,5],[806602223,5],[217421779,6],[164924693,6],[555140744,8],[614435426,3],[432897287,6],[167467368,4],[704725277,5],[812651449,5],[413999641,5],[986801716,6],[888271664,5],[383963551,5],[169253812,5],[489809536,4],[919745627,6],[828988207,4],[874527827,3],[420587416,5],[269540871,6],[406991364,5],[55691246,4],[720733021,6],[367860448,5],[481485622,4],[618809582,5],[864587310,6],[852005523,7],[166244857,6],[431171752,5],[373866675,5],[644642173,4],[801569480,4],[17552663,4],[599427657,6],[468175327,6],[556807266,6],[896510679,7],[66790086,4],[789548689,3],[328732304,4],[751363202,5],[849600546,6],[320090595,6],[33882977,5],[530779519,5],[317073340,6],[183084739,5],[41191704,5],[785452373,7],[135483697,5],[873347847,6],[702193761,4],[706620349,6],[777122960,8],[718392276,5],[489561862,6],[183090953,5],[112790229,5],[240856267,6],[849489344,4],[456612460,5],[467510959,8],[427782219,5],[331145315,4],[303148030,7],[221906714,6],[527271606,6],[979683980,5],[747439287,4],[306975095,5],[100686610,5],[107044363,6],[943165002,5],[53471920,5]],"1093125211":[[779553377,63],[905403774,66],[319444747,59],[727952818,61],[521573371,62],[353377836,63],[807000481,66],[674990297,62],[597209688,60],[895395975,59],[251007671,62],[703212775,62],[441244713,65],[837979015,62],[269540874,62],[896474842,62],[720733036,60],[522906890,59],[481485638,62],[166244875,63],[53980382,61],[666080563,63],[151999868,63],[435327457,65],[462220619,60],[183084744,60],[104175451,60],[873347847,59],[768394376,62],[718392279,60],[78601148,58],[134829885,61],[456612463,67],[332067092,61],[583261546,63],[301673207,61],[979683991,60],[107044364,60],[53471934,64],[930397870,63],[669561892,60],[60253177,59],[691403509,62],[109618053,65],[200913698,59],[819612914,61],[490772971,64],[599662237,64],[82829073,63],[836965837,61],[65217010,63],[486813788,61],[756443457,68],[150280976,63],[324095431,66],[77453359,60],[566916197,65],[556748466,63],[884871521,59],[552449962,63],[518932700,64],[669373848,59],[903784598,62],[734140531,65],[926896589,64],[822576267,63],[272564540,64],[725793996,61],[712894698,66],[968688820,62],[608092103,58],[579422910,63],[855777062,60],[754726738,62],[423512840,66],[716477647,60],[572794323,63],[893566528,64],[993164937,63],[286664593,63],[583284913,60],[677365648,67],[36707770,62],[449172186,58],[822355784,63],[251032007,66],[559318967,60],[667318465,64],[466173943,64],[583269542,67],[266149088,63],[823442065,65],[224683300,64],[408715193,66],[60972026,58],[115073195,61],[425762192,57],[730472217,62],[774460101,63],[727159790,61]],"1095935398":[[517311750,6],[779553368,11],[359901870,8],[905403767,9],[899221565,11],[502420527,10],[865228431,8],[449626174,11],[632158866,9],[727952816,10],[60402820,9],[92967414,9],[474064842,9],[862301166,11],[353377831,11],[555354324,7],[518441848,10],[807000467,11],[943221661,9],[394416111,7],[148813978,8],[930002483,9],[957535220,10],[220584264,10],[132060070,9],[806602223,11],[217421779,8],[164924693,8],[555140744,11],[614435426,11],[432897287,7],[167467368,10],[704725277,10],[812651449,10],[413999641,10],[986801716,10],[888271664,10],[383963550,10],[169253812,10],[489809536,8],[919745627,10],[828988207,10],[874527828,8],[420587416,8],[269540869,11],[406991364,11],[55691246,10],[720733021,11],[367860448,8],[481485622,10],[618809582,8],[864587310,7],[852005523,10],[166244857,8],[431171752,8],[373866675,9],[644642173,8],[801569480,7],[17552662,10],[599427656,9],[468175326,10],[556807265,10],[896510679,11],[66790086,10],[789548689,10],[328732305,11],[751363202,6],[849600546,10],[320090595,12],[33882977,9],[530779519,5],[317073340,11],[183084738,10],[41191704,11],[785452373,11],[135483697,10],[873347847,9],[702193759,8],[706620349,10],[777122960,11],[718392276,10],[489561862,9],[183090953,11],[112790229,8],[240856267,11],[849489344,10],[456612460,10],[467510959,10],[427782219,8],[331145315,9],[303148030,11],[221906714,8],[527271606,12],[979683980,9],[747439287,8],[306975095,6],[100686610,10],[107044363,7],[943165002,10],[53471920,7]],"1137989174":[[779553368,87],[359901873,53],[905403767,90],[899221565,134],[502420527,93],[865228432,77],[449626174,107],[632158867,88],[727952818,93],[60402821,70],[92967414,87],[353377832,112],[807000467,96],[943221662,93],[394416111,88],[148813978,82],[930002483,96],[957535224,96],[220584264,90],[806602225,68],[217421782,66],[555140744,93],[614435427,71],[432897287,97],[167467370,142],[704725277,75],[812651449,102],[986801716,97],[888271664,80],[383963550,72],[489809536,125],[919745627,97],[828988207,90],[874527827,135],[420587416,86],[269540869,65],[406991365,91],[720733022,83],[367860451,69],[618809583,75],[864587310,84],[852005524,79],[166244858,80],[431171754,59],[373866677,63],[644642173,60],[801569481,71],[17552662,65],[599427656,87],[468175326,104],[556807265,70],[896510679,72],[66790086,90],[328732307,93],[751363203,80],[849600547,108],[320090596,69],[530779520,90],[183084739,94],[41191704,80],[785452378,90],[135483697,64],[702193759,92],[706620351,74],[777122960,79],[489561863,92],[183090953,76],[849489344,96],[456612460,83],[467510959,79],[427782219,83],[331145315,85],[303148030,91],[221906714,77],[979683980,98],[306975095,79],[100686610,90],[107044364,66],[943165002,72],[53471920,75],[449676409,119],[633672784,88],[885192267,94],[992905746,65],[341313291,93],[368153383,73],[713560156,71],[477344425,70],[829930232,104],[173061358,61],[268265774,93],[170755447,101],[169058864,100],[109618042,81],[353907518,73],[476101724,78],[963979607,111],[228789490,112],[729874057,77],[578778123,75]],"1325957594":[[517311750,31],[779553368,28],[359901871,28],[905403767,30],[899221565,26],[502420527,26],[865228431,29],[449626174,30],[632158866,26],[727952816,23],[60402822,28],[92967414,29],[474064842,30],[862301166,27],[353377831,29],[555354324,23],[518441848,25],[807000467,31],[943221661,30],[394416111,31],[148813978,24],[930002483,30],[957535220,31],[220584264,26],[132060070,27],[806602223,28],[217421779,26],[164924693,23],[555140744,32],[614435426,25],[167467368,24],[704725277,28],[812651449,27],[413999641,27],[986801716,23],[888271664,30],[383963550,31],[169253812,28],[489809536,33],[919745627,25],[828988207,27],[874527827,28],[420587416,26],[269540869,31],[406991364,25],[55691246,29],[367860448,23],[481485622,31],[618809582,27],[864587310,30],[852005523,28],[166244858,31],[431171752,30],[373866675,26],[644642174,27],[801569480,29],[17552662,30],[599427657,29],[468175327,28],[556807265,23],[896510679,24],[66790086,30],[789548690,28],[328732304,25],[751363202,25],[849600546,27],[320090595,21],[33882977,30],[530779519,28],[317073341,31],[183084738,26],[41191705,28],[785452373,28],[873347847,28],[702193759,28],[706620349,28],[777122961,29],[718392277,29],[489561862,28],[183090954,25],[112790229,30],[240856267,26],[849489344,31],[456612460,29],[467510959,32],[427782219,27],[331145316,25],[221906715,21],[527271606,26],[979683982,24],[747439290,26],[306975095,28],[100686610,27],[107044363,28],[943165003,29],[53471922,24],[274059747,33],[447571286,27],[715686840,29],[53342778,30]],"1344724498":[[517311750,8],[779553368,8],[359901870,8],[905403767,10],[899221565,10],[502420527,8],[865228431,8],[449626174,11],[632158866,8],[727952816,10],[60402820,8],[92967414,8],[474064842,10],[862301166,11],[353377831,9],[555354324,9],[518441848,8],[807000467,10],[943221661,9],[394416111,9],[148813978,8],[930002483,8],[957535220,10],[220584264,10],[132060070,9],[806602223,10],[217421779,10],[164924693,9],[555140744,11],[614435426,11],[432897287,9],[167467368,8],[704725277,10],[812651449,9],[413999641,9],[986801716,10],[888271664,10],[383963550,10],[169253812,10],[489809536,9],[919745627,11],[828988207,9],[874527828,7],[420587416,10],[269540869,11],[406991364,11],[55691246,8],[720733021,9],[367860448,9],[481485622,9],[618809582,8],[864587310,7],[852005523,9],[166244857,9],[431171752,9],[373866675,7],[644642173,9],[801569480,8],[17552662,10],[599427656,9],[468175326,10],[556807265,9],[896510679,11],[66790086,9],[789548689,11],[328732305,10],[751363202,5],[849600546,10],[320090595,10],[33882977,10],[530779519,8],[317073340,10],[183084738,11],[41191704,9],[785452373,11],[135483697,9],[873347847,9],[702193759,8],[706620349,11],[777122960,11],[718392276,10],[489561862,9],[183090953,9],[112790229,8],[240856267,11],[849489344,9],[456612460,11],[467510959,10],[427782219,10],[331145315,9],[303148030,10],[221906714,8],[527271606,12],[979683980,10],[747439287,8],[306975095,8],[100686610,10],[107044363,8],[943165002,9],[53471920,8]],"1426736851":[[449626181,52],[474064856,52],[862301169,49],[216391707,54],[930002502,47],[217421779,52],[704725281,52],[812651451,45],[413999646,53],[369922196,52],[383963553,48],[406991367,52],[166244862,53],[801569481,50],[507734877,48],[789548690,48],[328732306,55],[849600553,54],[135483699,50],[832015651,51],[718392291,52],[183090957,54],[36424325,47],[503069731,50],[427782225,50],[303148039,52],[527271623,50],[100686629,52],[943165021,50],[53471925,53],[501944760,50],[593020247,54],[400839518,55],[488889245,48],[963979621,54],[967868685,51],[864373136,50],[845723994,50],[668115602,55],[924660835,50],[909428422,54],[250408197,50],[384294284,54],[19954856,46],[410963079,54],[987668488,50],[441261793,54],[198914271,54],[547521379,55],[386230342,51],[101514759,51],[430797691,52],[974242448,50],[407453560,54],[510172280,52],[663460070,52],[531250199,54],[61030519,51],[999067067,48],[218676154,50],[792526380,50],[420785176,49],[828953373,51],[88551614,51],[659901596,51],[389292823,51],[780444115,49],[288223855,51],[547809975,49],[316268797,54],[986840020,56],[867521382,54],[165318385,54],[617928241,51],[951139191,53],[27741532,52],[525255460,50],[71693425,53],[185013282,54],[785592579,53],[730972461,48],[602356402,53],[606767513,54],[399659174,51],[104287445,51],[487986756,48],[900303988,51],[417781719,54],[317748061,50],[984863338,51],[513857776,47],[932015733,53],[960826537,55],[217970687,50],[810113472,53],[33844296,51],[721785434,48],[218789500,49],[889935757,52],[38566170,53]],"1462498410":[[517311755,104],[779553372,84],[359901871,86],[905403768,108],[632158876,71],[60402825,88],[92967414,79],[474064842,100],[862301168,95],[518441850,86],[807000470,127],[394416117,72],[148813983,92],[930002491,95],[220584265,84],[806602223,71],[217421783,94],[164924693,241],[555140746,121],[704725281,342],[986801717,78],[888271674,138],[383963550,154],[169253815,133],[919745629,97],[874527828,98],[55691246,85],[720733025,83],[481485626,78],[618809583,112],[431171766,269],[373866694,195],[644642177,83],[801569481,76],[556807267,117],[789548693,71],[328732305,111],[33882979,112],[317073345,145],[785452381,83],[135483701,79],[873347848,220],[706620354,192],[777122964,294],[718392292,83],[112790234,101],[456612463,95],[221906717,132],[979683991,87],[912752756,86],[306975099,121],[107044364,103],[943165011,85],[53471926,100],[969612070,86],[547490894,87],[184760491,89],[181302523,73],[498578150,90],[669251629,89],[360603265,88],[707728842,69],[296100002,103],[983378982,85],[294557183,104],[12131979,85],[15121746,117],[78969635,84],[313410661,73],[920241841,88],[195597694,101],[315603026,90],[919227772,100],[117808711,90],[364589997,87],[36141314,84],[257562188,85],[223862356,112],[385270696,102],[349114392,91],[584922802,99],[70838752,81],[414798895,88],[556748466,118],[528403606,94],[684646646,117],[59720297,99],[409813391,112],[187972126,107],[468783470,81],[604152514,89],[630352551,91],[126638089,90],[77961948,101],[922892672,93],[477704820,102],[958682209,86],[136256234,91],[342308551,111],[916693772,85]],"1796831549":[[517311750,31],[779553373,32],[359901871,28],[905403767,30],[899221566,31],[502420528,36],[449626174,32],[727952816,31],[60402820,35],[92967414,29],[353377831,34],[555354324,34],[807000467,33],[394416111,36],[930002483,35],[220584265,35],[132060070,37],[217421779,31],[164924693,33],[614435426,33],[432897290,33],[704725278,35],[812651449,33],[383963553,31],[169253813,34],[919745629,32],[828988210,34],[874527827,32],[269540869,34],[406991365,35],[55691247,35],[720733022,32],[367860448,35],[481485624,34],[618809582,31],[852005524,37],[166244857,35],[431171752,31],[373866676,31],[801569481,33],[17552661,28],[599427656,35],[468175326,34],[556807265,28],[896510680,31],[66790086,35],[789548689,33],[328732305,33],[751363202,31],[849600546,39],[320090596,30],[530779519,33],[317073340,33],[183084738,38],[41191705,34],[785452373,33],[135483698,29],[873347847,32],[702193760,34],[706620350,36],[718392278,35],[489561862,34],[112790231,35],[240856267,33],[456612461,34],[467510960,32],[331145316,37],[221906718,31],[747439288,33],[100686615,33],[107044364,33],[943165002,34],[53471920,33],[264112260,34],[908418556,30],[367649568,35],[592309991,32],[519337958,32],[950071282,33],[250408176,32],[340691025,34],[332716030,29],[450503175,33],[838989418,31],[112948850,31],[433811071,31],[251629618,34],[405886192,36],[116482902,33],[172844756,33],[127921611,33],[919244086,35],[322037030,34],[259106838,36],[453885337,32],[93324158,35],[20367593,36],[495950114,36],[13895907,32],[571940313,32]],"1838295370":[[905403785,55],[449626175,61],[474064843,60],[862301168,57],[555354338,58],[518441852,54],[943221678,52],[148813985,49],[741614014,52],[919745641,51],[828988216,46],[896510683,49],[530779521,50],[785452384,50],[873347862,60],[240856269,51],[849489351,47],[331145315,46],[221906718,55],[107044370,54],[366549824,58],[261278341,51],[757797305,62],[859673116,57],[155670414,63],[284635454,49],[687960475,57],[948104508,58],[827840070,48],[858253820,51],[250408182,56],[909558311,52],[781382258,59],[728369752,52],[705352481,53],[647584688,47],[190170926,42],[197527900,57],[639220113,60],[37120532,63],[279235968,56],[56552038,62],[257217217,52],[551469294,52],[813137173,55],[17866595,50],[208310239,62],[442033939,54],[87137659,50],[218789518,50],[202828937,48],[984630697,50],[477525961,49],[137992974,53],[697991961,46],[405895525,51],[398626655,60],[180021799,52],[375189246,49],[581934824,51],[395624457,59],[965484396,51],[927587342,48],[313410665,56],[65128627,61],[447135099,55],[394781473,50],[674746559,56],[585672808,53],[383587385,49],[850187141,55],[45920196,49],[158564799,64],[912469067,50],[535554697,64],[521473631,65],[177158286,53],[259769159,62],[46362480,54],[683955104,52],[916179288,60],[289291554,54],[795847151,59],[578823351,45],[96317197,57],[756235816,50],[71733672,57],[375928705,46],[287544596,48],[178138191,49],[951965722,53],[266801,48],[283374596,55],[587046139,56],[85513485,54],[850947576,60],[79917824,58],[156167447,60],[805436800,61],[160665092,53]],"1922600978":[[574448084,46],[555140746,53],[661991762,50],[888271676,52],[618809591,54],[801569486,53],[298206263,53],[320090600,49],[702193760,48],[331145316,56],[221906718,51],[903086587,60],[666781869,60],[967664391,70],[963148006,59],[913236699,50],[589061930,60],[512589205,56],[133286173,49],[711368071,60],[111301750,50],[25911045,49],[999067066,53],[115415301,58],[419955930,50],[691360391,64],[193349736,51],[5194983,61],[497441175,61],[4789969,53],[13232396,59],[599662295,53],[675149731,51],[424555325,55],[993725837,55],[373440642,52],[821991612,66],[417985877,49],[487279738,62],[109294597,54],[260160769,67],[641379396,57],[944832607,57],[359957157,64],[326589526,57],[27188444,56],[778663434,51],[456479713,53],[694782304,42],[68099156,54],[409725117,56],[461739064,60],[594000829,55],[485885399,60],[994544717,48],[163632575,51],[586662308,56],[475367287,64],[852332754,59],[632641070,53],[357207802,54],[354623612,49],[839078624,51],[747779688,67],[400973112,52],[980352621,53],[138929728,56],[819360381,50],[66596065,61],[180240774,55],[764176763,56],[274934912,53],[781729966,55],[495333072,53],[52077069,54],[144305469,51],[208071019,52],[157972357,59],[587531509,63],[5659701,46],[754178802,52],[9747458,48],[918814045,56],[947301264,51],[502137143,54],[395034507,53],[664624753,54],[630118730,54],[909322634,66],[819302618,58],[271501743,59],[210994230,59],[811494471,50],[613368025,52],[728529418,55],[14958271,56],[673191321,61],[155043609,54],[31434555,50],[982940625,53]],"2043144636":[[739021958,26],[497166222,28],[983863046,26],[160610517,24],[419587449,24],[262545614,30],[959740145,23],[330356691,28],[714335911,23],[497876383,25],[406889140,22],[216399590,26],[301897140,26],[178998688,27],[579849009,28],[601466122,28],[826139148,25],[59667572,26],[69857754,27],[843358494,27],[941633483,25],[331921400,24],[528301489,23],[296294644,24],[574283545,26],[575692621,26],[954004964,26],[262271594,29],[451482933,26],[759155596,25],[244171920,26],[25766886,23],[49006152,28],[6910600,26],[439606523,26],[438787101,25],[894037188,25],[509776461,26],[871125970,26],[119076848,26],[236103763,22],[501018918,21],[112825160,25],[99902826,26],[165032950,25],[886385105,27],[176144293,25],[13283489,26],[156946613,24],[427611537,25],[282870796,24],[981105294,26],[397686256,27],[367617432,26],[897290040,29],[630464579,29],[83169467,27],[999196513,23],[888450748,24],[533754345,29],[571595053,26],[799203659,23],[351360165,26],[110784286,24],[75546940,25],[23876198,24],[996109588,24],[392619906,25],[793267431,28],[766802807,23],[249289701,24],[995716746,24],[589937228,23],[859858058,28],[93769998,24],[722192911,27],[798703820,25],[680637220,28],[524346351,25],[746467381,27],[188349847,25],[396311667,24],[210585207,26],[481490977,26],[329039168,25],[502018966,28],[829664228,26],[245026874,27],[893897226,25],[749655052,24],[51433467,27],[931782242,26],[211517685,26],[873188823,22],[156387065,26],[776723750,27],[19951455,24],[838708609,27],[95544447,27],[610864075,25]],"2128103473":[[517311749,14],[779553368,15],[359901870,12],[905403767,14],[899221565,14],[502420527,13],[865228431,11],[449626174,13],[632158866,12],[727952816,11],[60402820,12],[92967414,12],[474064842,10],[862301166,14],[353377831,13],[555354324,13],[518441848,11],[807000467,14],[943221661,13],[394416111,12],[148813978,13],[930002483,12],[957535220,13],[220584264,11],[132060070,13],[806602223,11],[217421779,15],[164924693,15],[555140744,14],[614435426,16],[432897287,14],[167467368,10],[704725277,14],[812651449,13],[413999641,14],[986801716,14],[888271664,13],[383963550,14],[169253812,14],[489809536,14],[919745627,12],[828988207,14],[874527828,13],[420587416,15],[269540869,13],[406991364,13],[55691246,12],[720733021,12],[367860448,13],[481485622,12],[618809582,15],[864587310,13],[852005523,13],[166244857,16],[431171752,16],[373866675,15],[644642173,12],[801569480,12],[17552661,13],[599427656,13],[468175326,13],[556807265,12],[896510679,16],[66790086,14],[789548689,12],[328732304,12],[751363202,15],[849600546,13],[320090595,12],[33882977,14],[530779519,12],[317073340,14],[183084738,12],[41191704,13],[785452373,13],[135483697,12],[873347847,15],[702193759,11],[706620349,12],[777122960,13],[718392276,12],[489561862,12],[183090953,12],[112790229,13],[240856267,13],[849489344,12],[456612460,14],[467510959,10],[427782219,12],[331145315,11],[303148030,10],[221906714,12],[527271606,15],[979683980,14],[747439287,12],[306975095,13],[100686610,13],[107044363,13],[943165002,12],[53471920,12]],"-2091623303":[[517311749,16],[779553368,17],[359901870,13],[905403767,16],[899221565,15],[502420527,14],[865228431,14],[449626174,15],[632158866,15],[727952816,16],[60402820,14],[92967414,17],[474064842,13],[862301166,16],[353377831,13],[555354324,13],[518441848,12],[807000467,13],[943221661,20],[394416111,15],[148813978,15],[930002483,12],[957535220,17],[220584264,15],[132060070,18],[806602223,17],[217421779,14],[164924693,14],[555140744,15],[614435426,16],[432897287,13],[167467368,13],[704725277,15],[812651449,15],[413999641,15],[986801716,17],[888271664,17],[383963550,14],[169253812,14],[489809536,15],[919745627,16],[828988207,17],[874527827,14],[420587416,15],[269540869,16],[406991364,15],[55691246,14],[720733021,17],[367860448,16],[481485622,15],[618809582,17],[864587310,16],[852005523,13],[166244857,16],[431171752,19],[373866675,13],[644642173,14],[801569480,14],[17552661,17],[599427656,16],[468175326,12],[556807265,12],[896510679,18],[66790086,18],[789548689,18],[328732304,15],[751363202,14],[849600546,17],[320090595,18],[33882977,16],[530779519,17],[317073340,19],[183084738,15],[41191704,13],[785452373,15],[135483697,17],[873347847,18],[702193759,16],[706620349,18],[777122960,15],[718392276,15],[489561862,16],[183090953,15],[112790229,14],[240856267,18],[849489344,15],[456612460,16],[467510959,15],[427782219,16],[331145315,16],[303148030,16],[221906714,15],[527271606,15],[979683980,17],[747439287,14],[306975095,19],[100686610,15],[107044363,16],[943165002,18],[53471920,15]],"-2087777902":[[517311750,14],[779553368,16],[359901870,11],[905403767,16],[899221565,17],[502420528,17],[865228431,12],[449626174,16],[632158867,16],[727952816,16],[60402820,15],[92967414,14],[474064842,10],[862301166,17],[353377832,16],[555354324,16],[518441848,11],[807000467,14],[148813978,15],[930002483,13],[957535222,17],[220584265,14],[132060070,18],[806602223,13],[555140744,16],[614435426,14],[432897287,16],[167467368,10],[812651449,18],[413999641,16],[986801716,17],[888271665,13],[383963551,13],[169253814,15],[489809536,15],[919745627,15],[828988208,17],[874527828,13],[420587416,18],[269540870,15],[406991364,13],[55691246,10],[720733021,14],[367860448,13],[481485622,12],[618809582,21],[864587310,13],[852005523,15],[166244857,18],[431171752,15],[801569480,12],[599427656,12],[468175327,16],[556807266,14],[896510679,18],[328732304,13],[849600547,12],[320090595,14],[33882977,17],[530779520,16],[317073340,16],[183084738,14],[785452373,14],[873347849,13],[702193760,17],[706620350,12],[777122961,14],[718392276,13],[489561863,13],[183090953,14],[112790230,15],[240856267,14],[849489346,13],[456612463,16],[467510959,11],[427782219,13],[331145316,13],[303148031,15],[221906714,13],[527271606,13],[979683980,14],[747439287,13],[100686611,15],[107044363,15],[943165002,11],[53471920,11],[579986211,15],[405173294,15],[387546104,15],[301735349,17],[645362681,12],[23783154,14],[647277882,15],[162825210,16],[132427717,19],[174372584,17],[806903956,13],[765729344,14],[6935246,12],[672225767,17]],"-2043200314":[[517311749,7],[779553368,6],[359901870,5],[905403767,4],[899221565,5],[502420527,4],[865228432,5],[449626174,4],[632158866,5],[727952816,4],[60402820,6],[92967414,5],[474064842,7],[862301166,7],[353377831,6],[555354324,6],[518441848,4],[807000467,7],[943221661,6],[394416112,4],[148813978,5],[930002484,4],[957535220,4],[220584264,5],[132060071,4],[806602223,4],[217421779,7],[164924693,6],[555140744,6],[614435426,3],[432897287,5],[167467368,4],[704725277,6],[812651449,7],[413999641,4],[986801716,6],[888271664,6],[383963551,6],[169253812,4],[489809536,4],[919745627,7],[828988207,4],[874527827,3],[420587416,6],[269540871,6],[406991364,4],[55691246,5],[720733021,7],[367860448,4],[481485622,4],[618809582,6],[864587310,6],[852005523,7],[166244857,7],[431171752,4],[373866675,5],[644642173,4],[801569480,4],[17552663,4],[599427657,7],[468175327,5],[556807266,7],[896510679,7],[66790086,4],[789548689,3],[328732304,4],[751363202,4],[849600546,7],[320090595,5],[33882977,5],[530779519,4],[317073340,7],[183084739,6],[41191704,4],[785452373,7],[135483697,4],[873347847,5],[702193761,5],[706620349,6],[777122960,6],[718392276,4],[489561862,7],[183090953,5],[112790229,7],[240856267,6],[849489344,5],[456612460,5],[467510959,6],[427782219,4],[331145315,5],[303148030,7],[221906714,5],[527271606,6],[979683980,4],[747439287,5],[306975095,4],[100686610,5],[107044363,6],[943165002,5],[53471920,5]],"-2030835643":[[517311749,14],[779553368,15],[359901870,13],[905403767,14],[899221565,14],[502420527,14],[865228431,11],[449626174,13],[632158866,14],[727952816,12],[60402820,12],[92967414,13],[474064842,8],[862301166,14],[353377831,13],[555354324,13],[518441848,11],[807000467,14],[943221661,13],[394416111,13],[148813978,13],[930002483,12],[957535220,12],[220584264,11],[132060070,13],[806602223,12],[217421779,15],[164924693,15],[555140744,15],[614435426,13],[432897287,13],[167467368,9],[704725277,13],[812651449,13],[413999641,14],[986801716,14],[888271664,16],[383963550,13],[169253812,13],[489809536,14],[919745627,12],[828988207,13],[874527828,14],[420587416,15],[269540869,12],[406991364,12],[55691246,10],[720733021,12],[367860448,13],[481485622,12],[618809582,15],[864587310,14],[852005523,13],[166244857,16],[431171752,16],[373866675,14],[644642173,12],[801569480,12],[17552661,13],[599427656,13],[468175326,13],[556807265,12],[896510679,16],[66790086,14],[789548689,12],[328732304,11],[751363202,15],[849600546,15],[320090595,12],[33882977,14],[530779519,12],[317073340,14],[183084738,11],[41191704,13],[785452373,13],[135483697,12],[873347847,15],[702193759,11],[706620349,11],[777122960,13],[718392276,11],[489561862,13],[183090953,11],[112790229,13],[240856267,14],[849489344,11],[456612460,12],[467510959,14],[427782219,12],[331145315,11],[303148030,9],[221906714,14],[527271606,15],[979683980,14],[747439287,12],[306975095,11],[100686610,13],[107044363,12],[943165002,12],[53471920,13]],"-1963939495":[[517311750,51],[779553368,51],[359901870,40],[905403768,36],[899221565,40],[502420527,44],[865228431,32],[449626174,42],[632158866,35],[727952816,41],[60402820,45],[92967414,34],[474064842,40],[862301166,37],[353377831,49],[555354324,37],[518441848,46],[807000467,52],[943221663,50],[394416111,58],[930002483,36],[957535220,47],[220584264,37],[132060070,49],[806602223,41],[217421779,44],[164924693,36],[555140744,44],[614435426,45],[432897287,39],[167467368,61],[704725277,60],[812651449,38],[413999641,46],[986801716,45],[888271665,29],[489809536,38],[919745627,46],[828988207,39],[874527827,34],[420587416,35],[269540869,82],[406991364,30],[55691246,36],[720733021,47],[367860448,38],[481485623,53],[618809584,44],[864587310,41],[852005523,46],[166244859,53],[431171752,32],[373866675,39],[644642174,29],[801569481,48],[17552661,32],[599427656,32],[468175326,41],[556807265,44],[896510679,43],[66790086,44],[789548689,52],[328732304,46],[751363202,40],[849600546,39],[320090595,38],[33882977,39],[317073340,48],[183084738,47],[41191704,32],[785452373,59],[135483697,40],[873347847,68],[702193759,35],[706620349,46],[777122960,53],[718392276,36],[489561862,43],[240856267,47],[849489344,39],[456612460,43],[467510959,35],[427782219,35],[331145316,32],[303148030,37],[221906714,40],[527271607,38],[979683980,41],[747439287,39],[306975095,39],[100686610,39],[107044363,42],[53471920,46],[140279861,41],[308272188,35],[669871811,44],[667821012,38],[240098407,36],[799004814,33],[59907303,55]],"-1886763509":[[517311749,17],[779553368,17],[359901870,14],[905403767,15],[899221565,15],[502420527,14],[865228431,16],[449626174,17],[632158866,18],[727952816,16],[60402820,16],[92967414,17],[474064842,13],[862301166,18],[353377834,18],[555354325,14],[518441848,12],[807000467,14],[943221661,18],[394416111,15],[148813978,14],[930002483,12],[957535220,19],[220584264,15],[132060070,18],[806602223,17],[217421780,17],[164924693,14],[555140744,17],[614435426,16],[432897287,13],[167467368,13],[704725278,13],[812651449,15],[413999641,16],[986801716,16],[888271664,17],[383963550,14],[169253812,16],[489809536,18],[919745627,16],[828988207,17],[874527827,15],[420587416,15],[269540869,16],[406991364,15],[55691246,14],[720733022,16],[367860449,16],[481485622,14],[618809582,19],[864587310,15],[852005524,17],[166244857,16],[431171753,14],[373866675,13],[644642173,15],[801569480,15],[17552661,18],[599427656,17],[468175326,12],[556807265,12],[896510679,19],[66790087,15],[789548689,17],[328732304,15],[751363202,14],[849600547,16],[320090595,18],[33882977,16],[530779519,17],[317073340,19],[183084738,15],[41191704,13],[785452373,15],[135483697,17],[873347847,19],[702193760,18],[706620349,18],[777122960,16],[718392276,15],[489561862,19],[183090953,16],[112790229,15],[240856267,19],[849489344,16],[467510959,16],[427782219,14],[331145316,17],[303148030,16],[221906714,15],[527271606,16],[979683980,18],[747439287,15],[306975095,20],[100686610,14],[107044364,16],[943165002,18],[53471920,17],[36541547,15]],"-1842859596":[[636804787,61],[557518004,54],[178998688,63],[865123507,66],[860413498,56],[374952893,66],[509776464,55],[623472769,60],[501018924,55],[793003650,59],[922606325,59],[156946616,52],[293819728,63],[108389848,57],[701919062,62],[635000509,63],[183776958,57],[745257887,70],[658566216,58],[249289701,58],[995716747,58],[251731130,57],[575394997,60],[467100015,67],[421896165,64],[500392922,59],[617071733,62],[489910876,59],[123225447,66],[252367849,66],[665426895,58],[521388602,79],[656672371,58],[145742740,58],[978122940,52],[988062649,59],[413509226,60],[700939684,65],[583129468,66],[995651921,58],[260704423,71],[836326140,58],[718022518,61],[976588331,54],[821024301,57],[710692572,62],[31574997,67],[706384427,52],[484088820,63],[221653946,58],[982748995,56],[266654040,61],[626349874,68],[450690016,53],[297117038,63],[602433815,58],[136500454,66],[406580652,54],[383416465,57],[298000841,71],[699006280,63],[758301142,62],[415695300,61],[773077099,58],[950451953,58],[371413077,58],[821778836,69],[565024935,61],[712894702,59],[436246807,57],[459350740,62],[502209290,59],[716471189,67],[639186308,57],[778427524,58],[753158384,65],[566079476,67],[102552624,60],[91372710,56],[625088390,61],[262541079,69],[719478958,64],[583316430,59],[376924214,64],[701037263,52],[467626548,53],[764938559,69],[119542852,59],[411229084,58],[129985270,58],[697291113,74],[775670257,60],[518296450,66],[609207718,65],[890488281,61],[244845391,64],[194113241,59],[910145481,51],[899683545,61],[614283729,60]],"-1718786349":[[517311749,62],[779553368,64],[60402820,66],[862301169,56],[353377832,57],[943221661,68],[148813979,61],[806602224,62],[986801716,61],[919745627,61],[406991364,86],[367860449,66],[481485622,62],[618809587,56],[852005523,60],[166244858,55],[373866676,58],[801569481,59],[599427656,60],[468175326,66],[896510679,61],[66790088,57],[530779520,63],[317073341,63],[183084739,55],[785452373,57],[135483701,65],[873347847,61],[706620350,63],[718392276,60],[112790229,56],[849489345,64],[467510959,60],[221906715,61],[979683984,59],[747439296,65],[100686610,57],[107044367,59],[943165009,57],[648328478,67],[430252580,67],[924875341,64],[836326140,58],[132041730,56],[534476191,62],[162638416,66],[407344145,60],[239143493,62],[773874727,57],[216433755,56],[699929496,62],[792854393,64],[571940313,63],[116341035,54],[583129469,60],[226374901,58],[487938535,63],[849443398,69],[129252587,57],[842437862,54],[433853775,59],[594010329,73],[931782242,60],[217269494,60],[625654612,67],[797431565,62],[76236424,65],[116175115,62],[466894408,61],[383254750,72],[569408233,56],[938428908,57],[32754024,59],[930477774,62],[601133574,62],[333619811,63],[437852683,74],[350449209,61],[496807273,65],[426787269,66],[681281303,61],[56676707,58],[704329037,58],[411826145,62],[842074148,60],[441341219,59],[816612288,60],[790894563,73],[455033159,62],[783657285,64],[771028393,60],[724841705,59],[291524812,56],[153345552,73],[20027329,65],[698793357,61],[585186883,65],[95894036,61],[49909430,62],[358839311,66]],"-1590996086":[[779553374,49],[359901878,46],[905403774,49],[502420534,46],[865228431,46],[632158871,47],[862301166,49],[353377835,47],[555354327,43],[394416121,42],[148813979,47],[930002488,45],[220584269,47],[217421786,50],[164924704,47],[555140750,46],[614435426,48],[812651456,49],[888271665,45],[383963550,49],[489809540,48],[828988216,45],[55691246,46],[720733021,44],[852005523,44],[166244865,47],[431171753,46],[373866675,48],[17552661,47],[468175326,46],[789548692,45],[751363202,45],[317073347,45],[183084756,46],[777122961,47],[456612464,51],[427782229,49],[221906718,44],[100686611,47],[53471923,47],[752105934,49],[314381461,47],[275160667,46],[656672358,46],[60953000,44],[65851265,48],[774587611,45],[28978715,46],[529873329,46],[303085380,47],[120272625,47],[902208839,47],[859434553,45],[497108527,51],[830798819,47],[153724252,49],[76268807,47],[613926436,47],[7739089,48],[825006129,46],[228212647,48],[282636370,48],[983113755,44],[97678514,45],[360803237,49],[872783122,49],[840783155,50],[538193572,46],[684912694,45],[441315644,48],[528494856,44],[831440155,45],[875504859,47],[628008985,45],[21819171,47],[816542009,45],[423863459,45],[330116012,46],[901271210,51],[333619810,41],[482785622,45],[801798469,48],[806620853,49],[106986001,49],[960948877,47],[556235331,47],[397637732,47],[409203457,53],[179282187,46],[147796097,45],[999288411,47],[102627211,47],[397640342,47],[69197318,45],[992029521,43],[855777057,46],[643750204,46],[998048154,46],[505434713,45],[419393695,48]],"-1487240467":[[517311749,30],[779553368,31],[359901870,30],[905403767,30],[899221566,31],[502420528,30],[865228431,27],[632158866,30],[862301166,29],[353377832,27],[555354324,29],[518441849,29],[807000468,31],[943221661,29],[394416111,29],[148813978,30],[930002483,32],[957535220,29],[220584265,26],[132060070,33],[806602223,30],[217421780,27],[164924693,26],[555140744,29],[614435427,27],[432897287,25],[167467368,28],[704725277,30],[812651449,29],[413999641,32],[986801716,30],[383963550,31],[169253812,33],[489809536,30],[919745627,31],[874527827,29],[420587416,29],[269540869,27],[720733022,32],[367860448,30],[481485622,30],[864587310,29],[166244857,30],[431171752,28],[373866675,30],[644642173,28],[801569482,29],[17552661,25],[599427656,30],[556807266,27],[896510679,30],[789548689,31],[328732304,31],[751363202,32],[849600546,31],[320090596,31],[33882978,27],[530779519,31],[317073341,29],[183084739,29],[41191704,26],[785452373,31],[135483699,30],[702193759,30],[706620349,31],[777122961,28],[489561862,27],[183090953,27],[112790230,29],[240856267,33],[849489344,30],[456612460,33],[467510959,28],[427782219,29],[331145315,30],[303148032,29],[221906714,33],[527271606,31],[979683980,32],[747439287,31],[107044363,32],[53471920,31],[401057540,30],[501334050,29],[537036347,29],[232236497,28],[912866995,28],[961344131,29],[547590073,29],[764176667,29],[902070024,28],[581338021,29],[950146915,31],[802801644,31],[925301815,31],[226365099,30],[834885544,28],[620971585,31],[19742316,30],[213546098,26]],"-1383484848":[[517311749,17],[779553368,17],[359901870,15],[905403767,17],[899221565,15],[502420527,15],[865228431,14],[449626174,17],[632158866,16],[727952816,16],[60402820,14],[92967414,17],[474064842,13],[862301166,16],[353377831,17],[555354324,16],[518441848,14],[807000467,15],[943221661,20],[394416111,16],[148813978,15],[930002483,12],[957535220,17],[220584264,15],[132060070,18],[806602223,17],[217421779,16],[164924693,14],[555140744,17],[614435426,17],[432897287,15],[167467368,14],[704725277,17],[812651449,16],[413999641,15],[986801716,17],[888271664,17],[383963550,14],[169253812,15],[489809536,15],[919745627,16],[828988207,18],[874527827,17],[420587416,17],[269540869,16],[406991364,15],[55691246,14],[720733021,16],[481485622,17],[618809582,17],[864587310,19],[852005523,14],[166244857,17],[431171752,19],[373866675,14],[644642173,17],[801569480,14],[17552661,17],[599427656,16],[468175326,14],[556807265,15],[896510679,18],[66790086,18],[789548689,18],[328732304,15],[751363202,16],[849600546,17],[320090595,16],[33882977,17],[530779519,17],[317073340,19],[183084738,16],[41191704,14],[785452373,16],[135483697,17],[873347847,17],[702193759,17],[706620349,18],[777122960,18],[718392276,15],[489561862,16],[183090953,15],[112790229,17],[240856267,18],[849489344,15],[456612460,16],[467510959,15],[427782219,16],[331145315,16],[303148030,17],[221906714,15],[527271606,16],[979683980,17],[747439287,15],[306975095,19],[100686610,15],[107044363,16],[943165002,18],[53471920,16],[724317139,19]],"-1367843655":[[517311750,34],[779553368,30],[359901871,27],[905403767,32],[899221565,26],[502420527,27],[865228431,29],[449626174,29],[632158866,26],[727952816,25],[60402822,28],[92967414,26],[474064842,31],[862301166,28],[353377831,27],[555354324,24],[518441848,26],[807000467,30],[943221661,30],[394416111,30],[148813978,27],[930002483,31],[957535220,28],[220584264,27],[132060070,26],[806602223,28],[217421779,24],[164924693,22],[555140744,32],[614435426,26],[432897287,26],[167467368,25],[704725277,30],[812651449,25],[413999641,26],[986801716,24],[888271664,28],[383963550,31],[169253812,30],[489809536,30],[919745627,26],[828988207,32],[874527827,29],[420587416,28],[269540869,28],[406991364,25],[55691246,31],[720733022,28],[367860448,27],[481485622,35],[618809582,26],[864587310,31],[852005523,26],[166244858,28],[431171752,29],[373866675,27],[644642173,31],[801569480,29],[17552662,25],[599427656,29],[468175327,28],[556807265,24],[896510679,22],[66790086,29],[789548690,26],[328732304,28],[751363202,25],[849600546,26],[320090595,24],[33882977,28],[530779519,31],[317073341,30],[183084738,28],[41191705,29],[785452373,28],[135483697,29],[873347847,28],[702193759,28],[706620349,27],[777122961,27],[718392277,28],[489561862,28],[183090954,24],[112790229,29],[240856267,26],[849489344,27],[456612460,29],[467510959,27],[427782219,27],[331145316,27],[303148030,29],[221906714,26],[527271606,27],[979683980,30],[747439289,31],[306975095,28],[100686610,27],[107044363,29],[943165003,28],[53471922,23]],"-1278686646":[[517311749,16],[779553368,17],[359901870,13],[905403767,16],[899221565,15],[502420527,14],[865228431,14],[449626174,15],[632158866,15],[727952816,16],[60402820,14],[92967414,17],[474064842,13],[862301166,16],[353377831,13],[555354324,13],[518441848,12],[807000467,13],[943221661,20],[394416111,15],[148813978,15],[930002483,12],[957535220,17],[220584264,15],[132060070,18],[806602223,17],[217421779,14],[164924693,14],[555140744,15],[614435426,16],[432897287,13],[167467368,13],[704725277,15],[812651449,15],[413999641,15],[986801716,17],[888271664,17],[383963550,14],[169253812,14],[489809536,15],[919745627,16],[828988207,17],[874527827,14],[420587416,15],[269540869,16],[406991364,15],[55691246,14],[720733021,17],[367860448,16],[481485622,15],[618809582,17],[864587310,16],[852005523,13],[166244857,16],[431171752,19],[373866675,13],[644642173,14],[801569480,14],[17552661,17],[599427656,16],[468175326,12],[556807265,12],[896510679,18],[66790086,18],[789548689,18],[328732304,15],[751363202,14],[849600546,17],[320090595,18],[33882977,16],[530779519,17],[317073340,19],[183084738,15],[41191704,13],[785452373,15],[135483697,17],[873347847,18],[702193759,16],[706620349,18],[777122960,15],[718392276,15],[489561862,16],[183090953,15],[112790229,14],[240856267,18],[849489344,15],[456612460,16],[467510959,15],[427782219,16],[331145315,16],[303148030,16],[221906714,15],[527271606,15],[979683980,17],[747439287,14],[306975095,19],[100686610,15],[107044363,16],[943165002,18],[53471920,15]],"-1208358067":[[517311749,18],[779553368,21],[359901873,21],[905403767,22],[899221565,20],[502420527,21],[865228433,18],[449626174,22],[632158867,21],[727952818,20],[60402821,22],[92967414,24],[474064842,22],[862301166,23],[353377832,20],[518441848,23],[807000467,22],[943221661,20],[394416111,20],[148813978,22],[930002483,28],[957535220,21],[220584264,17],[132060070,20],[806602225,23],[217421780,21],[164924693,28],[555140744,20],[614435426,22],[432897287,18],[167467370,22],[704725277,24],[812651449,22],[413999641,18],[986801716,19],[888271664,20],[383963550,20],[169253812,23],[489809536,21],[919745627,21],[828988207,23],[874527827,17],[420587416,20],[269540869,24],[406991364,19],[55691246,21],[720733021,19],[367860450,20],[481485622,21],[618809582,21],[864587310,23],[852005523,25],[166244858,16],[431171754,22],[373866676,21],[644642173,22],[801569481,21],[17552661,20],[599427656,22],[468175326,19],[556807265,18],[896510679,21],[66790086,19],[789548689,22],[328732307,20],[751363202,20],[849600547,21],[320090596,19],[33882977,17],[530779519,22],[317073340,21],[183084738,24],[41191704,22],[785452378,18],[135483697,22],[873347847,22],[702193759,19],[706620350,21],[777122960,18],[489561863,21],[183090953,22],[112790229,21],[240856267,23],[849489344,21],[456612460,22],[467510959,24],[427782219,20],[331145315,20],[303148030,22],[221906714,20],[527271606,21],[979683980,24],[747439287,19],[306975095,23],[100686610,20],[107044363,19],[943165002,21],[53471920,24],[213092212,19],[776990685,19]],"-1176092191":[[779553368,61],[359901872,63],[905403767,59],[899221567,61],[502420531,61],[865228431,59],[632158866,61],[727952816,64],[92967415,53],[862301167,67],[807000470,64],[943221661,64],[394416112,70],[930002484,67],[957535222,63],[806602223,61],[217421780,70],[164924693,64],[432897289,61],[167467368,58],[704725277,79],[812651449,72],[888271664,75],[383963552,64],[489809536,61],[919745631,56],[874527827,55],[406991365,71],[55691246,63],[720733021,57],[367860449,68],[481485622,75],[864587310,57],[852005523,59],[166244857,58],[431171752,61],[644642174,59],[801569481,81],[17552662,64],[468175326,74],[556807265,57],[896510679,63],[66790086,57],[789548690,72],[849600546,61],[530779523,67],[317073341,60],[183084742,62],[41191704,64],[785452373,69],[135483697,57],[873347849,60],[706620350,63],[718392279,69],[489561862,65],[183090954,81],[240856273,69],[456612460,68],[427782220,77],[303148030,177],[221906714,56],[747439288,55],[306975096,60],[943165003,329],[53471923,72],[214329862,68],[522495705,63],[297649463,71],[522437065,64],[2843997,65],[193613731,78],[312635551,59],[868627365,62],[873547615,63],[688864707,74],[913840686,74],[490162379,67],[408129798,57],[882592116,67],[360145594,72],[742039743,68],[240856267,67],[695507488,59],[898650445,65],[759273922,66],[619622849,80],[701037259,57],[385106390,55],[131359383,56],[698086699,74],[456583092,63],[718545573,69],[47957928,66],[412314108,71],[457545402,74],[365451187,60],[274484782,77],[722483587,68],[615401133,52],[681751763,65]],"-1121316185":[[458128238,38],[865228432,38],[360468040,34],[92967430,32],[353377840,35],[673191325,37],[957535223,32],[220584264,34],[590049539,37],[164924700,36],[555140745,34],[614435436,36],[564310248,32],[167467376,36],[413999641,37],[383963550,38],[169253815,35],[919745627,33],[896474853,35],[55691263,38],[801569491,36],[17552662,38],[468175333,35],[896510681,37],[66790092,35],[33882985,38],[317073353,36],[183084742,38],[135483700,37],[873347850,36],[641892525,39],[364909525,35],[183090967,37],[456612469,36],[303148041,35],[527271613,34],[306975096,34],[107044363,31],[943165002,37],[53471929,35],[952101986,35],[148091173,37],[683605042,36],[244596027,38],[332846706,38],[297718260,38],[38237849,40],[980781946,36],[168437470,37],[286900055,37],[972565268,36],[877910720,33],[714907653,36],[43025626,33],[395016578,34],[802859797,37],[8062263,39],[447559488,38],[213207466,36],[45061290,39],[717986939,35],[586694077,38],[313410679,37],[807399935,35],[810620479,38],[518970568,31],[522160970,35],[928793264,34],[676944969,40],[379942012,36],[790768761,37],[959664214,38],[57572388,36],[473861469,34],[477704828,35],[958273256,36],[189378553,36],[978828586,37],[445093203,36],[921188771,34],[119290442,39],[362170643,37],[122936103,35],[120235089,36],[790355487,38],[751105777,36],[431628234,38],[414039224,37],[526785783,37],[24992897,37],[800378096,35],[433111230,33],[631767400,35],[882552231,35],[329169980,36],[325086300,39],[135618133,36],[554245784,37],[555103910,34],[213971428,36]],"-1112664773":[[517311751,14],[779553370,17],[359901870,12],[905403768,14],[899221565,13],[502420527,13],[865228431,16],[449626174,15],[632158867,12],[727952816,16],[60402822,19],[92967414,15],[474064842,13],[862301167,16],[353377832,12],[555354324,12],[518441850,15],[807000469,16],[943221661,12],[394416111,14],[148813979,16],[930002485,11],[957535220,13],[220584265,12],[132060070,17],[806602223,12],[217421779,13],[164924694,14],[555140745,10],[614435426,11],[432897289,10],[167467368,15],[704725277,17],[812651453,17],[413999644,14],[986801716,16],[888271664,14],[383963550,16],[169253815,13],[489809536,13],[919745627,13],[828988207,16],[874527828,19],[420587420,11],[269540869,15],[406991364,13],[55691247,13],[720733022,11],[367860448,15],[481485622,14],[618809585,13],[864587310,14],[852005523,18],[166244857,13],[431171752,15],[373866676,13],[644642176,14],[801569481,15],[17552664,14],[599427657,14],[468175327,17],[556807265,12],[896510679,17],[66790086,14],[789548690,15],[328732305,15],[751363203,12],[849600546,14],[320090597,17],[33882977,13],[530779519,14],[317073340,16],[183084738,17],[41191705,17],[785452373,16],[135483699,13],[873347847,12],[702193760,14],[706620349,16],[777122960,14],[718392276,10],[489561862,16],[183090954,14],[112790229,13],[240856267,15],[849489344,16],[456612460,15],[467510959,13],[427782219,12],[331145315,13],[303148030,17],[221906716,15],[527271608,16],[979683980,15],[747439288,13],[306975095,16],[100686610,15],[107044364,14],[943165002,16],[53471920,14]],"-1013080196":[[517311750,32],[779553368,30],[359901870,30],[905403767,30],[899221566,31],[502420528,30],[865228432,29],[632158866,30],[92967414,28],[474064843,30],[862301166,30],[353377832,27],[555354324,29],[518441849,29],[943221661,28],[394416111,29],[148813979,30],[930002483,32],[957535220,26],[220584265,23],[132060070,32],[806602223,31],[217421780,27],[164924693,24],[555140744,29],[432897287,25],[167467370,29],[704725277,30],[986801716,30],[383963550,26],[169253812,33],[489809536,30],[919745628,28],[874527827,29],[269540869,26],[720733021,27],[367860448,30],[481485622,26],[864587310,29],[852005527,28],[166244857,30],[431171752,27],[373866675,29],[644642173,28],[801569482,29],[17552661,25],[599427656,28],[556807265,28],[896510679,31],[66790087,29],[789548689,31],[751363202,31],[849600546,32],[33882978,27],[530779519,31],[317073341,29],[183084739,29],[41191704,27],[785452373,31],[702193759,29],[706620349,30],[777122961,28],[489561862,25],[183090953,27],[112790229,29],[240856267,30],[849489346,29],[456612461,28],[467510959,26],[427782219,29],[331145315,27],[303148030,26],[221906714,33],[979683980,32],[747439287,31],[100686611,31],[107044363,34],[53471920,31],[387546106,31],[74123423,27],[373001036,29],[156563090,32],[862833726,28],[919009785,30],[130157051,26],[71622838,28],[21506322,33],[577380490,29],[48597122,24],[380341319,22],[197991936,30],[467556033,28],[760926187,29],[244845377,28],[800644473,27],[717290121,29],[225703192,31],[68293990,24],[866983966,26],[288711922,29]],"-960771562":[[359901870,78],[899221565,80],[632158867,65],[727952817,65],[92967416,72],[474064842,67],[353377836,55],[518441849,76],[807000468,56],[220584265,67],[217421786,64],[555140745,76],[432897287,67],[704725289,82],[413999642,59],[986801725,53],[919745628,67],[874527828,78],[269540870,68],[720733022,63],[864587312,79],[852005526,52],[431171756,67],[373866682,53],[801569483,55],[556807268,59],[896510680,71],[66790087,62],[328732304,63],[751363209,72],[317073343,52],[183084741,59],[135483700,57],[873347848,73],[489561870,61],[849489347,70],[427782234,56],[303148031,74],[747439287,61],[100686610,60],[943165005,67],[53471921,53],[728606499,57],[858946078,91],[617071734,75],[510539916,77],[948880622,67],[283169047,68],[977497457,70],[25600217,58],[350438071,62],[95941797,78],[305269965,59],[2267590,63],[599904280,75],[835212292,52],[652800176,61],[518932701,67],[547726861,72],[840783155,73],[490562531,79],[138108209,74],[781931660,56],[30991808,89],[855131061,64],[182080071,109],[69074080,64],[636531838,55],[112948844,79],[389106445,87],[387803255,58],[251677693,70],[501746863,87],[835589882,79],[197879703,87],[311002989,69],[395527262,52],[669561880,59],[650828297,71],[552306137,67],[651317552,87],[719812818,84],[488302885,73],[209997934,82],[905803282,68],[384299638,64],[76908432,66],[752112481,95],[648664656,70],[351948887,63],[742661439,78],[613091597,61],[347935592,74],[797972596,82],[282942580,62],[283860031,56],[256558676,48],[820397259,68],[589508991,67],[408477780,86]],"-946946414":[[517311752,21],[779553371,21],[905403770,22],[899221565,22],[865228433,18],[449626177,24],[632158866,21],[92967417,23],[474064843,24],[555354324,21],[943221661,21],[394416115,18],[930002486,23],[220584264,19],[217421780,21],[812651450,19],[986801720,22],[888271664,23],[383963552,22],[169253816,24],[828988214,21],[269540870,23],[406991364,20],[481485623,26],[618809587,22],[864587312,19],[852005527,19],[373866677,22],[644642174,24],[599427656,20],[556807265,21],[896510686,24],[66790086,21],[789548694,25],[328732307,24],[751363202,22],[849600549,20],[33882981,21],[530779525,23],[317073344,23],[41191705,22],[785452373,25],[873347853,23],[706620350,24],[489561864,22],[112790229,20],[240856274,17],[331145322,21],[303148032,22],[221906715,18],[527271606,23],[100686612,21],[943165002,23],[53471920,27],[102084976,25],[638956012,21],[53791654,23],[617508216,21],[916371431,22],[534119098,21],[807687785,19],[524972883,27],[154224897,27],[497673485,18],[580772320,25],[430836949,23],[487282335,16],[208071024,24],[510368407,19],[558731826,27],[759656490,26],[551173720,22],[247577356,25],[862062812,20],[653210562,18],[670986167,22],[581162351,21],[510172254,24],[188102575,19],[818006594,22],[160344429,23],[965182451,24],[529563524,27],[74192850,21],[852011098,22],[149287815,25],[40388534,23],[670442454,22],[675402583,24],[468406901,22],[981572413,21],[299800222,23],[872296322,27],[854498839,19],[25533863,23],[723761268,21],[414590431,25],[490505348,27],[501322115,22],[38484307,23]],"-906730025":[[517311749,4],[779553368,5],[359901870,4],[905403767,3],[899221565,5],[502420527,3],[865228432,4],[449626174,3],[632158866,4],[727952816,3],[60402820,5],[92967414,4],[474064842,7],[862301166,6],[353377831,4],[555354324,4],[518441848,4],[807000467,7],[943221661,5],[394416112,4],[148813978,4],[930002484,3],[957535220,3],[220584264,5],[132060071,4],[806602223,4],[217421779,6],[164924693,6],[555140744,5],[614435426,3],[432897287,4],[167467368,3],[704725277,4],[812651449,4],[413999641,4],[986801716,5],[888271664,5],[383963551,4],[169253812,3],[489809536,4],[919745627,5],[828988207,4],[874527827,3],[420587416,4],[269540871,5],[406991364,3],[55691246,4],[720733021,6],[367860448,3],[481485622,3],[618809582,4],[864587310,5],[852005523,7],[166244857,5],[431171752,4],[373866675,3],[644642173,4],[801569480,3],[17552663,4],[599427657,6],[468175327,5],[556807266,6],[896510679,7],[66790086,3],[789548689,3],[328732304,3],[751363202,4],[849600546,6],[320090595,5],[33882977,5],[530779519,3],[317073340,6],[183084739,4],[41191704,4],[785452373,7],[135483697,3],[873347847,5],[702193761,4],[706620349,5],[777122960,5],[718392276,3],[489561862,6],[183090953,4],[112790229,4],[240856267,5],[849489344,4],[456612460,3],[467510959,5],[427782219,4],[331145315,4],[303148030,7],[221906714,4],[527271606,5],[979683980,4],[747439287,4],[306975095,4],[100686610,3],[107044363,5],[943165002,4],[53471920,3]],"-862639388":[[517311749,16],[779553368,17],[359901870,16],[905403767,16],[899221565,18],[502420527,18],[865228431,14],[449626174,15],[632158866,17],[727952816,16],[60402820,14],[92967414,17],[474064842,14],[862301166,16],[353377831,14],[555354324,15],[518441848,13],[807000467,15],[943221662,16],[394416111,16],[148813978,16],[930002483,12],[957535220,17],[220584264,15],[132060070,18],[806602223,19],[217421779,16],[164924693,14],[555140744,17],[614435426,19],[432897287,15],[167467368,15],[704725277,16],[812651449,15],[413999641,19],[986801716,17],[888271664,17],[383963550,17],[169253812,14],[489809536,15],[919745627,17],[828988207,19],[874527827,15],[420587416,17],[269540869,16],[406991364,17],[55691246,14],[720733021,16],[367860448,17],[481485622,15],[618809582,17],[864587310,16],[852005523,13],[166244857,16],[431171752,19],[373866675,16],[644642173,15],[801569480,15],[17552661,17],[599427656,16],[468175326,15],[556807265,14],[896510679,18],[66790086,18],[789548689,19],[328732304,15],[751363202,16],[849600546,17],[320090595,17],[33882977,19],[530779519,18],[183084738,16],[41191704,15],[785452373,15],[135483697,17],[873347847,18],[702193759,19],[706620349,18],[777122960,18],[718392276,15],[489561862,18],[183090953,16],[112790229,17],[240856267,18],[849489344,16],[456612460,16],[467510959,15],[427782219,16],[331145315,16],[303148030,16],[221906714,19],[527271606,15],[979683980,19],[747439287,17],[306975095,19],[100686610,18],[107044363,18],[943165002,18],[53471920,17],[777773891,16]],"-848879898":[[449626175,35],[862301171,33],[555354331,34],[518441849,37],[807000467,30],[394416116,31],[148813986,32],[220584265,34],[132060074,35],[806602225,32],[217421783,30],[432897298,31],[167467379,36],[812651450,33],[888271676,34],[919745634,35],[874527827,33],[153460468,32],[406991364,34],[720733022,32],[644642180,35],[17552670,35],[66790095,33],[789548692,33],[751363209,34],[320090601,33],[33882977,34],[530779521,32],[317073344,35],[873347849,35],[702193760,32],[706620353,34],[718392283,35],[489561863,33],[112790231,32],[467510964,34],[303148031,34],[221906720,35],[527271617,33],[979683980,37],[306975095,31],[100686617,37],[943165012,36],[626147845,35],[104109950,36],[917005602,33],[567468250,34],[910418400,33],[816386649,37],[903351711,33],[948880610,34],[703371393,37],[453718224,35],[822417361,33],[398431603,34],[918590137,33],[115020499,34],[480260373,34],[324591399,32],[391252634,30],[170685924,36],[47557949,35],[220484295,36],[936959693,36],[701469929,37],[862293931,35],[70770586,35],[622920253,31],[8052726,33],[909719398,33],[602433822,34],[168306041,34],[19929178,32],[252190293,36],[774460112,34],[595646018,32],[681059164,35],[703820688,34],[219806995,33],[767613365,34],[435327448,34],[760673598,33],[354580096,33],[453885338,35],[660184470,31],[307000527,32],[129841269,36],[449849670,32],[299018236,32],[160229167,35],[91378588,35],[555257759,36],[946814762,31],[353088465,36],[845201,36],[396796255,33],[347152806,34],[695852641,34],[667585750,32],[127921617,35]],"-774542587":[[865228442,76],[449626175,72],[92967425,71],[518441866,88],[943221669,80],[957535222,86],[555140755,97],[369922197,88],[888271676,131],[683783301,106],[431171767,69],[151519474,95],[749051104,102],[104109965,108],[718392293,74],[518097073,108],[953023078,95],[465447944,64],[101989990,132],[432959291,79],[368674597,94],[73130075,78],[775719695,86],[669679088,81],[134062100,86],[664449335,92],[387417120,129],[847785584,89],[162189257,83],[295540804,80],[452778338,76],[220484294,94],[260611107,63],[659981192,136],[42880734,79],[157046645,67],[85457894,99],[686226403,89],[981649426,91],[895109067,104],[962724775,123],[328919199,75],[564907499,132],[926663139,113],[416427424,93],[493568286,93],[337534347,81],[518970578,112],[843138033,73],[50396843,116],[261167478,85],[432611777,69],[864016081,75],[332852082,117],[916261007,94],[764510545,113],[497439651,78],[471533031,103],[761034853,116],[885296650,100],[879980147,89],[824390200,95],[297980167,96],[857820480,99],[346578104,97],[240098418,76],[496253343,101],[315607218,89],[359775116,96],[400685836,98],[66920154,90],[625088407,66],[247719809,92],[422316086,98],[379144913,93],[136761615,85],[832121092,90],[519436675,97],[629822835,83],[492898570,91],[861293420,92],[431824044,123],[590480531,95],[722046252,98],[541277345,121],[104532420,101],[296549854,80],[532220556,98],[507828881,90],[130896811,88],[455952044,109],[671798202,103],[914988538,78],[226830737,115],[481466556,105],[268528397,73],[126070543,101],[727135610,111],[218185505,86],[316467301,85]],"-740687052":[[965816938,68],[327077998,78],[126378691,70],[557884159,66],[53439153,69],[882377273,76],[866136349,70],[278657233,68],[352956391,74],[78095337,73],[692319744,74],[540877125,79],[167447682,79],[239281939,78],[96477139,81],[572499114,63],[653766508,73],[195211877,77],[490505426,81],[272394249,78],[369307756,75],[156108036,73],[723470018,71],[970883420,71],[322206457,69],[270933750,68],[26319511,66],[453284718,74],[679518113,70],[341566140,77],[540705363,63],[452357417,68],[438378475,72],[93590740,69],[591487607,79],[614085914,77],[473230111,84],[606530898,66],[362834579,80],[579422968,73],[331940866,64],[479954827,77],[353912869,78],[551841075,74],[736770728,76],[919232160,72],[876495672,63],[168306073,120],[439921511,79],[208347321,64],[274894019,78],[895165219,71],[710683592,69],[156712275,76],[93324179,79],[245842184,72],[8348050,74],[96522078,79],[266810,61],[518374930,72],[928117774,76],[946918817,81],[778148679,71],[204377608,82],[283277509,74],[794768784,91],[749987820,68],[123690510,67],[437088982,82],[695987156,83],[831440171,70],[285844971,72],[919502053,99],[962664936,63],[503055575,70],[121886794,62],[858063371,72],[63425496,67],[71313194,78],[263112355,76],[659594610,78],[314076756,74],[205732394,63],[661812973,69],[289150658,83],[747978517,62],[47321232,65],[123226331,71],[641645047,80],[769788026,83],[788666610,82],[431487576,80],[854760024,69],[471251637,75],[16049584,76],[166561137,73],[906168716,84],[433583172,67],[212972167,77],[463930445,80]],"-703861492":[[779553369,34],[905403767,30],[899221565,33],[502420528,36],[865228432,34],[449626174,34],[727952816,34],[60402820,35],[92967414,30],[862301166,36],[353377831,34],[555354324,34],[807000467,35],[394416111,36],[930002483,35],[957535222,34],[220584265,35],[132060070,37],[217421779,31],[164924693,35],[614435427,31],[432897290,34],[704725278,35],[812651449,33],[888271664,34],[383963555,34],[169253813,35],[919745629,34],[874527827,34],[420587420,35],[269540869,34],[406991365,35],[55691247,35],[720733021,35],[367860448,35],[481485624,36],[618809583,36],[864587311,35],[852005524,37],[373866676,33],[644642175,32],[801569480,31],[17552661,27],[599427656,35],[468175326,36],[789548689,34],[328732305,33],[751363202,33],[849600546,39],[320090596,31],[33882978,30],[530779519,33],[317073340,33],[183084738,38],[41191705,34],[785452373,33],[135483697,32],[702193760,35],[706620350,36],[777122960,36],[489561862,34],[240856267,34],[849489350,35],[456612461,34],[467510960,33],[221906718,30],[747439288,35],[100686615,33],[107044364,33],[53471920,36],[747782150,34],[877249320,36],[67514104,32],[33070402,33],[36840176,34],[707719092,36],[899330573,32],[587011656,30],[306565801,33],[555917120,34],[999128127,34],[913190620,36],[259163285,33],[690145023,31],[74196754,35],[742144792,34],[922892672,35],[311533573,34],[470994771,36],[25600215,37],[82206875,34],[36424315,34],[872353649,35],[483329965,32],[237796802,36],[129709033,35],[525095994,32],[275535093,34],[852912749,37],[800796209,35]],"-671914411":[[517311750,115],[779553368,61],[359901870,74],[899221565,88],[502420527,62],[865228432,76],[727952817,72],[60402825,78],[92967414,83],[474064842,74],[862301168,70],[353377831,52],[555354325,93],[518441848,118],[807000467,60],[943221662,71],[394416112,93],[148813978,64],[930002483,70],[132060071,121],[164924697,94],[614435426,159],[167467368,111],[812651449,236],[888271664,118],[383963550,121],[169253815,161],[489809536,140],[828988207,70],[420587417,105],[406991365,84],[720733022,112],[367860448,111],[618809583,121],[864587311,197],[852005524,125],[166244857,161],[431171753,304],[373866675,63],[644642173,134],[801569480,108],[468175327,74],[66790086,77],[789548689,94],[751363204,458],[849600547,120],[33882978,74],[530779520,201],[317073343,494],[183084739,148],[41191704,62],[785452373,321],[135483699,177],[702193759,78],[183090953,132],[112790231,148],[240856270,68],[456612461,100],[467510959,122],[303148030,139],[221906714,64],[527271610,86],[979683982,156],[747439287,149],[306975097,80],[100686611,96],[107044363,202],[943165005,129],[53471920,94],[81432310,67],[796581526,128],[177192686,125],[347935589,121],[220954068,103],[934688212,90],[267930979,60],[177158280,55],[292674913,76],[903784573,87],[217686313,124],[290396307,90],[430006424,125],[329494963,81],[312682995,76],[482031334,114],[16691991,85],[969121358,93],[535890799,98],[297718249,85],[245992207,87],[60900260,103],[929658630,87],[185665340,109],[628366297,73],[664126927,57],[473861471,80],[378820373,89],[664039,84],[902622972,91],[321824716,83]],"-651621524":[[449626182,84],[474064845,82],[253945714,111],[406991366,91],[19192784,100],[815709796,91],[789548696,96],[41191705,97],[135483711,100],[873347858,101],[768394370,99],[849489362,82],[456612467,85],[534197191,84],[303148045,86],[538279440,96],[979683981,88],[107044364,101],[531250188,97],[759818951,119],[82412683,74],[442714477,96],[20226445,132],[830556131,115],[842074150,122],[858876580,97],[744845719,96],[224565826,78],[699279748,125],[215980402,115],[328043794,82],[665625814,99],[86422799,123],[764431654,96],[913145572,125],[502824440,89],[962651556,112],[992029523,93],[251500790,138],[50476226,132],[890117837,101],[395034473,153],[449849685,104],[283277462,102],[715652977,106],[364590020,95],[171688406,115],[157593702,83],[690234350,100],[641399688,105],[577225277,125],[333518842,123],[234982527,89],[571494467,130],[235739702,106],[28468883,101],[471098535,107],[343161844,94],[141753493,132],[613592279,103],[297699279,115],[78770721,112],[295877771,103],[868379176,101],[837177591,78],[983530600,72],[798745818,73],[66554224,76],[661786809,87],[936260586,90],[504655627,101],[749834278,112],[143217031,142],[325037159,105],[396817769,136],[319165316,90],[570768024,97],[807867368,102],[924660835,89],[127733341,118],[996696941,86],[263808537,92],[242030195,81],[895521920,77],[269529671,98],[123558684,129],[775429750,91],[267482550,105],[78022030,110],[756431749,93],[903784585,110],[905803293,105],[880399980,96],[482785620,97],[969259033,89],[130239731,106],[515230766,92],[802082620,98],[450617657,95],[692375532,87]],"-635280127":[[517311750,34],[779553368,30],[359901870,30],[905403768,34],[899221566,31],[502420528,31],[632158866,31],[862301166,29],[353377832,28],[555354324,29],[518441849,29],[394416111,30],[148813979,32],[930002483,32],[220584265,28],[132060070,33],[806602223,31],[217421780,28],[164924693,24],[555140744,30],[432897287,27],[167467368,30],[704725277,31],[986801719,30],[383963550,32],[169253812,33],[919745628,29],[828988207,31],[874527827,29],[420587416,29],[269540871,31],[720733021,28],[367860448,31],[481485622,30],[618809583,28],[864587311,30],[852005527,28],[166244857,31],[431171752,27],[373866675,32],[644642173,32],[801569482,29],[17552661,25],[599427656,29],[896510679,31],[66790087,33],[789548689,31],[328732304,31],[751363202,31],[849600546,33],[320090595,29],[530779519,34],[317073340,32],[183084738,32],[41191704,28],[785452373,30],[702193759,30],[706620349,31],[489561863,30],[183090953,27],[112790230,30],[240856267,33],[849489346,31],[456612461,29],[467510959,29],[427782219,29],[331145315,25],[303148031,33],[221906714,33],[527271606,31],[979683980,32],[747439287,30],[107044363,34],[53471920,31],[374194555,30],[344228150,28],[30481107,29],[570354194,31],[287007624,31],[228805153,29],[993725837,33],[728903636,31],[421294475,29],[700228074,28],[754150179,27],[936489487,30],[381028156,29],[866045468,28],[977497457,29],[531130564,32],[809915733,30],[596096084,32],[524888190,30],[398490593,27],[964733867,30],[178417472,29],[281402301,32],[616274952,33],[863777159,30],[307644613,32]],"-606567752":[[517311749,13],[779553368,15],[359901870,12],[905403767,14],[899221565,14],[502420527,12],[865228431,11],[449626174,13],[632158866,12],[727952816,11],[60402820,12],[92967414,12],[474064842,8],[862301166,14],[353377831,10],[555354324,13],[518441848,11],[807000467,13],[943221661,13],[394416111,11],[148813978,13],[930002483,12],[957535220,12],[220584264,11],[132060070,13],[806602223,11],[217421779,12],[164924693,15],[555140744,14],[614435426,13],[432897287,12],[167467368,9],[704725277,13],[812651449,13],[413999641,14],[986801716,14],[888271664,13],[383963550,10],[169253812,12],[489809536,14],[919745627,12],[828988207,13],[874527828,12],[420587416,15],[269540869,12],[406991364,12],[55691246,9],[720733021,12],[367860448,13],[481485622,12],[618809582,15],[864587310,13],[852005523,12],[166244857,16],[431171752,13],[373866675,11],[644642173,12],[801569480,12],[17552661,14],[599427656,13],[468175326,13],[556807265,12],[896510679,16],[66790086,14],[789548689,12],[328732304,11],[751363202,15],[849600546,13],[320090595,11],[33882977,13],[530779519,9],[317073340,14],[183084738,11],[41191704,13],[785452373,12],[135483697,10],[873347847,14],[702193759,10],[706620349,10],[777122960,12],[718392276,10],[489561862,12],[183090953,11],[112790229,12],[240856267,12],[849489344,10],[456612460,12],[467510959,9],[427782219,12],[331145315,10],[303148030,9],[221906714,12],[527271606,14],[979683980,14],[747439287,11],[306975095,11],[100686610,13],[107044363,12],[943165002,12],[53471920,11]],"-600281204":[[517311749,18],[779553368,21],[359901870,13],[905403767,23],[899221565,21],[865228431,19],[449626174,24],[632158866,20],[727952817,21],[60402821,22],[92967414,23],[474064842,21],[862301166,23],[353377832,22],[555354324,21],[518441848,21],[807000467,23],[943221661,21],[394416111,19],[148813978,20],[957535220,20],[220584264,17],[132060070,19],[806602223,21],[217421779,18],[164924693,22],[555140744,22],[614435426,25],[432897287,19],[167467369,17],[704725277,23],[812651449,21],[413999642,15],[986801716,19],[888271664,20],[383963550,20],[169253812,23],[489809536,23],[919745628,21],[828988207,23],[874527827,17],[420587416,22],[269540869,23],[406991364,19],[55691246,21],[720733021,21],[367860448,22],[481485622,21],[618809582,21],[864587310,23],[852005523,24],[166244857,19],[431171752,21],[373866675,16],[644642173,22],[17552661,21],[599427656,19],[468175326,19],[556807265,18],[896510679,23],[789548689,23],[328732304,22],[751363202,20],[849600546,23],[320090595,18],[33882977,16],[530779519,22],[317073340,20],[183084738,25],[41191704,23],[785452373,21],[135483697,21],[873347847,23],[702193759,20],[706620349,20],[777122960,19],[718392277,23],[489561862,22],[183090953,20],[112790229,19],[240856267,24],[849489344,20],[456612460,22],[467510959,24],[427782219,21],[331145315,20],[303148031,24],[221906714,20],[527271606,22],[979683980,23],[747439287,19],[306975095,21],[100686611,20],[107044363,19],[943165002,22],[53471920,23],[279497004,22],[264350792,19],[446427122,22],[260577741,19]],"-600105873":[[517311749,19],[779553368,19],[359901870,20],[905403767,21],[899221565,19],[502420527,21],[865228431,18],[632158866,21],[727952816,18],[60402820,19],[92967414,20],[474064842,19],[862301166,20],[353377831,19],[555354324,21],[518441848,18],[807000467,20],[943221662,20],[394416111,20],[148813978,18],[930002483,20],[957535220,21],[220584264,21],[132060071,22],[806602223,17],[217421779,21],[164924693,19],[555140744,21],[614435426,18],[432897287,16],[167467368,21],[704725277,22],[812651449,20],[413999641,19],[986801716,22],[888271664,16],[383963550,18],[169253812,20],[489809536,18],[919745627,17],[828988209,16],[874527827,18],[420587416,20],[269540869,19],[406991364,19],[55691246,16],[720733021,19],[367860448,22],[481485622,17],[618809582,20],[864587311,22],[852005523,17],[166244857,19],[431171752,20],[373866675,18],[644642173,19],[801569480,21],[17552661,15],[599427656,21],[468175326,18],[556807265,20],[896510680,21],[66790086,22],[789548689,20],[328732304,21],[751363203,20],[849600546,21],[320090595,17],[33882977,17],[317073340,24],[41191704,19],[785452373,20],[135483697,19],[873347848,20],[702193759,21],[706620349,22],[777122960,21],[718392276,17],[489561862,21],[183090953,21],[112790229,20],[240856267,20],[849489344,23],[456612460,22],[467510959,21],[427782219,21],[331145315,20],[303148030,21],[221906714,20],[527271606,17],[979683980,20],[747439288,15],[306975095,18],[100686610,18],[107044363,16],[943165002,19],[53471920,19],[534231595,20],[406613771,19],[228762780,19]],"-580278930":[[517311749,70],[905403767,88],[899221567,71],[502420528,94],[632158867,75],[60402821,98],[92967414,81],[807000467,78],[943221669,70],[394416114,73],[217421780,81],[432897288,85],[167467368,83],[704725277,75],[812651451,79],[413999642,89],[986801718,78],[383963560,81],[828988208,79],[420587428,72],[269540869,79],[406991364,94],[720733021,78],[367860449,75],[166244857,87],[556807266,78],[896510680,99],[66790094,85],[849600548,85],[183084738,106],[41191705,76],[785452382,89],[135483700,97],[777122967,92],[183090954,84],[112790231,81],[240856268,86],[849489347,87],[427782221,101],[303148031,96],[979683985,71],[107044364,78],[943165003,84],[453718225,89],[286433979,93],[302814727,73],[354620069,80],[124267741,85],[586516324,86],[137352727,84],[147602271,75],[359950547,98],[87393951,93],[394159644,82],[916015453,82],[444903700,80],[701469941,86],[702811047,94],[592303853,85],[699929495,90],[607677663,80],[305632650,96],[989897535,85],[69942325,78],[363278510,76],[447821650,102],[806536939,88],[802762954,80],[66732405,86],[377692182,110],[195211804,76],[890610708,81],[40388534,91],[749597950,90],[781931663,87],[572815717,75],[227631813,85],[134965501,89],[57077703,81],[887613938,69],[813137172,90],[72305095,84],[566730821,71],[913400975,81],[848977500,93],[86585343,93],[744827989,99],[410296434,105],[742015317,78],[209291404,83],[524089351,74],[114805466,81],[59987059,79],[323222387,79],[355993083,93],[582582267,81],[376339386,78],[605357194,86],[958682204,77],[281178406,79]],"-577494253":[[392835183,18],[457543909,17],[925320773,18],[143128840,18],[656400547,17],[587416004,13],[165442984,19],[89636288,18],[300397497,16],[811248171,19],[391176713,18],[459302332,18],[42035479,16],[345065545,17],[623830278,17],[657520556,19],[648538419,18],[198372721,18],[480243772,13],[880878096,17],[642170784,15],[707894701,17],[569811292,18],[599830253,14],[173351994,16],[844544370,17],[878295489,18],[212077777,18],[740422798,18],[221390823,18],[886842141,18],[582843897,19],[9051630,19],[473953836,19],[802712618,15],[273883066,15],[365772992,19],[754254286,17],[778749561,14],[465843467,19],[974174162,14],[139598387,17],[317095158,19],[852017990,17],[551535310,21],[359951871,14],[338441589,17],[711648129,14],[940022951,16],[402791710,18],[835232915,14],[565881759,18],[443872639,16],[240188477,18],[815549207,15],[863008841,19],[404660566,20],[361812179,17],[401799942,16],[291816840,14],[442175636,18],[759905495,18],[95029266,14],[113199281,19],[536797890,19],[259743879,18],[255990098,16],[44706801,18],[49769267,14],[695507489,16],[534197201,17],[306565816,21],[879207655,16],[538279432,18],[509911281,18],[912752757,18],[226033149,16],[107044369,17],[331241237,17],[696348220,17],[473213763,20],[961804658,16],[857449024,17],[28558383,18],[291634886,17],[222774826,14],[464082631,21],[730327092,16],[952867642,19],[994508582,17],[889788747,19],[225443293,20],[256278020,20],[894502220,18],[600978,17],[929794436,18],[861199292,17],[713804243,22],[844115390,17],[770044344,16]],"-576443611":[[359901871,30],[502420530,29],[449626176,35],[632158868,33],[727952817,30],[92967433,30],[474064842,30],[862301171,32],[353377835,32],[518441848,33],[807000468,31],[394416124,36],[957535221,30],[217421788,32],[164924695,32],[614435430,30],[167467370,35],[812651452,32],[413999650,34],[888271674,29],[383963551,34],[169253812,30],[489809538,33],[420587419,32],[406991364,33],[367860451,34],[481485625,31],[618809595,29],[864587316,31],[852005528,32],[808107526,33],[431171756,31],[373866677,33],[468175328,32],[66790086,29],[328732309,32],[751363209,33],[183084747,31],[785452373,31],[702193759,34],[706620349,31],[777122962,31],[489561863,31],[112790230,33],[240856271,30],[849489352,33],[331145331,31],[303148032,33],[107044364,35],[943165004,30],[854642060,31],[907206708,32],[185352202,31],[891036564,31],[621166281,33],[517336138,32],[627794393,31],[544326376,27],[228212650,33],[991204665,35],[95513852,30],[676816793,29],[566666123,33],[743781993,28],[836965830,33],[473861474,30],[904400873,32],[912504855,34],[916161514,34],[969612068,31],[245992208,31],[432647893,33],[432127017,33],[230530576,32],[918018807,31],[16246502,31],[976036163,29],[715758088,33],[887657226,29],[18337641,33],[784435643,33],[664420315,30],[288530060,29],[421054985,31],[617704179,29],[267930979,31],[799764156,29],[386129553,32],[860353278,31],[211679073,30],[521388602,30],[2748082,28],[740188398,32],[965840288,31],[509265514,32],[141320235,32],[99781915,31],[956268403,28],[979825400,33],[157046638,33]],"-492540558":[[482302217,73],[733973323,71],[469827073,69],[151875369,69],[264046886,68],[501018920,75],[736996154,71],[648603882,70],[125643728,66],[282870800,64],[145780969,67],[986862083,74],[651701474,68],[964591227,69],[646998734,88],[82412667,66],[318375511,61],[340577080,75],[438965723,67],[954008882,77],[811433800,82],[910872697,66],[648437950,72],[100881712,76],[866323992,76],[795243583,72],[86658901,75],[273101542,74],[652746757,72],[372694741,67],[143830927,69],[895390801,75],[829949132,63],[529680405,66],[164130216,64],[626838921,70],[483767096,64],[13966671,71],[38421027,65],[716315414,85],[114949153,72],[609741739,82],[147450710,87],[354444028,76],[224542338,85],[85525580,78],[755558311,78],[546577358,64],[809665722,81],[300354937,77],[496717395,77],[463055381,75],[663663567,65],[620346133,74],[34890054,75],[772973238,57],[330832950,67],[827128537,68],[255301414,74],[844977067,69],[737320913,75],[255349370,77],[734574226,67],[417677399,67],[260322713,61],[218468994,78],[613072381,71],[596916069,65],[864681711,71],[951131502,76],[244829658,76],[761546027,80],[521880248,78],[527042517,76],[498799745,81],[22230503,75],[825549565,69],[120381649,75],[442563328,80],[879887445,83],[648793610,71],[890958172,77],[389790616,70],[235175072,81],[441074138,69],[487122350,64],[783768184,74],[505434723,70],[509767741,73],[351220913,64],[8062256,69],[366264984,75],[619738008,70],[696170976,71],[979905948,70],[820793969,66],[709504354,71],[714052679,72],[89908374,78],[432940670,78]],"-413861179":[[779553368,76],[353377833,63],[807000467,67],[943221678,77],[806602223,71],[164924693,64],[413999653,67],[169253813,65],[489809537,72],[420587427,70],[166244865,66],[644642187,64],[801569481,71],[328732306,68],[33882979,62],[785452377,70],[706620353,63],[832015652,68],[112790239,68],[575485113,67],[768000139,67],[107044364,59],[945096814,71],[517666965,75],[217008251,62],[405639139,71],[177776951,68],[912546418,74],[873333592,70],[944466490,71],[269625332,64],[369857999,65],[667961481,66],[391917630,68],[772056989,65],[541360865,71],[158169557,68],[330832959,67],[898587525,58],[223036995,76],[272424179,78],[881446496,63],[881036789,65],[147796091,67],[174846167,69],[889662352,71],[127058391,71],[59853102,69],[647277882,70],[58732437,66],[928344204,66],[297117046,67],[296100002,68],[173793351,68],[54436676,70],[617928238,69],[850819032,66],[30051300,64],[564379291,64],[417977598,63],[128261012,67],[580623554,70],[847726681,68],[632420968,69],[430821127,67],[305435361,70],[259395645,70],[765153128,70],[122188237,72],[82009811,66],[610470739,77],[291729271,64],[123225447,70],[803115751,72],[572333090,68],[147140308,77],[629545008,77],[490544655,65],[515324053,66],[936300630,70],[470365339,73],[127921613,80],[615969365,72],[592309990,68],[583316429,66],[972526998,68],[775391706,68],[807883653,67],[766009320,70],[656400767,65],[303155227,66],[690902129,76],[185013284,70],[610586946,77],[413544571,67],[179197664,68],[157972321,64],[535580900,72],[329151991,68],[48745485,66]],"-407920866":[[517311752,47],[905403774,48],[632158868,47],[92967414,46],[862301166,49],[353377833,46],[555354327,46],[807000468,47],[394416114,46],[148813979,47],[930002487,45],[220584269,49],[167467372,44],[704725281,46],[888271668,45],[383963550,49],[489809540,47],[874527827,50],[420587419,47],[720733021,48],[367860448,50],[618809587,47],[852005523,44],[431171756,49],[801569486,42],[599427657,46],[468175333,50],[556807276,49],[328732311,46],[751363202,49],[320090597,52],[530779520,51],[183084756,47],[41191717,45],[706620353,49],[489561865,49],[112790234,51],[331145315,48],[303148032,47],[221906718,46],[747439288,47],[306975099,50],[561521792,47],[842683491,45],[59508969,45],[317129019,42],[699164131,52],[807538516,48],[395542937,49],[156563090,45],[638724363,47],[868718391,47],[573698161,45],[282303430,45],[103221486,49],[197879701,48],[718989200,45],[805973614,48],[253945712,49],[213546099,44],[476755412,49],[895071988,42],[600538385,49],[709090745,46],[582254289,49],[250192741,48],[247559904,49],[507187854,50],[891696480,46],[367939804,46],[495254011,49],[849675978,46],[878009465,48],[988375256,47],[452225655,46],[56701646,48],[931770554,46],[304854023,48],[929518687,45],[745242371,47],[143919690,48],[441341226,46],[994633062,51],[990693752,46],[123558668,49],[667810182,43],[694782292,50],[901375987,49],[344228166,50],[59632329,45],[506184,50],[699751698,45],[897151357,48],[68187506,48],[738313535,48],[797972602,50],[968595162,47],[886825782,45],[576496949,48],[664420310,48]],"-384850278":[[517311749,25],[779553368,25],[359901872,22],[905403767,23],[899221565,23],[502420527,22],[865228432,24],[449626174,21],[632158866,23],[727952816,21],[60402820,21],[92967415,18],[474064842,27],[862301166,23],[353377831,18],[555354324,22],[518441848,23],[807000467,25],[943221661,23],[394416112,23],[148813979,24],[930002483,24],[957535222,22],[220584265,22],[132060070,19],[806602223,25],[217421779,25],[164924693,21],[555140744,22],[614435426,24],[432897288,23],[167467368,21],[704725277,24],[812651449,29],[413999641,22],[986801716,26],[888271664,23],[383963550,23],[169253814,21],[919745627,21],[828988207,24],[874527827,22],[420587418,24],[269540869,22],[406991367,20],[55691246,18],[720733022,25],[367860448,22],[481485622,21],[618809582,27],[864587310,23],[852005523,24],[166244857,29],[431171752,20],[373866675,20],[644642173,20],[801569480,24],[17552662,25],[599427658,20],[468175328,21],[556807265,16],[896510679,23],[66790086,22],[789548689,23],[328732305,22],[751363202,20],[849600546,20],[320090595,23],[33882977,23],[530779519,25],[317073340,21],[183084738,26],[41191704,22],[785452373,22],[135483697,25],[873347847,21],[702193759,23],[706620349,23],[777122960,27],[718392276,22],[489561862,24],[183090954,20],[112790229,21],[240856267,24],[849489344,21],[456612460,22],[467510959,25],[427782219,24],[331145316,23],[303148032,21],[221906714,20],[527271608,28],[979683980,25],[747439288,23],[306975095,20],[100686611,27],[107044363,25],[943165002,21],[53471920,21],[308689473,21]],"-368467311":[[359901870,65],[905403767,76],[502420536,68],[865228435,77],[632158872,80],[862301171,74],[518441852,69],[220584265,74],[132060072,80],[299800223,74],[888271666,74],[720733022,75],[468175329,74],[751363209,81],[317073342,72],[135483698,70],[489561862,70],[427782220,57],[303148031,88],[82206875,73],[638510972,80],[852011097,66],[678918016,71],[705001359,85],[300817061,73],[905166772,66],[655893407,71],[437225885,77],[200655,78],[398431610,74],[875445001,77],[810162171,75],[540447592,76],[925856065,72],[501322122,72],[77741400,77],[636531850,89],[185857779,67],[747978518,68],[332842197,73],[463068652,81],[795242014,77],[273144880,73],[654856205,77],[151533841,72],[238290185,73],[70085781,77],[110716833,74],[147734992,71],[889054306,80],[256996757,80],[371940876,71],[296549854,70],[629792455,68],[63167091,64],[602508982,79],[35097670,77],[182095059,71],[178321109,78],[53471929,83],[220521220,67],[358383371,80],[733611714,68],[649802605,73],[547557756,69],[78922923,71],[600407088,71],[745672598,72],[417715598,72],[994869408,64],[606055969,74],[212034860,81],[457545404,67],[864847383,84],[705621743,69],[439813542,69],[882222594,83],[905795572,75],[110440100,71],[506919024,65],[100382286,66],[480485746,72],[466300201,80],[19396790,69],[114477018,78],[559416718,64],[850341312,90],[346578094,84],[628398000,77],[14762896,65],[418246970,66],[837389446,77],[944832603,76],[418315672,79],[123946461,68],[245896452,74],[239523789,78],[759204305,76],[476558764,80],[233978704,69]],"-305057966":[[453224256,15],[955087335,14],[724765126,15],[365728647,15],[648432742,16],[518833950,41],[109397776,12],[271009115,13],[450828192,14],[325990380,39],[672259042,17],[149758807,15],[25107910,16],[242938045,15],[770130882,16],[392683637,13],[661667266,13],[943485692,14],[781910226,13],[93204457,12],[650929702,13],[394044940,14],[150341193,12],[531424253,12],[271359080,13],[974199904,15],[10696690,14],[719965461,11],[933714229,13],[285845329,13],[279225135,14],[619888190,17],[170861474,15],[375931783,12],[645371496,17],[84433533,16],[460957699,16],[23641296,13],[471725112,12],[120132822,14],[713759926,16],[282266967,14],[301338754,13],[802572892,14],[646633352,17],[112818580,16],[258733138,15],[204733234,31],[575228254,14],[808500641,15],[23564583,16],[338614635,14],[73102279,14],[932733523,17],[23480996,22],[292528353,13],[926322911,12],[521866946,12],[722714402,14],[43862269,15],[812207782,15],[189205207,12],[768394385,13],[255990106,16],[718392290,16],[795347056,21],[240856274,14],[849489360,15],[286900058,16],[314627884,16],[919670808,16],[769028148,13],[747439303,10],[904642802,10],[490553494,17],[107044370,12],[330592261,15],[182031076,15],[4120296,13],[533393009,15],[753021471,11],[727678801,13],[505140993,13],[876662048,16],[945650279,15],[765544668,15],[273748615,18],[826023737,14],[990955868,19],[935081767,12],[50396853,15],[202802572,16],[307896264,15],[889788744,15],[2094719,18],[101994218,15],[793542146,15],[380453845,11],[921460747,15],[578687632,11]],"-304007324":[[359901870,26],[502420530,29],[449626174,28],[60402820,31],[92967415,27],[353377831,29],[943221661,29],[930002488,30],[132060072,30],[555140747,28],[614435430,31],[432897287,27],[704725280,28],[703212758,31],[413999641,27],[986801717,28],[888271664,29],[489809540,30],[919745632,31],[874527832,33],[420587417,28],[269540871,29],[406991365,28],[720733021,30],[481485626,29],[864587313,30],[166244857,30],[431171753,29],[373866675,30],[801569480,31],[896510688,30],[66790091,30],[751363202,28],[849600547,28],[33882982,29],[530779519,28],[183084738,29],[135483697,28],[873347848,30],[706620351,29],[777122960,28],[489561869,27],[112790231,27],[456612469,28],[427782224,30],[121801048,30],[303148032,31],[979683980,29],[747439289,32],[100686613,29],[107044372,32],[889662355,31],[273090962,30],[893204360,29],[134218459,30],[212243106,31],[451329369,29],[312929054,31],[335318897,29],[188102580,27],[854642062,28],[954187309,31],[792047015,31],[560916602,29],[425320183,33],[33388730,30],[898605278,27],[562363758,31],[871361912,29],[178138191,31],[741606348,29],[552169014,29],[984978730,29],[823413288,29],[465448041,27],[196278088,28],[597423024,29],[765153127,30],[920395599,28],[702730281,29],[406926784,28],[922198415,31],[295485064,29],[883075817,31],[366549821,30],[843816921,31],[813216004,30],[38566165,27],[601740214,28],[677246682,29],[942614005,28],[595750012,30],[233830753,30],[511965085,30],[605715268,30],[561339114,29],[225593954,31],[342864795,31],[215478718,28],[968063498,30]],"-290327439":[[517311750,16],[779553368,19],[359901870,17],[905403769,20],[899221565,17],[502420528,18],[865228432,20],[449626174,15],[632158868,18],[727952816,20],[60402820,16],[92967415,19],[474064842,15],[353377836,16],[555354326,18],[518441848,16],[807000467,16],[943221662,15],[394416111,19],[148813978,15],[930002483,13],[957535220,20],[220584264,16],[132060070,24],[806602223,18],[217421780,20],[164924693,15],[555140745,22],[432897287,14],[167467368,17],[704725278,15],[812651449,20],[986801717,14],[888271664,20],[383963551,17],[489809536,15],[919745629,16],[874527827,16],[406991364,18],[55691246,14],[720733022,20],[367860448,17],[481485624,18],[618809583,17],[864587311,15],[852005524,20],[166244858,20],[431171753,15],[373866675,18],[644642173,17],[801569481,24],[17552661,21],[599427659,17],[556807265,14],[896510679,20],[328732304,15],[751363202,18],[320090596,13],[530779520,23],[317073340,20],[183084738,17],[785452373,17],[135483697,20],[873347847,19],[702193760,20],[706620350,16],[777122960,20],[718392276,16],[183090954,17],[112790229,15],[240856267,21],[849489344,18],[467510959,19],[427782222,19],[527271607,17],[747439288,17],[306975095,21],[100686610,16],[107044364,20],[943165003,16],[53471920,16],[384097525,16],[55183806,19],[3106686,18],[729874061,17],[752531116,21],[455768173,17],[59628687,21],[261889364,17],[913840687,16],[85425644,16],[449172184,20],[495173438,22],[687978810,18],[61006410,18],[880720950,17],[220935635,21],[909252369,18],[308305664,19],[250580174,20]],"-283360499":[[779553369,82],[60402821,59],[862301167,72],[943221661,113],[148813978,71],[930002484,79],[217421779,81],[555140746,81],[432897287,119],[167467370,99],[704725277,146],[812651450,82],[413999643,82],[888271666,79],[919745630,66],[269540869,73],[406991368,78],[55691247,99],[720733021,129],[367860448,216],[618809585,67],[864587310,75],[852005523,77],[431171755,82],[801569480,107],[17552661,98],[599427656,83],[789548691,100],[849600547,79],[320090600,66],[530779519,73],[135483699,78],[873347849,72],[706620350,82],[777122961,62],[718392277,77],[112790233,62],[467510964,94],[331145315,66],[303148030,76],[221906720,77],[527271606,107],[306975095,68],[100686610,71],[107044363,116],[943165010,71],[79529666,77],[654606296,64],[891772306,65],[827675735,61],[137336584,68],[580395755,75],[781643285,75],[989636855,88],[190514162,80],[605025171,69],[794758317,79],[982812233,76],[216156289,70],[666781851,80],[831707963,75],[647935775,73],[245026874,71],[668402829,65],[142468983,70],[990235523,82],[601740212,73],[497127724,77],[182368008,75],[690145026,81],[120404371,69],[727236248,69],[469774520,74],[997941875,74],[943546922,75],[558476733,75],[517295414,79],[395865046,67],[87137650,78],[251496439,73],[609298281,82],[356035220,89],[948226112,69],[298016968,71],[82175825,80],[266554252,76],[104287435,71],[850004175,56],[819644570,75],[324643316,66],[326653501,68],[378028448,69],[15555479,71],[696060044,73],[710454047,66],[300583733,62],[197712968,76],[156758067,74],[503006523,59],[380084399,72]],"-205274512":[[517311749,23],[779553368,23],[359901870,19],[905403767,21],[502420527,22],[727952816,21],[92967416,24],[474064842,24],[555354324,22],[518441848,19],[943221663,20],[148813978,20],[930002485,17],[220584266,21],[132060072,23],[806602223,23],[164924696,22],[432897287,20],[704725278,19],[812651449,19],[413999642,20],[986801717,20],[888271664,20],[383963550,24],[169253813,20],[489809536,21],[919745627,19],[420587417,25],[269540869,27],[55691247,22],[367860448,27],[864587310,28],[373866677,24],[644642173,22],[17552661,22],[599427656,28],[468175326,19],[556807266,20],[328732307,25],[751363202,24],[320090595,15],[33882977,19],[530779519,27],[183084738,23],[41191705,25],[135483698,26],[702193759,29],[706620350,23],[777122960,26],[489561863,21],[183090953,24],[112790230,24],[467510961,24],[331145316,24],[303148030,24],[221906714,20],[527271606,17],[979683981,24],[747439287,19],[306975095,21],[107044363,21],[943165003,22],[53471922,22],[974242447,30],[850341309,20],[434025512,21],[84861351,24],[36237590,22],[441295468,21],[774587610,18],[564255500,24],[400034001,21],[706569303,20],[825684135,26],[792191690,28],[585770587,23],[219504509,19],[688199414,22],[4120294,21],[785089222,20],[190908286,21],[358791853,25],[896480501,22],[66554216,19],[143217030,21],[969194251,26],[956446592,26],[973442670,23],[189205179,23],[895165214,21],[775040399,19],[130112665,20],[493949705,18],[134111762,22],[74123423,21],[358383360,27],[656672355,20],[127921611,22],[495883682,19],[901271210,20]],"-180984719":[[779553373,14],[359901872,15],[865228431,15],[632158870,13],[92967426,12],[862301166,18],[555354328,17],[518441850,19],[807000468,17],[943221663,18],[957535223,17],[220584267,19],[132060087,18],[806602224,19],[164924695,13],[614435433,13],[432897288,21],[209086736,16],[812651455,18],[413999641,19],[986801717,17],[169253827,17],[489809537,16],[919745638,14],[828988213,13],[420587426,14],[55691259,15],[720733027,18],[367860454,14],[481485623,12],[618809585,18],[852005523,15],[166244859,13],[373866675,14],[644642174,20],[468175326,13],[66790087,18],[789548697,15],[328732316,18],[751363202,15],[320090597,15],[317073345,14],[41191704,13],[785452373,15],[873347853,15],[706620356,13],[183090953,18],[112790230,16],[240856268,19],[467510961,15],[427782226,17],[303148031,21],[221906720,17],[527271622,14],[979683981,20],[747439287,14],[100686614,19],[107044365,14],[53471922,15],[639266369,23],[988375255,16],[563826826,15],[657292531,18],[617970171,19],[190135184,21],[811846150,16],[789551541,14],[968543135,15],[444307129,16],[28561535,17],[272342069,18],[582633681,21],[608271396,16],[169205811,15],[673208557,11],[148638837,16],[414546222,15],[114957032,17],[775391708,14],[593922932,16],[646998711,15],[571896601,15],[423512818,18],[330396854,18],[667558086,14],[790821375,17],[600131790,18],[119992686,20],[821871310,17],[793542146,15],[540081046,25],[130283843,18],[867465286,12],[532220527,18],[175601169,14],[113832908,15],[703676500,18],[453739275,15],[661183080,20],[38566170,17]],"-154035910":[[517311749,42],[779553368,45],[359901870,36],[905403767,44],[899221565,53],[502420527,36],[865228431,35],[449626174,50],[632158866,42],[727952816,35],[60402820,41],[92967414,41],[474064842,38],[862301166,39],[353377831,38],[555354324,34],[518441848,39],[807000467,38],[943221661,51],[394416111,48],[148813978,38],[930002484,52],[957535220,40],[220584264,38],[132060070,48],[806602223,48],[217421779,47],[164924693,26],[555140744,53],[614435426,51],[432897287,39],[167467368,47],[704725277,54],[812651449,37],[413999641,35],[986801716,46],[888271664,51],[383963550,46],[169253812,41],[489809536,46],[919745627,45],[828988207,44],[874527827,35],[420587416,49],[269540869,60],[406991364,37],[55691246,37],[720733021,31],[367860448,34],[481485622,47],[618809582,44],[864587310,43],[852005523,46],[166244857,38],[431171752,31],[373866675,48],[644642173,39],[801569480,54],[17552661,33],[599427656,49],[468175326,48],[556807265,41],[896510679,48],[66790086,39],[789548689,40],[328732304,44],[751363202,39],[849600546,49],[320090595,47],[33882977,37],[530779519,51],[317073340,44],[183084738,46],[41191704,36],[785452374,49],[135483697,49],[873347847,54],[702193759,41],[706620349,49],[777122960,52],[718392276,35],[489561862,50],[183090953,42],[112790229,40],[240856267,43],[849489344,45],[456612460,32],[467510959,38],[427782219,38],[331145315,42],[303148030,38],[221906714,33],[527271606,59],[979683980,48],[747439287,42],[306975095,41],[100686610,44],[107044363,42],[943165002,40],[53471920,41]],"-45478100":[[517311749,15],[779553368,15],[359901871,18],[905403767,16],[899221565,15],[502420527,17],[865228431,18],[449626174,17],[632158866,16],[727952816,17],[60402820,19],[92967414,18],[474064842,15],[862301166,16],[353377832,17],[555354324,15],[518441848,15],[807000467,15],[943221662,20],[394416111,14],[148813978,18],[930002484,18],[957535220,16],[220584264,16],[132060070,16],[806602223,16],[217421779,16],[164924693,15],[555140744,20],[614435426,17],[432897287,18],[167467368,17],[704725277,18],[812651450,18],[413999641,17],[986801716,17],[888271664,15],[383963550,16],[169253812,18],[489809536,16],[919745627,14],[828988207,19],[874527827,16],[420587416,17],[269540869,18],[406991364,18],[55691246,14],[720733021,15],[367860448,20],[481485622,17],[618809583,19],[864587310,16],[852005523,18],[166244857,16],[431171752,16],[373866676,15],[644642173,18],[17552662,15],[599427656,18],[468175327,13],[556807265,14],[896510679,18],[66790086,16],[789548689,16],[328732304,17],[751363202,18],[849600546,14],[320090595,16],[33882977,17],[530779519,17],[317073340,16],[183084738,18],[41191704,16],[785452373,16],[135483697,17],[873347847,17],[702193759,15],[706620349,17],[777122960,17],[718392276,16],[489561862,19],[183090955,17],[112790229,21],[240856267,19],[849489344,18],[456612460,17],[467510959,18],[427782219,18],[331145316,15],[303148031,17],[221906714,18],[527271606,19],[979683981,17],[747439287,18],[306975095,14],[100686610,15],[107044363,17],[943165002,18],[53471920,17],[576217591,18]],"-32621679":[[494195628,8],[246460598,10],[359901872,12],[905403772,9],[899221568,9],[502420554,13],[865228442,9],[449626175,11],[727952819,14],[60402830,9],[862301185,10],[392912493,8],[555354327,9],[242487882,9],[351041154,11],[394416112,14],[930002487,8],[957535228,10],[220584268,10],[557969583,11],[806602234,10],[217421789,12],[164924703,12],[62375277,12],[614435440,18],[432897306,10],[167467379,10],[704725284,10],[810302261,10],[369922194,10],[888271672,11],[383963564,11],[741614006,11],[489809540,10],[919745630,9],[789313998,13],[874527841,10],[420587420,9],[269540871,14],[896474844,12],[683783293,12],[864587323,11],[852005526,11],[451942199,13],[431171755,12],[28221385,10],[644642174,12],[822576243,12],[151999867,15],[468175334,13],[556807265,12],[815709802,8],[595167594,11],[936938793,8],[427611537,10],[812357752,13],[811011340,11],[386672421,9],[1886401,12],[885809648,12],[128917440,10],[66993530,12],[501117463,11],[328744492,10],[861903938,10],[853992505,11],[850660488,13],[823034153,12],[57405412,12],[528079300,10],[778789331,11],[613309715,10],[619149958,12],[878916086,12],[635582690,13],[793267432,8],[311708490,12],[712735092,13],[11570438,11],[366922888,11],[139353059,12],[93770000,9],[225355950,14],[49498411,12],[553126451,13],[168457157,10],[502373365,11],[498654590,12],[387842129,12],[977983445,11],[693164859,10],[339865003,12],[86851021,12],[103481205,11],[475697415,12],[534638718,12],[306588526,11],[29574660,13],[95941797,14],[934887563,8]],"-31571037":[[517311755,26],[779553374,28],[359901871,27],[899221570,26],[502420527,27],[449626176,30],[632158869,24],[727952817,29],[60402821,25],[92967415,29],[474064842,28],[862301167,28],[353377831,25],[807000469,29],[394416113,26],[930002483,28],[957535225,28],[220584265,26],[806602223,27],[555140753,30],[614435428,27],[432897287,26],[167467373,27],[413999641,28],[888271667,26],[169253812,28],[919745629,29],[269540869,28],[406991364,27],[55691246,27],[367860448,25],[618809584,26],[852005525,26],[166244857,26],[431171753,27],[644642176,26],[599427658,30],[468175326,28],[896510679,26],[789548691,28],[328732305,26],[751363203,27],[849600546,29],[320090595,27],[317073341,30],[183084738,28],[785452374,26],[135483698,27],[873347847,27],[702193761,28],[706620349,29],[777122961,28],[718392278,29],[240856272,26],[456612465,25],[427782220,25],[303148030,28],[221906715,25],[747439289,28],[306975100,26],[53471921,30],[643388670,27],[449172186,28],[95432311,29],[743781982,29],[214212076,29],[917077155,23],[921731777,26],[807396776,26],[832044790,27],[959922774,26],[195434137,26],[817774877,30],[465215830,25],[724184585,29],[987252040,28],[748395571,28],[128051304,27],[713560156,28],[995418691,28],[298206261,26],[500104787,29],[327505702,28],[804050144,27],[210489351,27],[249622423,25],[605242028,29],[916847144,27],[489615150,27],[389519834,26],[25910993,26],[936327594,27],[663663556,27],[948880614,27],[124585124,26],[735132057,27],[267482548,29],[497287981,28],[260611082,26],[300224539,29]],"-23063370":[[905403769,53],[862301169,49],[353377832,51],[555354325,49],[394416111,58],[930002502,47],[704725281,52],[812651450,52],[383963553,50],[153460486,53],[618809593,52],[17552663,52],[507734877,49],[66790086,55],[328732306,55],[849600553,55],[530779519,52],[890488300,47],[785452373,55],[873347860,52],[706620353,51],[718392291,51],[36424329,47],[427782225,50],[527271623,49],[100686614,50],[53471925,53],[860295484,55],[903894887,52],[997751146,52],[835143260,52],[326322248,55],[918809263,55],[58946752,50],[829444108,51],[223377174,55],[703993395,52],[144871377,51],[112948847,53],[82193471,57],[924867734,45],[134304507,52],[394821700,53],[582914996,53],[150677643,48],[451190366,50],[145840982,55],[958007742,56],[59628712,51],[309601169,58],[443146433,53],[28904102,48],[877341465,50],[959791984,51],[131166824,49],[617070360,52],[668932993,54],[632932307,53],[787471914,53],[946160961,53],[567077424,53],[248818411,49],[352213066,53],[169308269,54],[943084913,52],[633833584,52],[365477293,54],[803515006,54],[457468470,55],[607805870,53],[691841471,54],[685846666,51],[675638116,53],[709090769,49],[632033195,52],[928740691,53],[652360139,51],[178138198,50],[165535619,54],[621059500,53],[410262497,50],[64014298,52],[101821984,51],[5659703,55],[608620694,51],[773105671,49],[267401874,51],[909148600,53],[630315743,56],[683593520,52],[825707296,50],[815371929,51],[43299718,54],[862078660,52],[659945430,51],[89629071,52],[768070323,53],[992029529,53],[95011306,55],[924718919,52]],"-5036954":[[102552640,72],[449626180,65],[474064859,76],[587896349,65],[627853245,67],[164924698,75],[895395975,81],[704725277,59],[812651453,80],[413999653,88],[369922207,72],[294736199,80],[618809582,81],[254652653,69],[166244866,79],[101569644,74],[644642177,72],[708835267,84],[556807278,66],[789548692,66],[849600559,83],[702193765,69],[777122970,81],[795347063,68],[471116735,75],[979683991,67],[8254389,74],[272913912,74],[362525292,66],[162393345,67],[172807403,81],[558476736,64],[571494482,65],[872674388,76],[384863439,88],[413509219,72],[362523474,74],[372434763,71],[627821545,72],[316188689,73],[7393117,76],[4120301,72],[731602805,72],[365768906,75],[907537316,67],[950423984,61],[930543114,85],[507628918,75],[864579496,72],[274484786,72],[710346282,69],[864116160,69],[157864893,86],[497026938,80],[953038604,71],[631524038,74],[743154892,67],[984207334,75],[274856151,64],[394085636,67],[543509435,78],[413544571,66],[518970594,69],[27015747,73],[5233153,80],[502758809,73],[789290589,72],[51557417,89],[253125449,77],[287892499,71],[340577086,69],[412868573,69],[599839530,69],[776990692,71],[459046024,72],[841499588,77],[40388534,78],[822649951,67],[850947593,74],[949176014,68],[671864448,70],[225703205,66],[201424649,75],[747569614,71],[43862275,60],[100881728,81],[72860646,66],[630118733,83],[445803719,76],[844759486,73],[306932945,80],[744347768,72],[105820459,77],[203620343,90],[21745885,80],[632991155,78],[464988566,73],[274445132,77],[502143291,72],[359240758,81]]}'
+  '{"31976334":[[517311749,19],[359901870,22],[905403767,17],[899221565,19],[502420529,18],[865228431,21],[449626174,22],[632158867,20],[727952816,18],[60402820,20],[92967415,20],[474064842,21],[555354324,19],[518441848,22],[807000469,18],[943221663,20],[394416111,20],[148813978,21],[930002483,21],[220584264,18],[132060070,20],[806602223,20],[217421779,19],[555140744,22],[432897287,20],[167467368,20],[704725278,21],[812651450,19],[413999644,17],[888271664,21],[169253812,22],[489809536,20],[919745632,23],[828988207,23],[874527827,18],[269540869,22],[55691246,21],[720733021,21],[367860449,19],[481485622,22],[618809582,19],[864587310,20],[852005523,21],[166244857,21],[431171753,17],[373866676,19],[644642175,23],[801569483,24],[17552663,18],[599427656,21],[468175326,20],[556807265,16],[66790086,20],[328732305,18],[751363202,17],[849600546,20],[320090598,21],[33882977,16],[317073340,19],[183084739,21],[41191705,20],[785452374,20],[873347848,20],[702193759,20],[706620350,24],[777122960,22],[718392276,21],[489561862,22],[183090954,21],[112790229,21],[240856267,21],[456612460,17],[467510959,20],[331145315,21],[303148030,20],[221906716,19],[527271607,19],[979683981,21],[747439287,23],[306975097,18],[100686610,21],[107044363,17],[943165002,21],[53471920,21],[567308900,19],[421749944,20],[776865233,22],[736255507,19],[990486276,20],[234540976,22],[963453212,22],[937442019,18],[862293920,23],[326322243,19],[890488284,20],[471040617,24],[742661439,22],[116439555,21],[61006409,19],[936383014,19]],"65462911":[[779553376,43],[359901870,46],[905403774,49],[899221566,46],[632158868,47],[60402821,43],[474064843,45],[862301166,49],[353377833,48],[555354327,42],[394416114,45],[148813979,47],[220584269,47],[217421786,48],[614435426,48],[167467372,41],[812651456,49],[828988211,47],[420587419,47],[55691246,46],[720733021,45],[481485625,46],[852005523,42],[166244865,47],[431171753,46],[373866683,44],[644642180,43],[17552665,45],[896510679,47],[751363202,43],[41191714,45],[785452377,45],[183090958,50],[456612462,45],[427782219,43],[331145315,46],[747439292,47],[100686611,47],[53471922,46],[493355776,45],[286655128,45],[956556884,45],[375794656,49],[845724001,45],[206308936,48],[359255220,47],[705709741,47],[895521904,47],[274893941,41],[935386165,44],[467635272,46],[271118085,43],[175432469,46],[212243109,50],[876589606,45],[692375533,47],[35744365,46],[471116723,47],[866599604,51],[747978519,44],[971348278,44],[65010629,45],[848417029,45],[2799698,39],[303593325,46],[712191755,40],[29220768,48],[669952966,46],[471046174,47],[33070403,43],[230436279,48],[628673029,48],[549635989,44],[393212983,46],[660179143,45],[890898956,40],[428069860,45],[81432315,48],[12491695,45],[417977601,44],[593386838,46],[726982392,48],[330154711,43],[564052589,48],[891709515,48],[437491342,47],[752815986,48],[680377114,49],[840609601,46],[279479088,49],[647584689,48],[287892493,47],[344020955,51],[176087773,47],[227631814,44],[677951094,48],[30051300,44],[775941680,47],[525176575,41],[395208169,47]],"91451568":[[779553371,12],[359901871,13],[865228431,11],[449626175,11],[632158866,15],[727952819,14],[60402820,14],[92967416,15],[474064842,9],[862301169,14],[353377834,11],[555354326,13],[518441848,10],[807000472,16],[943221663,14],[394416114,14],[148813978,14],[930002487,9],[957535220,12],[220584267,12],[132060075,12],[217421781,14],[164924694,14],[614435430,14],[432897289,12],[167467375,13],[704725279,10],[812651451,12],[413999641,14],[986801717,10],[383963550,12],[489809536,19],[919745630,9],[828988207,14],[874527829,12],[420587419,13],[269540869,14],[406991365,14],[720733021,15],[367860448,13],[481485622,13],[852005524,11],[166244859,8],[431171754,13],[644642174,12],[599427658,11],[468175326,16],[556807265,12],[896510683,14],[66790087,12],[849600547,10],[320090595,12],[33882977,15],[530779521,16],[317073340,14],[183084739,12],[135483700,14],[873347852,15],[702193759,10],[706620349,12],[777122962,10],[489561863,14],[183090953,12],[112790231,9],[240856267,12],[849489344,10],[467510963,11],[331145321,10],[221906714,13],[979683980,16],[747439288,12],[306975097,14],[100686610,15],[107044363,13],[943165002,11],[53471920,11],[969121356,15],[666781849,12],[710346279,12],[862259528,16],[658907162,10],[909148578,10],[196278083,13],[265947629,11],[588949489,14],[860228488,9],[79582321,12],[23167076,12],[619367357,14],[31163027,21],[738834808,14],[688234239,13],[269555220,11],[441841729,14],[82193470,17],[324834154,11],[729801146,10],[890053904,11],[792191692,13],[641379344,13]],"219749905":[[517311754,25],[779553370,23],[359901876,24],[905403770,22],[899221565,22],[865228433,18],[632158868,24],[727952817,23],[60402823,24],[92967417,19],[555354324,23],[518441849,22],[807000469,23],[394416115,18],[220584264,19],[164924694,22],[555140748,22],[614435426,24],[167467371,24],[704725278,18],[812651450,20],[169253813,21],[919745627,21],[406991364,20],[720733021,21],[481485623,26],[864587312,19],[166244858,19],[431171756,20],[373866677,27],[644642174,24],[17552666,18],[599427656,20],[468175328,23],[556807265,21],[896510692,22],[66790087,21],[328732322,23],[751363202,22],[33882980,24],[530779525,23],[317073340,23],[183084741,23],[785452373,25],[135483701,22],[702193763,24],[706620350,24],[777122961,21],[718392283,22],[849489355,21],[427782220,25],[303148032,22],[527271606,23],[747439288,20],[306975105,19],[943165002,24],[805973626,26],[559900113,25],[82829059,20],[371051852,23],[884931362,22],[125884505,19],[203334952,20],[533219758,24],[5714782,25],[19929148,22],[818644476,22],[650902999,19],[944496963,22],[325037154,21],[458128230,23],[336891429,22],[202173528,25],[156844913,25],[548727181,24],[173793340,24],[421344995,22],[465111487,20],[465706432,21],[555708526,26],[962147098,22],[520589567,23],[187940970,16],[423611055,22],[266149082,22],[347138641,24],[657214793,23],[746467382,24],[885504965,22],[668115596,27],[937305649,20],[806024690,21],[970870576,18],[408129797,19],[207868292,25],[311209102,20],[100881709,22],[749051105,25],[213475316,23],[858171746,21]],"222816842":[[905403769,55],[862301169,48],[351041161,49],[217421779,54],[555140757,57],[704725281,52],[369922196,52],[383963553,50],[55691264,54],[367860455,53],[166244862,54],[801569481,51],[17552666,54],[599427673,55],[556807267,51],[328732306,55],[849600553,55],[135483705,50],[702193764,51],[784546905,54],[303148039,54],[527271623,50],[100686629,53],[943165021,55],[594915256,50],[998120312,55],[215807321,56],[926211065,52],[191531400,50],[548255473,53],[871075993,54],[905346971,52],[710346285,53],[306975113,53],[210712379,59],[59720321,55],[800796213,54],[9356700,46],[628271029,53],[417569857,54],[397637729,53],[671303126,52],[558978558,55],[515182601,50],[777662259,50],[965484389,53],[669224446,55],[290392761,53],[368153383,54],[884931365,49],[921867444,55],[37022994,50],[3307457,52],[119542847,53],[639338148,49],[965976001,58],[719478961,52],[506709653,49],[535454388,56],[135036039,48],[322460838,57],[262869511,55],[659581665,52],[665871535,55],[173452184,51],[932525110,53],[895964651,52],[600407087,55],[939063725,53],[892361979,48],[325957035,55],[748960869,55],[764921864,54],[311682411,54],[985534857,50],[186667289,55],[821738694,55],[788111866,51],[22637663,56],[767692425,56],[300285042,53],[786262767,52],[836380026,55],[394039211,47],[49475103,53],[984211073,52],[656375824,53],[175445352,54],[844663296,55],[407603386,58],[208310235,55],[474693338,54],[162393345,50],[337723847,54],[385626491,53],[436406994,51],[715509072,53],[522437084,56],[283756623,52],[476944440,58]],"226958187":[[517311749,13],[779553368,14],[359901870,14],[905403767,15],[899221565,14],[502420527,13],[865228431,15],[449626174,16],[632158866,15],[727952816,13],[60402820,13],[92967414,17],[474064842,14],[862301166,16],[353377832,12],[555354324,16],[518441848,11],[807000468,13],[943221661,14],[394416111,16],[148813978,14],[930002483,13],[957535220,15],[220584264,11],[132060070,15],[806602223,14],[217421779,17],[164924693,15],[555140744,18],[614435426,13],[432897288,13],[167467368,15],[704725277,14],[812651449,12],[413999641,12],[986801716,16],[888271664,15],[383963550,14],[169253812,15],[489809536,17],[919745627,15],[828988207,16],[874527827,13],[420587416,14],[269540869,16],[406991364,14],[55691246,16],[720733021,13],[367860448,15],[481485622,12],[618809582,15],[864587310,14],[852005523,15],[166244857,16],[431171752,15],[373866675,15],[644642173,16],[801569480,12],[17552661,15],[599427656,15],[468175326,12],[556807265,17],[896510679,16],[66790086,15],[328732304,16],[751363202,13],[849600546,13],[320090595,14],[33882977,13],[530779519,16],[317073340,16],[183084738,15],[41191704,16],[785452373,15],[135483698,11],[873347847,14],[702193759,15],[706620349,16],[777122960,14],[718392276,13],[489561862,15],[183090953,13],[112790229,13],[240856267,15],[849489344,15],[456612460,13],[467510959,14],[427782219,14],[331145315,14],[303148030,17],[221906714,13],[527271606,14],[979683980,15],[747439287,11],[306975095,16],[100686610,15],[107044364,14],[943165002,13],[53471920,15],[335179851,14]],"239814608":[[517311750,6],[779553369,8],[359901872,11],[905403768,9],[899221566,6],[502420538,7],[865228433,7],[449626174,11],[632158870,10],[727952828,5],[60402820,8],[92967419,8],[474064855,11],[862301166,10],[353377837,7],[555354325,6],[518441853,9],[807000471,10],[943221662,8],[394416113,8],[148813978,6],[930002486,7],[957535224,6],[220584268,6],[132060071,5],[806602223,9],[217421783,7],[164924696,9],[555140758,10],[614435427,8],[432897307,11],[167467371,7],[704725285,10],[812651450,10],[413999643,6],[986801719,6],[888271665,6],[169253814,5],[489809539,11],[919745631,8],[828988213,7],[420587420,7],[406991364,11],[55691248,7],[720733026,7],[367860448,7],[481485623,10],[618809582,4],[864587320,6],[852005529,7],[166244871,6],[431171754,9],[373866679,5],[644642179,8],[801569480,6],[17552663,5],[599427657,10],[468175326,8],[556807266,6],[896510679,10],[66790086,8],[789548694,8],[328732314,9],[751363205,5],[849600546,9],[320090596,9],[33882977,7],[530779519,5],[317073344,8],[183084740,10],[41191704,8],[785452375,10],[135483699,7],[702193772,8],[706620350,7],[777122970,6],[718392280,8],[489561866,8],[183090968,9],[112790235,8],[240856271,9],[849489346,8],[456612473,8],[467510963,7],[427782219,8],[331145316,7],[303148035,10],[221906715,10],[527271612,6],[979683984,6],[747439290,5],[306975102,9],[100686611,9],[107044376,7],[943165007,7],[53471925,8],[112005979,9],[775446497,7],[809567040,6],[508431958,8]],"240865250":[[517311750,24],[359901870,23],[905403768,26],[899221566,24],[865228432,26],[449626174,25],[632158866,23],[727952816,24],[60402826,23],[92967414,23],[474064844,26],[862301166,25],[353377833,24],[555354324,25],[518441850,23],[943221663,23],[394416111,25],[148813978,27],[930002483,26],[957535220,23],[220584264,26],[132060071,26],[806602224,25],[217421779,25],[555140745,24],[614435426,26],[167467368,26],[812651449,22],[413999641,26],[888271664,24],[383963551,24],[828988207,25],[269540870,25],[406991365,25],[720733021,24],[481485623,28],[864587310,24],[852005527,23],[166244857,24],[373866675,23],[801569481,24],[17552666,24],[599427656,25],[556807265,25],[896510681,23],[66790086,25],[789548691,25],[751363202,24],[849600547,23],[33882977,23],[530779519,26],[183084738,23],[785452374,23],[135483698,23],[702193761,24],[718392277,21],[489561864,26],[183090955,26],[456612460,29],[467510960,22],[427782220,25],[331145317,27],[221906714,26],[747439290,28],[100686618,20],[943165002,25],[53471921,24],[301674070,25],[706640133,27],[866599604,25],[95896677,28],[359950547,27],[223080981,25],[79582322,24],[49216690,23],[366245939,25],[458128229,27],[242487866,27],[34234637,26],[899069763,25],[938108730,25],[706384425,24],[948226112,27],[511965087,25],[454768124,26],[912469055,22],[705321270,26],[125905687,24],[600538391,25],[614085913,25],[551732586,26],[101372962,25],[321946448,25],[625206644,27],[491418665,26],[734251823,25],[445093203,21],[409813391,23],[549973349,23],[321824715,24]],"304412621":[[517311749,17],[779553368,17],[359901871,15],[905403767,19],[899221565,20],[502420527,20],[865228431,16],[632158866,20],[727952816,17],[60402821,17],[92967414,19],[474064842,19],[862301166,17],[353377831,15],[555354324,18],[518441848,17],[807000469,20],[943221661,17],[148813978,18],[930002483,16],[957535220,20],[220584265,19],[806602223,18],[217421781,18],[614435426,18],[432897287,18],[167467369,20],[704725277,17],[413999641,18],[986801716,19],[888271664,18],[383963550,18],[169253813,19],[489809536,17],[919745627,18],[828988210,14],[874527827,17],[420587416,16],[269540869,20],[406991364,19],[55691246,15],[720733023,18],[367860448,18],[481485622,18],[618809583,20],[864587310,21],[852005523,21],[166244857,19],[431171752,20],[373866675,19],[644642173,18],[801569480,21],[17552662,16],[599427658,19],[468175328,20],[556807266,17],[66790087,19],[789548690,16],[328732305,18],[751363202,18],[849600546,20],[320090595,18],[33882979,19],[530779519,19],[317073340,20],[183084738,18],[41191704,17],[785452373,17],[135483697,17],[873347847,17],[777122960,20],[489561862,22],[183090954,19],[112790229,18],[240856267,18],[849489345,17],[456612460,17],[427782221,19],[331145315,17],[303148030,17],[221906714,19],[527271606,19],[979683981,17],[747439287,19],[306975095,19],[100686610,17],[107044363,17],[53471920,20],[643747234,19],[500060559,17],[991124157,16],[42658434,18],[779954357,20],[305269964,18],[735246590,18],[243288428,21],[95432311,17],[866599603,18],[654402732,20],[20367593,21]],"305556867":[[517311749,17],[779553368,19],[359901870,19],[905403767,20],[899221565,18],[502420527,21],[865228431,18],[632158866,21],[727952816,18],[60402820,19],[92967414,20],[474064842,20],[862301166,21],[353377831,19],[555354324,21],[518441848,16],[807000467,20],[943221662,18],[394416111,19],[148813978,18],[957535220,21],[220584264,20],[132060070,21],[806602223,18],[217421779,21],[164924693,19],[555140744,21],[614435426,18],[432897287,16],[167467368,20],[704725277,22],[812651449,21],[413999641,19],[986801716,22],[888271664,17],[383963550,18],[169253812,20],[489809536,20],[919745627,16],[828988209,17],[874527827,18],[420587416,18],[269540869,19],[406991364,19],[55691246,15],[720733021,19],[367860448,22],[481485622,19],[618809582,19],[864587311,19],[852005523,19],[431171752,21],[373866675,18],[644642173,18],[801569480,22],[17552661,17],[599427656,21],[468175326,19],[556807265,18],[896510680,19],[66790086,20],[789548689,20],[328732305,17],[751363203,20],[849600546,21],[320090595,17],[33882977,16],[317073340,24],[183084738,25],[41191704,19],[785452373,20],[135483697,19],[873347848,22],[702193759,21],[706620349,22],[777122960,21],[718392276,19],[489561862,21],[183090953,20],[240856267,20],[849489344,21],[456612460,19],[467510959,21],[427782219,21],[331145315,18],[303148030,21],[221906714,20],[527271606,17],[979683980,20],[747439287,20],[306975095,18],[100686610,19],[107044363,16],[943165002,19],[53471920,19],[405886192,17],[208063620,19],[30519807,20],[594078541,20],[103481204,18]],"363887855":[[517311750,6],[779553369,8],[359901871,10],[905403767,7],[865228431,6],[449626174,11],[632158869,11],[727952816,9],[60402820,7],[92967414,7],[474064842,10],[862301166,10],[353377833,9],[555354325,8],[518441848,7],[807000469,9],[943221661,9],[394416112,10],[148813978,7],[930002483,8],[957535220,9],[220584264,11],[132060070,10],[806602223,9],[217421782,9],[164924693,8],[555140744,12],[614435426,11],[432897287,8],[167467369,8],[704725280,8],[812651449,9],[413999641,9],[986801717,7],[888271664,9],[169253814,5],[489809536,8],[828988209,8],[874527828,6],[420587416,7],[406991364,11],[55691246,7],[720733021,9],[367860448,9],[481485622,7],[618809582,4],[864587310,8],[166244857,8],[431171752,8],[373866675,6],[644642173,8],[801569480,6],[17552662,10],[599427657,10],[468175326,8],[556807266,7],[896510679,10],[66790086,8],[328732305,10],[751363202,6],[320090596,9],[33882977,7],[530779519,5],[317073341,11],[41191704,7],[785452373,11],[135483697,8],[873347847,9],[702193759,7],[706620349,13],[777122960,14],[489561862,9],[183090956,7],[112790230,9],[456612461,11],[467510960,13],[427782222,6],[331145316,7],[303148035,10],[221906715,10],[527271612,7],[979683980,11],[747439289,7],[306975095,8],[100686610,11],[107044364,7],[943165002,10],[53471921,11],[71107439,8],[916161507,11],[21945869,8],[376231936,10],[721063485,8],[35595057,8],[416755066,10],[340900290,10],[800796210,11],[543808335,9],[610470729,7],[888893344,9]],"374972231":[[779553377,63],[905403774,65],[458128243,59],[738545686,65],[469250498,64],[120329970,59],[353377836,59],[555354336,60],[807000481,65],[962310698,62],[895395975,60],[251007671,62],[209086746,62],[703212775,60],[986801729,63],[812886551,62],[655650051,59],[441244713,62],[874527831,64],[420587428,62],[269540874,62],[685400728,61],[720733022,60],[481485638,62],[815581696,61],[644642181,64],[53980371,61],[666080563,63],[468175335,60],[556807267,66],[815709796,66],[435327457,62],[789548692,62],[46412911,60],[320090613,65],[33882996,61],[530779522,60],[183084744,59],[542813250,60],[873347847,58],[706620354,59],[255990100,58],[183090960,57],[456612463,65],[332067092,61],[221906725,66],[225923201,63],[979683991,58],[799372887,64],[226033155,67],[107044364,54],[241199285,58],[447369006,61],[54169439,62],[943972963,59],[604636360,63],[223421024,58],[503224317,62],[838387769,59],[473213756,61],[895082804,61],[541916032,60],[760623388,65],[892048661,61],[280423076,63],[478034690,65],[642243325,66],[379514965,61],[974657845,59],[500352402,63],[819009261,59],[42293115,62],[726392925,62],[608953624,64],[362904866,60],[864605982,62],[970934049,63],[66727715,62],[285878051,60],[151519476,64],[157753813,64],[284587699,63],[961836643,63],[858063318,61],[46745836,60],[327335749,62],[91099226,59],[555641019,63],[966780947,63],[845221767,65],[536791548,61],[239841998,58],[639672258,66],[861482978,58],[32288721,61],[608843598,62],[638307819,63],[460971467,57],[539373476,62],[212369408,62]],"494781908":[[779553371,63],[359901879,90],[899221565,100],[865228434,74],[449626182,101],[727952819,66],[60402824,77],[474064856,79],[148813985,74],[957535225,94],[627853239,141],[217421786,63],[986801720,76],[919745628,83],[420587420,58],[55691256,72],[720733022,82],[618809583,114],[852005534,69],[801569480,83],[17552661,76],[328732308,138],[751363210,85],[317073343,69],[183084743,107],[873347851,60],[489561872,79],[456612470,94],[303148031,94],[306975098,109],[100686610,58],[53471920,64],[615813431,80],[823363678,74],[880626252,60],[267930983,78],[827675738,69],[840745950,99],[5194965,77],[339517716,70],[449849673,94],[79619503,73],[890592470,110],[268365465,104],[680538360,76],[390042101,79],[481918044,81],[966394410,95],[179470473,91],[562489761,107],[288056135,130],[25085993,90],[768464201,93],[203474104,137],[2605514,98],[262728522,70],[939465647,83],[585690726,106],[999786411,116],[936489495,90],[739025986,59],[714375259,103],[353745231,76],[127921611,90],[848172387,120],[724927200,95],[965492895,90],[428069863,68],[59824355,87],[300177011,88],[606767488,82],[138826313,79],[133009131,75],[496895529,67],[183161573,86],[717986924,79],[643386304,84],[549572716,108],[753567094,77],[217117950,106],[580192953,82],[525255461,85],[16398980,67],[82736248,96],[971534157,107],[386476465,82],[337534327,87],[934027695,73],[678024427,64],[837628145,79],[412187125,112],[504736587,105],[676692723,75],[784239055,73],[752112484,79],[160120994,103],[432183331,77],[915042488,84],[591926650,65],[830714662,129]],"512250895":[[517311749,4],[779553368,5],[359901872,4],[905403769,5],[899221565,5],[502420528,4],[865228434,7],[449626177,4],[632158867,5],[60402821,3],[92967414,4],[474064842,7],[862301166,6],[353377837,4],[555354324,4],[518441848,4],[807000467,7],[943221661,5],[394416112,4],[148813978,4],[930002484,3],[957535223,3],[220584264,5],[132060071,4],[806602223,4],[217421779,6],[164924693,6],[555140744,5],[614435426,3],[432897288,3],[167467369,7],[704725279,5],[812651449,4],[413999645,4],[986801717,3],[888271664,5],[383963554,3],[169253813,4],[489809536,4],[919745627,5],[828988207,4],[874527827,3],[420587417,4],[269540871,5],[406991364,3],[55691246,4],[720733022,5],[367860448,3],[481485624,4],[618809582,4],[852005523,7],[166244863,4],[431171753,4],[373866675,3],[644642173,4],[801569481,6],[17552663,4],[599427658,3],[468175328,3],[896510679,7],[66790087,7],[789548689,3],[328732305,4],[751363203,4],[849600546,6],[320090595,7],[33882977,5],[530779520,3],[317073340,6],[183084739,4],[41191707,4],[785452376,5],[135483697,3],[873347849,4],[702193759,7],[706620350,5],[777122960,5],[718392279,6],[489561862,6],[183090954,3],[112790229,4],[240856269,5],[849489344,4],[456612460,3],[467510959,5],[427782219,4],[331145315,4],[303148030,7],[221906715,4],[527271608,5],[979683980,4],[747439287,4],[306975095,4],[100686610,3],[107044363,5],[943165003,7],[53471920,3],[993223207,3],[401741359,4],[842437860,4]],"532916128":[[779553369,34],[905403767,31],[502420528,36],[449626174,35],[727952816,34],[60402820,36],[92967414,32],[474064844,34],[353377831,34],[555354325,34],[807000468,36],[943221661,36],[394416111,37],[930002483,35],[957535222,34],[220584265,36],[132060070,37],[806602225,34],[217421779,34],[164924693,34],[432897290,35],[704725278,35],[812651449,33],[413999641,35],[986801720,33],[888271666,33],[383963553,33],[169253813,36],[919745629,33],[828988210,33],[874527827,34],[269540869,34],[406991365,35],[55691248,33],[720733021,35],[367860448,35],[618809583,36],[864587311,35],[852005524,37],[166244857,35],[373866676,31],[801569480,38],[17552661,28],[599427656,35],[468175326,34],[556807265,30],[789548689,33],[328732305,34],[751363202,33],[849600546,39],[320090596,33],[33882978,31],[530779519,32],[317073340,36],[183084738,38],[41191705,34],[785452373,33],[873347847,36],[777122960,36],[718392277,32],[489561862,34],[112790231,35],[240856267,33],[456612461,36],[467510960,34],[221906718,31],[100686611,36],[107044364,34],[53471920,36],[76236421,33],[969614397,36],[627585821,34],[77741398,35],[547726857,35],[877979090,35],[789450155,35],[66727700,34],[526876597,32],[258718442,35],[88551611,33],[153211493,32],[325227388,34],[13254489,35],[482070066,35],[278093623,34],[266784548,33],[720788590,35],[494195624,35],[409631100,33],[775446497,35],[283072971,32],[876956447,35],[112451665,33],[696945059,33],[166591232,34],[930551513,37],[926539590,32],[817389292,35],[62676673,34],[482759852,30]],"598143099":[[517311751,18],[779553368,15],[359901871,16],[905403767,18],[899221566,18],[502420527,21],[865228433,19],[449626175,18],[632158866,17],[727952817,22],[60402820,19],[92967414,18],[474064842,16],[862301167,15],[353377831,18],[555354324,17],[518441848,19],[807000467,18],[943221661,18],[394416111,20],[148813979,18],[930002483,14],[957535220,20],[220584265,18],[132060070,22],[806602223,18],[217421779,22],[164924693,20],[555140744,19],[614435427,14],[432897287,18],[167467368,16],[704725277,17],[812651453,20],[413999641,20],[986801717,17],[888271665,20],[383963550,20],[169253812,20],[489809536,19],[919745627,16],[828988207,17],[874527828,18],[420587416,17],[269540869,20],[406991364,21],[55691246,21],[720733022,16],[367860448,19],[481485622,15],[618809582,21],[864587310,17],[852005523,20],[166244857,20],[431171752,18],[373866675,16],[644642173,19],[801569480,21],[17552661,19],[599427656,21],[468175327,16],[556807266,16],[896510679,16],[66790086,19],[789548689,15],[328732304,18],[751363202,18],[849600546,19],[320090597,18],[33882979,15],[530779519,22],[317073340,20],[183084738,18],[41191704,14],[785452374,15],[135483698,16],[873347847,17],[702193760,20],[706620349,21],[777122961,17],[718392276,16],[489561862,22],[183090955,19],[112790229,20],[240856267,18],[849489344,17],[456612460,21],[467510959,20],[427782219,16],[331145315,17],[303148031,19],[221906714,17],[527271606,17],[979683981,14],[747439288,17],[306975096,18],[100686610,19],[107044363,19],[943165002,16],[53471920,18]],"636324142":[[517311749,4],[779553368,5],[359901870,5],[905403767,4],[899221565,6],[502420527,4],[865228431,6],[449626174,4],[632158867,5],[727952816,4],[60402820,7],[92967414,4],[474064842,7],[862301166,6],[353377832,5],[555354324,4],[518441848,4],[807000467,7],[943221661,5],[394416112,4],[148813978,4],[930002484,3],[957535220,4],[220584264,7],[132060071,4],[806602223,4],[217421779,7],[164924693,6],[555140744,5],[614435426,3],[432897287,6],[167467368,4],[704725277,5],[812651449,4],[413999641,6],[986801716,6],[888271664,5],[383963551,5],[169253812,4],[489809536,4],[919745627,5],[828988207,4],[874527827,3],[269540869,6],[406991364,3],[55691246,4],[720733021,8],[367860448,3],[481485622,4],[618809582,4],[864587310,6],[852005523,7],[166244857,6],[431171752,5],[373866675,3],[644642173,4],[801569480,4],[17552663,4],[599427657,9],[468175327,7],[556807266,6],[896510679,7],[66790086,4],[789548689,3],[328732304,4],[751363202,6],[849600546,6],[320090595,6],[33882977,6],[530779519,4],[317073340,7],[183084739,4],[41191704,6],[785452373,9],[135483697,3],[873347847,6],[702193759,6],[706620349,7],[777122960,5],[718392276,4],[489561862,6],[183090954,3],[112790229,4],[240856267,7],[849489344,4],[456612460,3],[467510959,5],[427782219,4],[331145315,4],[303148030,7],[221906714,6],[527271606,6],[979683980,4],[747439287,4],[306975095,4],[100686610,3],[107044363,5],[943165002,6],[53471920,3],[644859537,5]],"661504400":[[779553382,20],[359901876,24],[905403773,24],[899221577,24],[449626177,25],[632158867,23],[727952817,23],[521573372,22],[353377831,23],[555354339,23],[518441854,23],[807000477,26],[148813980,21],[957535224,21],[806602238,20],[299800229,24],[555140759,23],[614435427,20],[812651451,19],[413999646,22],[986801727,21],[383963552,21],[169253813,21],[489809536,21],[493112077,26],[828988219,23],[367860459,20],[618809594,23],[864587321,23],[801569483,21],[17552680,18],[599427656,20],[468175329,23],[556807265,19],[789548706,17],[751363202,23],[849600550,27],[320090601,26],[33882981,22],[702193763,24],[706620352,19],[718392285,22],[754031611,21],[427782238,19],[121801053,20],[303148035,23],[953023070,22],[882592124,22],[719812831,22],[622482417,24],[200618604,25],[555425550,21],[602354185,22],[336068240,22],[410514377,21],[899867591,22],[437491350,19],[581338027,25],[130861286,24],[604636354,24],[171928595,25],[454901889,26],[944823866,19],[120329985,20],[613612098,24],[140073126,26],[705797191,26],[317000445,24],[395624470,20],[787598935,23],[541795468,24],[739025987,24],[340691029,15],[522483102,24],[217008244,22],[588456911,23],[887779083,23],[982940617,23],[267482588,22],[792191694,20],[198914252,24],[662979154,24],[919322780,24],[194182701,22],[563965677,23],[585757025,14],[942995714,22],[340287741,23],[867093745,26],[784369148,21],[116240515,21],[57270765,24],[825513622,19],[712958440,22],[173421454,20],[896688588,20],[724927229,19],[639338153,26],[885504981,22],[151703714,19]],"684698259":[[215107856,30],[874890925,40],[15144563,27],[321449692,29],[767931295,27],[285204707,36],[569127029,42],[369070372,28],[928309510,30],[299269846,32],[864097863,32],[211354790,29],[409475649,25],[308976570,26],[747884705,23],[117120741,26],[594682794,33],[986801732,32],[11212415,31],[441244721,27],[455902176,38],[541338576,29],[406991374,37],[506709660,21],[417677359,32],[206210373,35],[442175641,34],[931790505,31],[422316083,34],[672077511,34],[612352458,38],[391931204,31],[777122969,26],[489561871,32],[494051496,42],[361737511,26],[208076200,29],[888122969,29],[769028152,25],[943165006,36],[127921611,35],[280869691,25],[721765,35],[170629443,30],[159420443,37],[93332416,27],[429786675,29],[901439199,27],[656043063,30],[305160306,34],[155767384,47],[230370925,44],[663053482,35],[268265801,30],[922679552,36],[214052793,36],[356475707,34],[857820464,40],[893897231,33],[39507825,29],[124187392,30],[648466910,37],[275788789,29],[674301997,24],[367939803,36],[395034521,37],[369165992,34],[515282470,31],[708763207,29],[586393264,33],[721304344,26],[300354990,29],[343127308,30],[71906975,31],[42658443,25],[953502943,32],[886921011,27],[442416900,36],[352254189,31],[988740343,31],[598632749,26],[466462693,31],[464489827,40],[125934226,40],[548122984,31],[931770622,36],[718963563,27],[111232209,35],[654618582,34],[461875420,32],[644789906,28],[560678964,28],[191185890,25],[937411389,27],[23783162,31],[741001013,28],[222401890,36],[744730609,36],[52019592,30],[471046187,28]],"718288442":[[517311749,17],[779553368,19],[359901870,18],[905403767,21],[899221565,18],[502420527,21],[865228431,18],[449626174,19],[632158866,21],[727952816,18],[60402820,19],[92967415,18],[474064842,19],[862301166,20],[353377831,17],[555354324,20],[518441848,14],[807000467,20],[943221662,18],[394416111,19],[148813978,17],[930002483,20],[957535220,21],[220584264,19],[132060071,20],[806602223,17],[217421779,19],[164924693,19],[555140744,18],[614435426,18],[432897287,16],[167467368,19],[704725277,20],[812651449,20],[413999641,18],[986801716,22],[888271664,16],[383963550,18],[169253812,20],[489809536,18],[919745627,16],[828988209,16],[874527827,17],[420587416,18],[269540869,19],[406991364,18],[55691246,15],[720733021,18],[367860448,22],[481485622,17],[618809582,19],[864587311,19],[852005523,17],[166244857,19],[431171752,19],[373866675,17],[644642173,18],[801569480,21],[17552661,15],[599427656,21],[468175326,16],[556807265,18],[896510680,18],[66790086,21],[789548689,19],[328732305,17],[751363203,20],[849600546,21],[320090595,16],[33882977,15],[530779519,21],[317073340,24],[183084738,25],[41191704,19],[785452373,20],[135483697,19],[873347847,19],[702193759,21],[706620349,22],[777122960,21],[718392276,17],[489561862,21],[183090953,20],[112790229,21],[240856267,20],[849489344,18],[456612460,19],[467510959,18],[427782219,19],[331145315,18],[303148030,21],[221906714,18],[527271606,16],[979683980,20],[747439287,17],[306975095,18],[100686610,18],[107044363,16],[943165002,19],[53471920,19]],"784481713":[[444588742,32],[426848623,38],[717672653,44],[370072895,40],[754652487,37],[733973323,38],[407474129,41],[973084493,36],[251731131,40],[452036023,45],[489809537,37],[269540869,32],[720733021,36],[367860449,31],[864587320,42],[852005533,36],[431171752,36],[66790086,40],[41191707,33],[873347850,39],[702193762,35],[718392277,34],[112790234,39],[240856270,35],[849489351,37],[527271608,34],[979683980,46],[306975096,37],[100686615,42],[151703706,36],[569521155,37],[866045478,39],[854498838,42],[330302319,35],[883892330,43],[383650849,49],[844663282,40],[271063516,39],[556271136,34],[788105733,33],[678681868,38],[316093570,29],[476558766,40],[442024482,41],[316188663,36],[529873334,44],[108163818,39],[44336560,51],[52925288,38],[424749135,40],[600047391,37],[910887089,45],[872783118,46],[567776389,35],[604152515,30],[793498911,44],[722431459,33],[126070018,41],[845201,38],[688798942,35],[294557189,35],[519361997,48],[198633822,39],[576938488,32],[86184707,32],[113973705,44],[502169118,38],[506252183,40],[667558086,35],[38178224,39],[194990487,38],[515175282,35],[263487889,40],[653460192,37],[81022692,36],[398431606,37],[515282457,35],[982893169,38],[332716030,32],[808107531,41],[667788774,39],[750827263,39],[126092915,38],[867541608,40],[556748464,47],[469636567,36],[371963542,44],[839020552,37],[431376531,38],[224683297,50],[611315822,38],[995417590,56],[535496757,35],[525255474,32],[787365318,42],[18937237,43],[590067929,41],[881036790,36],[534119098,34],[229205302,39]],"849132023":[[517311751,65],[359901870,72],[865228433,96],[632158867,71],[862301168,88],[943221661,77],[148813983,65],[164924694,64],[986801721,63],[406991364,57],[55691246,61],[618809582,72],[864587311,81],[852005529,54],[431171755,77],[373866675,58],[644642176,55],[328732305,75],[530779523,64],[183084738,71],[41191706,87],[135483700,52],[873347847,64],[702193762,55],[706620349,66],[777122960,65],[718392282,62],[489561862,63],[183090954,77],[112790231,67],[240856268,67],[456612461,81],[427782224,74],[331145317,62],[303148031,58],[979683982,85],[107044368,73],[992782212,65],[102084982,49],[920632345,64],[464884290,83],[963096032,88],[756235789,66],[868627370,74],[243890960,71],[758576727,74],[928515491,68],[764572418,66],[829504763,92],[331363144,72],[340900292,53],[299875116,70],[708331027,69],[889662352,79],[557969577,58],[10150482,58],[860032800,67],[754031610,69],[562737898,71],[657705590,64],[428096734,71],[3679607,90],[47584521,68],[295120906,75],[392147427,67],[762296235,64],[382223588,74],[886642647,83],[873405291,73],[401137995,66],[368332158,75],[317856290,92],[669251627,61],[217063481,74],[328043781,78],[857570388,76],[179197658,68],[375189226,63],[599129177,59],[844919951,76],[848172385,63],[702242421,79],[966822911,67],[126492333,54],[135175563,59],[128684669,79],[896931745,70],[946160958,53],[889391931,67],[869355856,59],[204067060,75],[907438053,65],[569915788,80],[331851285,81],[807510859,69],[913145571,72],[384863429,61],[347445332,69],[378145294,70],[182249841,69]],"878487799":[[517311750,6],[779553368,8],[359901870,6],[905403767,7],[899221565,9],[502420527,8],[865228431,5],[449626174,11],[632158866,8],[727952816,8],[60402820,8],[92967414,7],[474064842,9],[862301166,11],[353377831,9],[555354324,7],[518441848,6],[807000467,10],[943221661,9],[394416111,7],[148813978,7],[930002483,8],[957535220,9],[220584264,10],[132060070,9],[806602223,10],[217421779,8],[164924693,7],[555140744,11],[614435426,11],[432897287,7],[167467368,6],[704725277,10],[812651449,9],[413999641,9],[986801716,9],[888271664,10],[383963550,9],[169253812,10],[489809536,7],[919745627,10],[828988207,9],[874527828,7],[420587416,7],[269540869,11],[406991364,11],[55691246,7],[720733021,7],[367860448,8],[481485622,7],[618809582,4],[864587310,7],[852005523,9],[166244857,8],[431171752,8],[373866675,6],[644642173,8],[801569480,7],[17552662,10],[599427656,9],[468175326,8],[556807265,8],[896510679,11],[66790086,9],[789548689,10],[328732305,10],[751363202,5],[849600546,10],[320090595,10],[33882977,7],[530779519,5],[317073340,10],[183084738,10],[41191704,9],[785452373,11],[135483697,8],[873347847,7],[702193759,7],[706620349,10],[777122960,11],[718392276,9],[489561862,8],[183090953,9],[112790229,7],[240856267,11],[849489344,8],[456612460,8],[467510959,8],[427782219,8],[331145315,8],[303148030,10],[221906714,6],[527271606,12],[979683980,9],[747439287,8],[306975095,6],[100686610,10],[107044363,7],[943165002,9],[53471920,6]],"880938594":[[67514109,57],[779553377,63],[738545686,64],[469250498,64],[727952818,60],[788920159,60],[521573371,58],[353377836,58],[163364822,60],[807000481,65],[674990297,57],[167763230,62],[597209688,59],[251007667,58],[564310258,58],[209086746,62],[810302254,56],[151841915,60],[493112082,60],[595642275,62],[269540874,62],[685400728,59],[720733022,60],[175440176,60],[166244875,60],[431376529,64],[644642181,61],[468175335,60],[68469099,56],[462220619,57],[33882996,59],[985521853,58],[183084744,58],[41191706,63],[612352447,63],[768394376,62],[134829885,61],[456612463,65],[332067092,61],[583261546,60],[979683991,58],[747439287,61],[107044364,54],[53471934,62],[85473184,64],[923016651,61],[389434688,58],[15204743,61],[534730807,61],[701874348,59],[411913944,64],[93319944,59],[534872969,63],[581653012,66],[590815931,58],[894693233,62],[579038083,60],[416427426,63],[298652071,64],[509977847,64],[979905934,59],[696099136,61],[610695565,60],[877341477,59],[256558690,63],[244481742,56],[697558197,61],[391252651,63],[678279918,62],[287757023,60],[483767088,59],[240623837,62],[627794408,57],[215849667,60],[890117834,61],[406926795,65],[188757651,64],[362047812,62],[67668854,61],[613909967,61],[23400685,63],[267482571,61],[995156146,59],[395954211,58],[514782063,59],[819360392,60],[537355934,62],[889567427,65],[425668750,60],[961910530,61],[919501962,64],[19133214,64],[801385684,56],[980885838,61],[680766829,58],[20226450,66],[830480173,59],[343288763,59],[239841998,58],[816286873,61]],"908554960":[[517311751,38],[779553369,39],[359901870,48],[905403769,41],[899221567,38],[502420527,41],[449626174,39],[727952816,40],[92967414,33],[862301166,38],[555354324,37],[518441848,40],[807000469,36],[394416112,42],[148813986,35],[930002484,43],[220584265,37],[806602225,37],[217421784,34],[164924693,49],[555140746,41],[614435428,43],[432897287,40],[986801718,41],[489809536,41],[828988210,37],[269540869,32],[406991365,43],[55691246,38],[367860449,32],[852005525,40],[166244858,45],[801569480,31],[599427656,38],[468175326,35],[556807265,30],[896510679,37],[66790086,39],[789548689,40],[751363209,40],[320090595,34],[33882977,39],[183084739,47],[135483697,39],[702193759,40],[706620349,36],[718392277,34],[183090958,38],[112790233,40],[456612461,44],[427782219,38],[331145317,37],[303148031,41],[221906720,36],[527271607,33],[100686612,44],[107044363,49],[53471921,45],[330554621,39],[223036990,40],[123989685,41],[153724245,33],[929624958,42],[130910677,47],[454424584,45],[169760354,42],[20231231,38],[71733674,38],[124840228,41],[294123364,41],[391689748,41],[99637606,36],[699492136,35],[472532522,43],[785890281,38],[86922177,36],[412010851,39],[906075849,41],[842168691,41],[115511722,45],[65010628,38],[479926502,32],[598466087,40],[259330184,40],[930397852,40],[730574185,35],[2094708,38],[858569096,39],[10120574,44],[174541558,37],[71009783,44],[728636604,38],[102581850,43],[904188920,38],[82063573,47],[225128701,33],[486206376,41],[219512397,40],[739979742,37],[772566115,38]],"932844753":[[517311750,41],[449626177,36],[632158873,32],[60402822,33],[862301166,30],[555354324,32],[518441850,26],[943221666,32],[957535228,35],[220584266,32],[132060079,32],[164924696,34],[555140746,39],[614435434,34],[253945709,38],[888271677,29],[383963552,34],[828988211,35],[420587422,30],[406991369,41],[55691255,30],[367860448,31],[373866689,29],[17552661,29],[599427657,29],[789548691,38],[328732313,37],[33882978,34],[41191706,31],[702193762,35],[112790230,32],[331145315,33],[53471920,40],[901439194,40],[350093853,32],[901244076,35],[552306138,35],[582582270,29],[324834155,30],[554778654,32],[65059018,36],[934706545,37],[776723768,41],[820703947,36],[951358825,32],[614283731,39],[942359317,30],[950071294,33],[482716293,45],[308430142,33],[410514377,38],[936701038,32],[250859386,30],[899330584,34],[100382286,34],[669712197,37],[653956768,35],[618382938,31],[566928648,31],[821778829,33],[273748598,39],[543682292,32],[912252567,41],[303958748,33],[238614954,31],[655469339,37],[355148646,32],[170199693,34],[882471282,38],[287710957,33],[199683207,37],[787598932,36],[962383938,33],[880093775,35],[904400869,34],[35908040,38],[111893172,36],[638091725,36],[190328742,34],[586505687,27],[545813098,37],[447559489,34],[82412661,34],[525107442,38],[852479623,34],[983778531,32],[528476806,35],[498654575,36],[215980399,32],[174322226,31],[954353400,32],[981741263,36],[57077704,38],[722768665,29],[119795169,31],[862064961,37],[955300332,29],[425546342,30],[796988416,30],[639672250,35]],"1010042162":[[517311749,5],[779553368,5],[359901870,6],[905403767,4],[899221565,5],[502420527,4],[865228432,6],[449626174,4],[632158866,5],[727952816,5],[60402820,6],[92967414,4],[474064842,7],[862301166,6],[353377831,5],[555354324,5],[518441848,5],[807000467,7],[943221661,5],[394416112,5],[148813978,4],[930002484,5],[957535220,5],[220584264,6],[132060071,5],[806602223,5],[217421779,6],[164924693,6],[555140744,8],[614435426,3],[432897287,6],[167467368,4],[704725277,5],[812651449,5],[413999641,5],[986801716,6],[888271664,5],[383963551,5],[169253812,5],[489809536,4],[919745627,6],[828988207,4],[874527827,3],[420587416,5],[269540871,6],[406991364,5],[55691246,4],[720733021,6],[367860448,5],[481485622,4],[618809582,5],[864587310,6],[852005523,7],[166244857,6],[431171752,5],[373866675,5],[644642173,4],[801569480,4],[17552663,4],[599427657,6],[468175327,6],[556807266,6],[896510679,7],[66790086,4],[789548689,3],[328732304,4],[751363202,5],[849600546,6],[320090595,6],[33882977,5],[530779519,5],[317073340,6],[183084739,5],[41191704,5],[785452373,7],[135483697,5],[873347847,6],[702193761,4],[706620349,6],[777122960,8],[718392276,5],[489561862,6],[183090953,5],[112790229,5],[240856267,6],[849489344,4],[456612460,5],[467510959,8],[427782219,5],[331145315,4],[303148030,7],[221906714,6],[527271606,6],[979683980,5],[747439287,4],[306975095,5],[100686610,5],[107044363,6],[943165002,5],[53471920,5]],"1093125211":[[779553377,63],[905403774,66],[319444747,59],[727952818,61],[521573371,62],[353377836,63],[807000481,66],[674990297,62],[597209688,60],[895395975,59],[251007671,62],[703212775,62],[441244713,65],[837979015,62],[269540874,62],[896474842,62],[720733036,60],[522906890,59],[481485638,62],[166244875,63],[53980382,61],[666080563,63],[151999868,63],[435327457,65],[462220619,60],[183084744,60],[104175451,60],[873347847,59],[768394376,62],[718392279,60],[78601148,58],[134829885,61],[456612463,67],[332067092,61],[583261546,63],[301673207,61],[979683991,60],[107044364,60],[53471934,64],[930397870,63],[669561892,60],[60253177,59],[691403509,62],[109618053,65],[200913698,59],[819612914,61],[490772971,64],[599662237,64],[82829073,63],[836965837,61],[65217010,63],[486813788,61],[756443457,68],[150280976,63],[324095431,66],[77453359,60],[566916197,65],[556748466,63],[884871521,59],[552449962,63],[518932700,64],[669373848,59],[903784598,62],[734140531,65],[926896589,64],[822576267,63],[272564540,64],[725793996,61],[712894698,66],[968688820,62],[608092103,58],[579422910,63],[855777062,60],[754726738,62],[423512840,66],[716477647,60],[572794323,63],[893566528,64],[993164937,63],[286664593,63],[583284913,60],[677365648,67],[36707770,62],[449172186,58],[822355784,63],[251032007,66],[559318967,60],[667318465,64],[466173943,64],[583269542,67],[266149088,63],[823442065,65],[224683300,64],[408715193,66],[60972026,58],[115073195,61],[425762192,57],[730472217,62],[774460101,63],[727159790,61]],"1095935398":[[517311750,6],[779553368,11],[359901870,8],[905403767,9],[899221565,11],[502420527,10],[865228431,8],[449626174,11],[632158866,9],[727952816,10],[60402820,9],[92967414,9],[474064842,9],[862301166,11],[353377831,11],[555354324,7],[518441848,10],[807000467,11],[943221661,9],[394416111,7],[148813978,8],[930002483,9],[957535220,10],[220584264,10],[132060070,9],[806602223,11],[217421779,8],[164924693,8],[555140744,11],[614435426,11],[432897287,7],[167467368,10],[704725277,10],[812651449,10],[413999641,10],[986801716,10],[888271664,10],[383963550,10],[169253812,10],[489809536,8],[919745627,10],[828988207,10],[874527828,8],[420587416,8],[269540869,11],[406991364,11],[55691246,10],[720733021,11],[367860448,8],[481485622,10],[618809582,8],[864587310,7],[852005523,10],[166244857,8],[431171752,8],[373866675,9],[644642173,8],[801569480,7],[17552662,10],[599427656,9],[468175326,10],[556807265,10],[896510679,11],[66790086,10],[789548689,10],[328732305,11],[751363202,6],[849600546,10],[320090595,12],[33882977,9],[530779519,5],[317073340,11],[183084738,10],[41191704,11],[785452373,11],[135483697,10],[873347847,9],[702193759,8],[706620349,10],[777122960,11],[718392276,10],[489561862,9],[183090953,11],[112790229,8],[240856267,11],[849489344,10],[456612460,10],[467510959,10],[427782219,8],[331145315,9],[303148030,11],[221906714,8],[527271606,12],[979683980,9],[747439287,8],[306975095,6],[100686610,10],[107044363,7],[943165002,10],[53471920,7]],"1137989174":[[779553368,87],[359901873,53],[905403767,90],[899221565,134],[502420527,93],[865228432,77],[449626174,107],[632158867,88],[727952818,93],[60402821,70],[92967414,87],[353377832,112],[807000467,96],[943221662,93],[394416111,88],[148813978,82],[930002483,96],[957535224,96],[220584264,90],[806602225,68],[217421782,66],[555140744,93],[614435427,71],[432897287,97],[167467370,142],[704725277,75],[812651449,102],[986801716,97],[888271664,80],[383963550,72],[489809536,125],[919745627,97],[828988207,90],[874527827,135],[420587416,86],[269540869,65],[406991365,91],[720733022,83],[367860451,69],[618809583,75],[864587310,84],[852005524,79],[166244858,80],[431171754,59],[373866677,63],[644642173,60],[801569481,71],[17552662,65],[599427656,87],[468175326,104],[556807265,70],[896510679,72],[66790086,90],[328732307,93],[751363203,80],[849600547,108],[320090596,69],[530779520,90],[183084739,94],[41191704,80],[785452378,90],[135483697,64],[702193759,92],[706620351,74],[777122960,79],[489561863,92],[183090953,76],[849489344,96],[456612460,83],[467510959,79],[427782219,83],[331145315,85],[303148030,91],[221906714,77],[979683980,98],[306975095,79],[100686610,90],[107044364,66],[943165002,72],[53471920,75],[449676409,119],[633672784,88],[885192267,94],[992905746,65],[341313291,93],[368153383,73],[713560156,71],[477344425,70],[829930232,104],[173061358,61],[268265774,93],[170755447,101],[169058864,100],[109618042,81],[353907518,73],[476101724,78],[963979607,111],[228789490,112],[729874057,77],[578778123,75]],"1325957594":[[517311750,31],[779553368,28],[359901871,28],[905403767,30],[899221565,26],[502420527,26],[865228431,29],[449626174,30],[632158866,26],[727952816,23],[60402822,28],[92967414,29],[474064842,30],[862301166,27],[353377831,29],[555354324,23],[518441848,25],[807000467,31],[943221661,30],[394416111,31],[148813978,24],[930002483,30],[957535220,31],[220584264,26],[132060070,27],[806602223,28],[217421779,26],[164924693,23],[555140744,32],[614435426,25],[167467368,24],[704725277,28],[812651449,27],[413999641,27],[986801716,23],[888271664,30],[383963550,31],[169253812,28],[489809536,33],[919745627,25],[828988207,27],[874527827,28],[420587416,26],[269540869,31],[406991364,25],[55691246,29],[367860448,23],[481485622,31],[618809582,27],[864587310,30],[852005523,28],[166244858,31],[431171752,30],[373866675,26],[644642174,27],[801569480,29],[17552662,30],[599427657,29],[468175327,28],[556807265,23],[896510679,24],[66790086,30],[789548690,28],[328732304,25],[751363202,25],[849600546,27],[320090595,21],[33882977,30],[530779519,28],[317073341,31],[183084738,26],[41191705,28],[785452373,28],[873347847,28],[702193759,28],[706620349,28],[777122961,29],[718392277,29],[489561862,28],[183090954,25],[112790229,30],[240856267,26],[849489344,31],[456612460,29],[467510959,32],[427782219,27],[331145316,25],[221906715,21],[527271606,26],[979683982,24],[747439290,26],[306975095,28],[100686610,27],[107044363,28],[943165003,29],[53471922,24],[274059747,33],[447571286,27],[715686840,29],[53342778,30]],"1344724498":[[517311750,8],[779553368,8],[359901870,8],[905403767,10],[899221565,10],[502420527,8],[865228431,8],[449626174,11],[632158866,8],[727952816,10],[60402820,8],[92967414,8],[474064842,10],[862301166,11],[353377831,9],[555354324,9],[518441848,8],[807000467,10],[943221661,9],[394416111,9],[148813978,8],[930002483,8],[957535220,10],[220584264,10],[132060070,9],[806602223,10],[217421779,10],[164924693,9],[555140744,11],[614435426,11],[432897287,9],[167467368,8],[704725277,10],[812651449,9],[413999641,9],[986801716,10],[888271664,10],[383963550,10],[169253812,10],[489809536,9],[919745627,11],[828988207,9],[874527828,7],[420587416,10],[269540869,11],[406991364,11],[55691246,8],[720733021,9],[367860448,9],[481485622,9],[618809582,8],[864587310,7],[852005523,9],[166244857,9],[431171752,9],[373866675,7],[644642173,9],[801569480,8],[17552662,10],[599427656,9],[468175326,10],[556807265,9],[896510679,11],[66790086,9],[789548689,11],[328732305,10],[751363202,5],[849600546,10],[320090595,10],[33882977,10],[530779519,8],[317073340,10],[183084738,11],[41191704,9],[785452373,11],[135483697,9],[873347847,9],[702193759,8],[706620349,11],[777122960,11],[718392276,10],[489561862,9],[183090953,9],[112790229,8],[240856267,11],[849489344,9],[456612460,11],[467510959,10],[427782219,10],[331145315,9],[303148030,10],[221906714,8],[527271606,12],[979683980,10],[747439287,8],[306975095,8],[100686610,10],[107044363,8],[943165002,9],[53471920,8]],"1426736851":[[449626181,52],[474064856,52],[862301169,49],[216391707,54],[930002502,47],[217421779,52],[704725281,52],[812651451,45],[413999646,53],[369922196,52],[383963553,48],[406991367,52],[166244862,53],[801569481,50],[507734877,48],[789548690,48],[328732306,55],[849600553,54],[135483699,50],[832015651,51],[718392291,52],[183090957,54],[36424325,47],[503069731,50],[427782225,50],[303148039,52],[527271623,50],[100686629,52],[943165021,50],[53471925,53],[501944760,50],[593020247,54],[400839518,55],[488889245,48],[963979621,54],[967868685,51],[864373136,50],[845723994,50],[668115602,55],[924660835,50],[909428422,54],[250408197,50],[384294284,54],[19954856,46],[410963079,54],[987668488,50],[441261793,54],[198914271,54],[547521379,55],[386230342,51],[101514759,51],[430797691,52],[974242448,50],[407453560,54],[510172280,52],[663460070,52],[531250199,54],[61030519,51],[999067067,48],[218676154,50],[792526380,50],[420785176,49],[828953373,51],[88551614,51],[659901596,51],[389292823,51],[780444115,49],[288223855,51],[547809975,49],[316268797,54],[986840020,56],[867521382,54],[165318385,54],[617928241,51],[951139191,53],[27741532,52],[525255460,50],[71693425,53],[185013282,54],[785592579,53],[730972461,48],[602356402,53],[606767513,54],[399659174,51],[104287445,51],[487986756,48],[900303988,51],[417781719,54],[317748061,50],[984863338,51],[513857776,47],[932015733,53],[960826537,55],[217970687,50],[810113472,53],[33844296,51],[721785434,48],[218789500,49],[889935757,52],[38566170,53]],"1462498410":[[517311755,104],[779553372,84],[359901871,86],[905403768,108],[632158876,71],[60402825,88],[92967414,79],[474064842,100],[862301168,95],[518441850,86],[807000470,127],[394416117,72],[148813983,92],[930002491,95],[220584265,84],[806602223,71],[217421783,94],[164924693,241],[555140746,121],[704725281,342],[986801717,78],[888271674,138],[383963550,154],[169253815,133],[919745629,97],[874527828,98],[55691246,85],[720733025,83],[481485626,78],[618809583,112],[431171766,269],[373866694,195],[644642177,83],[801569481,76],[556807267,117],[789548693,71],[328732305,111],[33882979,112],[317073345,145],[785452381,83],[135483701,79],[873347848,220],[706620354,192],[777122964,294],[718392292,83],[112790234,101],[456612463,95],[221906717,132],[979683991,87],[912752756,86],[306975099,121],[107044364,103],[943165011,85],[53471926,100],[969612070,86],[547490894,87],[184760491,89],[181302523,73],[498578150,90],[669251629,89],[360603265,88],[707728842,69],[296100002,103],[983378982,85],[294557183,104],[12131979,85],[15121746,117],[78969635,84],[313410661,73],[920241841,88],[195597694,101],[315603026,90],[919227772,100],[117808711,90],[364589997,87],[36141314,84],[257562188,85],[223862356,112],[385270696,102],[349114392,91],[584922802,99],[70838752,81],[414798895,88],[556748466,118],[528403606,94],[684646646,117],[59720297,99],[409813391,112],[187972126,107],[468783470,81],[604152514,89],[630352551,91],[126638089,90],[77961948,101],[922892672,93],[477704820,102],[958682209,86],[136256234,91],[342308551,111],[916693772,85]],"1796831549":[[517311750,31],[779553373,32],[359901871,28],[905403767,30],[899221566,31],[502420528,36],[449626174,32],[727952816,31],[60402820,35],[92967414,29],[353377831,34],[555354324,34],[807000467,33],[394416111,36],[930002483,35],[220584265,35],[132060070,37],[217421779,31],[164924693,33],[614435426,33],[432897290,33],[704725278,35],[812651449,33],[383963553,31],[169253813,34],[919745629,32],[828988210,34],[874527827,32],[269540869,34],[406991365,35],[55691247,35],[720733022,32],[367860448,35],[481485624,34],[618809582,31],[852005524,37],[166244857,35],[431171752,31],[373866676,31],[801569481,33],[17552661,28],[599427656,35],[468175326,34],[556807265,28],[896510680,31],[66790086,35],[789548689,33],[328732305,33],[751363202,31],[849600546,39],[320090596,30],[530779519,33],[317073340,33],[183084738,38],[41191705,34],[785452373,33],[135483698,29],[873347847,32],[702193760,34],[706620350,36],[718392278,35],[489561862,34],[112790231,35],[240856267,33],[456612461,34],[467510960,32],[331145316,37],[221906718,31],[747439288,33],[100686615,33],[107044364,33],[943165002,34],[53471920,33],[264112260,34],[908418556,30],[367649568,35],[592309991,32],[519337958,32],[950071282,33],[250408176,32],[340691025,34],[332716030,29],[450503175,33],[838989418,31],[112948850,31],[433811071,31],[251629618,34],[405886192,36],[116482902,33],[172844756,33],[127921611,33],[919244086,35],[322037030,34],[259106838,36],[453885337,32],[93324158,35],[20367593,36],[495950114,36],[13895907,32],[571940313,32]],"1838295370":[[905403785,55],[449626175,61],[474064843,60],[862301168,57],[555354338,58],[518441852,54],[943221678,52],[148813985,49],[741614014,52],[919745641,51],[828988216,46],[896510683,49],[530779521,50],[785452384,50],[873347862,60],[240856269,51],[849489351,47],[331145315,46],[221906718,55],[107044370,54],[366549824,58],[261278341,51],[757797305,62],[859673116,57],[155670414,63],[284635454,49],[687960475,57],[948104508,58],[827840070,48],[858253820,51],[250408182,56],[909558311,52],[781382258,59],[728369752,52],[705352481,53],[647584688,47],[190170926,42],[197527900,57],[639220113,60],[37120532,63],[279235968,56],[56552038,62],[257217217,52],[551469294,52],[813137173,55],[17866595,50],[208310239,62],[442033939,54],[87137659,50],[218789518,50],[202828937,48],[984630697,50],[477525961,49],[137992974,53],[697991961,46],[405895525,51],[398626655,60],[180021799,52],[375189246,49],[581934824,51],[395624457,59],[965484396,51],[927587342,48],[313410665,56],[65128627,61],[447135099,55],[394781473,50],[674746559,56],[585672808,53],[383587385,49],[850187141,55],[45920196,49],[158564799,64],[912469067,50],[535554697,64],[521473631,65],[177158286,53],[259769159,62],[46362480,54],[683955104,52],[916179288,60],[289291554,54],[795847151,59],[578823351,45],[96317197,57],[756235816,50],[71733672,57],[375928705,46],[287544596,48],[178138191,49],[951965722,53],[266801,48],[283374596,55],[587046139,56],[85513485,54],[850947576,60],[79917824,58],[156167447,60],[805436800,61],[160665092,53]],"1922600978":[[574448084,46],[555140746,53],[661991762,50],[888271676,52],[618809591,54],[801569486,53],[298206263,53],[320090600,49],[702193760,48],[331145316,56],[221906718,51],[903086587,60],[666781869,60],[967664391,70],[963148006,59],[913236699,50],[589061930,60],[512589205,56],[133286173,49],[711368071,60],[111301750,50],[25911045,49],[999067066,53],[115415301,58],[419955930,50],[691360391,64],[193349736,51],[5194983,61],[497441175,61],[4789969,53],[13232396,59],[599662295,53],[675149731,51],[424555325,55],[993725837,55],[373440642,52],[821991612,66],[417985877,49],[487279738,62],[109294597,54],[260160769,67],[641379396,57],[944832607,57],[359957157,64],[326589526,57],[27188444,56],[778663434,51],[456479713,53],[694782304,42],[68099156,54],[409725117,56],[461739064,60],[594000829,55],[485885399,60],[994544717,48],[163632575,51],[586662308,56],[475367287,64],[852332754,59],[632641070,53],[357207802,54],[354623612,49],[839078624,51],[747779688,67],[400973112,52],[980352621,53],[138929728,56],[819360381,50],[66596065,61],[180240774,55],[764176763,56],[274934912,53],[781729966,55],[495333072,53],[52077069,54],[144305469,51],[208071019,52],[157972357,59],[587531509,63],[5659701,46],[754178802,52],[9747458,48],[918814045,56],[947301264,51],[502137143,54],[395034507,53],[664624753,54],[630118730,54],[909322634,66],[819302618,58],[271501743,59],[210994230,59],[811494471,50],[613368025,52],[728529418,55],[14958271,56],[673191321,61],[155043609,54],[31434555,50],[982940625,53]],"2043144636":[[739021958,26],[497166222,28],[983863046,26],[160610517,24],[419587449,24],[262545614,30],[959740145,23],[330356691,28],[714335911,23],[497876383,25],[406889140,22],[216399590,26],[301897140,26],[178998688,27],[579849009,28],[601466122,28],[826139148,25],[59667572,26],[69857754,27],[843358494,27],[941633483,26],[331921400,24],[528301489,23],[296294644,24],[574283545,26],[575692621,26],[954004964,26],[262271594,29],[451482933,26],[759155596,25],[244171920,26],[25766886,23],[49006152,28],[6910600,26],[439606523,26],[438787101,25],[894037188,25],[509776461,26],[871125970,26],[119076848,26],[236103763,22],[501018918,21],[112825160,25],[99902826,26],[165032950,25],[886385105,27],[176144293,25],[13283489,26],[156946613,24],[427611537,25],[282870796,24],[981105294,26],[397686256,27],[367617432,26],[897290040,29],[630464579,29],[83169467,27],[999196513,23],[888450748,24],[533754345,29],[571595053,26],[799203659,23],[351360165,26],[110784286,24],[75546940,25],[23876198,24],[996109588,24],[392619906,25],[793267431,28],[766802807,23],[249289701,24],[995716746,24],[589937228,23],[859858058,28],[93769998,24],[722192911,27],[798703820,25],[680637220,28],[524346351,25],[746467381,27],[188349847,25],[396311667,24],[210585207,26],[481490977,26],[329039168,25],[502018966,28],[829664228,26],[245026874,27],[893897226,25],[749655052,24],[51433467,27],[931782242,26],[211517685,26],[873188823,22],[156387065,26],[776723750,27],[19951455,24],[838708609,27],[95544447,27],[610864075,25]],"2128103473":[[517311749,14],[779553368,15],[359901870,12],[905403767,14],[899221565,14],[502420527,13],[865228431,11],[449626174,13],[632158866,12],[727952816,11],[60402820,12],[92967414,12],[474064842,10],[862301166,14],[353377831,13],[555354324,13],[518441848,11],[807000467,14],[943221661,13],[394416111,12],[148813978,13],[930002483,12],[957535220,13],[220584264,11],[132060070,13],[806602223,11],[217421779,15],[164924693,15],[555140744,14],[614435426,16],[432897287,14],[167467368,10],[704725277,14],[812651449,13],[413999641,14],[986801716,14],[888271664,13],[383963550,14],[169253812,14],[489809536,14],[919745627,12],[828988207,14],[874527828,13],[420587416,15],[269540869,13],[406991364,13],[55691246,12],[720733021,12],[367860448,13],[481485622,12],[618809582,15],[864587310,13],[852005523,13],[166244857,16],[431171752,16],[373866675,15],[644642173,12],[801569480,12],[17552661,13],[599427656,13],[468175326,13],[556807265,12],[896510679,16],[66790086,14],[789548689,12],[328732304,12],[751363202,15],[849600546,13],[320090595,12],[33882977,14],[530779519,12],[317073340,14],[183084738,12],[41191704,13],[785452373,13],[135483697,12],[873347847,15],[702193759,11],[706620349,12],[777122960,13],[718392276,12],[489561862,12],[183090953,12],[112790229,13],[240856267,13],[849489344,12],[456612460,14],[467510959,10],[427782219,12],[331145315,11],[303148030,10],[221906714,12],[527271606,15],[979683980,14],[747439287,12],[306975095,13],[100686610,13],[107044363,13],[943165002,12],[53471920,12]],"-2091623303":[[517311749,16],[779553368,17],[359901870,13],[905403767,16],[899221565,15],[502420527,14],[865228431,14],[449626174,15],[632158866,15],[727952816,16],[60402820,14],[92967414,17],[474064842,13],[862301166,16],[353377831,13],[555354324,13],[518441848,12],[807000467,13],[943221661,20],[394416111,15],[148813978,15],[930002483,12],[957535220,17],[220584264,15],[132060070,18],[806602223,17],[217421779,14],[164924693,14],[555140744,15],[614435426,16],[432897287,13],[167467368,13],[704725277,15],[812651449,15],[413999641,15],[986801716,17],[888271664,17],[383963550,14],[169253812,14],[489809536,15],[919745627,16],[828988207,17],[874527827,14],[420587416,15],[269540869,16],[406991364,15],[55691246,14],[720733021,17],[367860448,16],[481485622,15],[618809582,17],[864587310,16],[852005523,13],[166244857,16],[431171752,19],[373866675,13],[644642173,14],[801569480,14],[17552661,17],[599427656,16],[468175326,12],[556807265,12],[896510679,18],[66790086,18],[789548689,18],[328732304,15],[751363202,14],[849600546,17],[320090595,18],[33882977,16],[530779519,17],[317073340,19],[183084738,15],[41191704,13],[785452373,15],[135483697,17],[873347847,18],[702193759,16],[706620349,18],[777122960,15],[718392276,15],[489561862,16],[183090953,15],[112790229,14],[240856267,18],[849489344,15],[456612460,16],[467510959,15],[427782219,16],[331145315,16],[303148030,16],[221906714,15],[527271606,15],[979683980,17],[747439287,14],[306975095,19],[100686610,15],[107044363,16],[943165002,18],[53471920,15]],"-2087777902":[[517311750,14],[779553368,16],[359901870,11],[905403767,16],[899221565,17],[502420528,17],[865228431,12],[449626174,16],[632158867,16],[727952816,16],[60402820,15],[92967414,14],[474064842,10],[862301166,17],[353377832,16],[555354324,16],[518441848,11],[807000467,14],[148813978,15],[930002483,13],[957535222,17],[220584265,14],[132060070,18],[806602223,13],[555140744,16],[614435426,14],[432897287,16],[167467368,10],[812651449,18],[413999641,16],[986801716,17],[888271665,13],[383963551,13],[169253814,15],[489809536,15],[919745627,15],[828988208,17],[874527828,13],[420587416,18],[269540870,15],[406991364,13],[55691246,10],[720733021,14],[367860448,13],[481485622,12],[618809582,21],[864587310,13],[852005523,15],[166244857,18],[431171752,15],[801569480,12],[599427656,12],[468175327,16],[556807266,14],[896510679,18],[328732304,13],[849600547,12],[320090595,14],[33882977,17],[530779520,16],[317073340,16],[183084738,14],[785452373,14],[873347849,13],[702193760,17],[706620350,12],[777122961,14],[718392276,13],[489561863,13],[183090953,14],[112790230,15],[240856267,14],[849489346,13],[456612463,16],[467510959,11],[427782219,13],[331145316,13],[303148031,15],[221906714,13],[527271606,13],[979683980,14],[747439287,13],[100686611,15],[107044363,15],[943165002,11],[53471920,11],[579986211,15],[405173294,15],[387546104,15],[301735349,17],[645362681,12],[23783154,14],[647277882,15],[162825210,16],[132427717,19],[174372584,17],[806903956,13],[765729344,14],[6935246,12],[672225767,17]],"-2043200314":[[517311749,7],[779553368,6],[359901870,5],[905403767,4],[899221565,5],[502420527,4],[865228432,5],[449626174,4],[632158866,5],[727952816,4],[60402820,6],[92967414,5],[474064842,7],[862301166,7],[353377831,6],[555354324,6],[518441848,4],[807000467,7],[943221661,6],[394416112,4],[148813978,5],[930002484,4],[957535220,4],[220584264,5],[132060071,4],[806602223,4],[217421779,7],[164924693,6],[555140744,6],[614435426,3],[432897287,5],[167467368,4],[704725277,6],[812651449,7],[413999641,4],[986801716,6],[888271664,6],[383963551,6],[169253812,4],[489809536,4],[919745627,7],[828988207,4],[874527827,3],[420587416,6],[269540871,6],[406991364,4],[55691246,5],[720733021,7],[367860448,4],[481485622,4],[618809582,6],[864587310,6],[852005523,7],[166244857,7],[431171752,4],[373866675,5],[644642173,4],[801569480,4],[17552663,4],[599427657,7],[468175327,5],[556807266,7],[896510679,7],[66790086,4],[789548689,3],[328732304,4],[751363202,4],[849600546,7],[320090595,5],[33882977,5],[530779519,4],[317073340,7],[183084739,6],[41191704,4],[785452373,7],[135483697,4],[873347847,5],[702193761,5],[706620349,6],[777122960,6],[718392276,4],[489561862,7],[183090953,5],[112790229,7],[240856267,6],[849489344,5],[456612460,5],[467510959,6],[427782219,4],[331145315,5],[303148030,7],[221906714,5],[527271606,6],[979683980,4],[747439287,5],[306975095,4],[100686610,5],[107044363,6],[943165002,5],[53471920,5]],"-2030835643":[[517311749,14],[779553368,15],[359901870,13],[905403767,14],[899221565,14],[502420527,14],[865228431,11],[449626174,13],[632158866,14],[727952816,12],[60402820,12],[92967414,13],[474064842,8],[862301166,14],[353377831,13],[555354324,13],[518441848,11],[807000467,14],[943221661,13],[394416111,13],[148813978,13],[930002483,12],[957535220,12],[220584264,11],[132060070,13],[806602223,12],[217421779,15],[164924693,15],[555140744,15],[614435426,13],[432897287,13],[167467368,9],[704725277,13],[812651449,13],[413999641,14],[986801716,14],[888271664,16],[383963550,13],[169253812,13],[489809536,14],[919745627,12],[828988207,13],[874527828,14],[420587416,15],[269540869,12],[406991364,12],[55691246,10],[720733021,12],[367860448,13],[481485622,12],[618809582,15],[864587310,14],[852005523,13],[166244857,16],[431171752,16],[373866675,14],[644642173,12],[801569480,12],[17552661,13],[599427656,13],[468175326,13],[556807265,12],[896510679,16],[66790086,14],[789548689,12],[328732304,11],[751363202,15],[849600546,15],[320090595,12],[33882977,14],[530779519,12],[317073340,14],[183084738,11],[41191704,13],[785452373,13],[135483697,12],[873347847,15],[702193759,11],[706620349,11],[777122960,13],[718392276,11],[489561862,13],[183090953,11],[112790229,13],[240856267,14],[849489344,11],[456612460,12],[467510959,14],[427782219,12],[331145315,11],[303148030,9],[221906714,14],[527271606,15],[979683980,14],[747439287,12],[306975095,11],[100686610,13],[107044363,12],[943165002,12],[53471920,13]],"-1963939495":[[517311750,51],[779553368,51],[359901870,40],[905403768,36],[899221565,40],[502420527,44],[865228431,32],[449626174,42],[632158866,35],[727952816,41],[60402820,45],[92967414,34],[474064842,40],[862301166,37],[353377831,49],[555354324,37],[518441848,46],[807000467,52],[943221663,50],[394416111,58],[930002483,36],[957535220,47],[220584264,37],[132060070,49],[806602223,41],[217421779,44],[164924693,36],[555140744,44],[614435426,45],[432897287,39],[167467368,61],[704725277,60],[812651449,38],[413999641,46],[986801716,45],[888271665,29],[489809536,38],[919745627,46],[828988207,39],[874527827,34],[420587416,35],[269540869,82],[406991364,30],[55691246,36],[720733021,47],[367860448,38],[481485623,53],[618809584,44],[864587310,41],[852005523,46],[166244859,53],[431171752,32],[373866675,39],[644642174,29],[801569481,48],[17552661,32],[599427656,32],[468175326,41],[556807265,44],[896510679,43],[66790086,44],[789548689,52],[328732304,46],[751363202,40],[849600546,39],[320090595,38],[33882977,39],[317073340,48],[183084738,47],[41191704,32],[785452373,59],[135483697,40],[873347847,68],[702193759,35],[706620349,46],[777122960,53],[718392276,36],[489561862,43],[240856267,47],[849489344,39],[456612460,43],[467510959,35],[427782219,35],[331145316,32],[303148030,37],[221906714,40],[527271607,38],[979683980,41],[747439287,39],[306975095,39],[100686610,39],[107044363,42],[53471920,46],[140279861,41],[308272188,35],[669871811,44],[667821012,38],[240098407,36],[799004814,33],[59907303,55]],"-1886763509":[[517311749,17],[779553368,17],[359901870,14],[905403767,15],[899221565,15],[502420527,14],[865228431,16],[449626174,17],[632158866,18],[727952816,16],[60402820,16],[92967414,17],[474064842,13],[862301166,18],[353377834,18],[555354325,14],[518441848,12],[807000467,14],[943221661,18],[394416111,15],[148813978,14],[930002483,12],[957535220,19],[220584264,15],[132060070,18],[806602223,17],[217421780,17],[164924693,14],[555140744,17],[614435426,16],[432897287,13],[167467368,13],[704725278,13],[812651449,15],[413999641,16],[986801716,16],[888271664,17],[383963550,14],[169253812,16],[489809536,18],[919745627,16],[828988207,17],[874527827,15],[420587416,15],[269540869,16],[406991364,15],[55691246,14],[720733022,16],[367860449,16],[481485622,14],[618809582,19],[864587310,15],[852005524,17],[166244857,16],[431171753,14],[373866675,13],[644642173,15],[801569480,15],[17552661,18],[599427656,17],[468175326,12],[556807265,12],[896510679,19],[66790087,15],[789548689,17],[328732304,15],[751363202,14],[849600547,16],[320090595,18],[33882977,16],[530779519,17],[317073340,19],[183084738,15],[41191704,13],[785452373,15],[135483697,17],[873347847,19],[702193760,18],[706620349,18],[777122960,16],[718392276,15],[489561862,19],[183090953,16],[112790229,15],[240856267,19],[849489344,16],[467510959,16],[427782219,14],[331145316,17],[303148030,16],[221906714,15],[527271606,16],[979683980,18],[747439287,15],[306975095,20],[100686610,14],[107044364,16],[943165002,18],[53471920,17],[36541547,15]],"-1842859596":[[636804787,61],[557518004,54],[178998688,63],[865123507,66],[860413498,56],[374952893,66],[509776464,55],[623472769,60],[501018924,55],[793003650,59],[922606325,59],[156946616,52],[293819728,63],[108389848,57],[701919062,62],[635000509,63],[183776958,57],[745257887,70],[658566216,58],[249289701,58],[995716747,58],[251731130,57],[575394997,60],[467100015,67],[421896165,64],[500392922,59],[617071733,62],[489910876,59],[123225447,66],[252367849,66],[665426895,58],[521388602,79],[656672371,58],[145742740,58],[978122940,52],[988062649,59],[413509226,60],[700939684,65],[583129468,66],[995651921,58],[260704423,71],[836326140,58],[718022518,61],[976588331,54],[821024301,57],[710692572,62],[31574997,67],[706384427,52],[484088820,63],[221653946,58],[982748995,56],[266654040,61],[626349874,68],[450690016,53],[297117038,63],[602433815,58],[136500454,66],[406580652,54],[383416465,57],[298000841,71],[699006280,63],[758301142,62],[415695300,61],[773077099,58],[950451953,58],[371413077,58],[821778836,69],[565024935,61],[712894702,59],[436246807,57],[459350740,62],[502209290,59],[716471189,67],[639186308,57],[778427524,58],[753158384,65],[566079476,67],[102552624,60],[91372710,56],[625088390,61],[262541079,69],[719478958,64],[583316430,59],[376924214,64],[701037263,52],[467626548,53],[764938559,69],[119542852,59],[411229084,58],[129985270,58],[697291113,74],[775670257,60],[518296450,66],[609207718,65],[890488281,61],[244845391,64],[194113241,59],[910145481,51],[899683545,61],[614283729,60]],"-1718786349":[[517311749,62],[779553368,64],[60402820,66],[862301169,56],[353377832,57],[943221661,68],[148813979,61],[806602224,62],[986801716,61],[919745627,61],[406991364,86],[367860449,66],[481485622,62],[618809587,56],[852005523,60],[166244858,55],[373866676,58],[801569481,59],[599427656,60],[468175326,66],[896510679,61],[66790088,57],[530779520,63],[317073341,63],[183084739,55],[785452373,57],[135483701,65],[873347847,61],[706620350,63],[718392276,60],[112790229,56],[849489345,64],[467510959,60],[221906715,61],[979683984,59],[747439296,65],[100686610,57],[107044367,59],[943165009,57],[648328478,67],[430252580,67],[924875341,64],[836326140,58],[132041730,56],[534476191,62],[162638416,66],[407344145,60],[239143493,62],[773874727,57],[216433755,56],[699929496,62],[792854393,64],[571940313,63],[116341035,54],[583129469,60],[226374901,58],[487938535,63],[849443398,69],[129252587,57],[842437862,54],[433853775,59],[594010329,73],[931782242,60],[217269494,60],[625654612,67],[797431565,62],[76236424,65],[116175115,62],[466894408,61],[383254750,72],[569408233,56],[938428908,57],[32754024,59],[930477774,62],[601133574,62],[333619811,63],[437852683,74],[350449209,61],[496807273,65],[426787269,66],[681281303,61],[56676707,58],[704329037,58],[411826145,62],[842074148,60],[441341219,59],[816612288,60],[790894563,73],[455033159,62],[783657285,64],[771028393,60],[724841705,59],[291524812,56],[153345552,73],[20027329,65],[698793357,61],[585186883,65],[95894036,61],[49909430,62],[358839311,66]],"-1590996086":[[779553374,49],[359901878,46],[905403774,49],[502420534,46],[865228431,46],[632158871,47],[862301166,49],[353377835,47],[555354327,43],[394416121,42],[148813979,47],[930002488,45],[220584269,47],[217421786,50],[164924704,47],[555140750,46],[614435426,48],[812651456,49],[888271665,45],[383963550,49],[489809540,48],[828988216,45],[55691246,46],[720733021,44],[852005523,44],[166244865,47],[431171753,46],[373866675,48],[17552661,47],[468175326,46],[789548692,45],[751363202,45],[317073347,45],[183084756,46],[777122961,47],[456612464,51],[427782229,49],[221906718,44],[100686611,47],[53471923,47],[752105934,49],[314381461,47],[275160667,46],[656672358,46],[60953000,44],[65851265,48],[774587611,45],[28978715,46],[529873329,46],[303085380,47],[120272625,47],[902208839,47],[859434553,45],[497108527,51],[830798819,47],[153724252,49],[76268807,47],[613926436,47],[7739089,48],[825006129,46],[228212647,48],[282636370,48],[983113755,44],[97678514,45],[360803237,49],[872783122,49],[840783155,50],[538193572,46],[684912694,45],[441315644,48],[528494856,44],[831440155,45],[875504859,47],[628008985,45],[21819171,47],[816542009,45],[423863459,45],[330116012,46],[901271210,51],[333619810,41],[482785622,45],[801798469,48],[806620853,49],[106986001,49],[960948877,47],[556235331,47],[397637732,47],[409203457,53],[179282187,46],[147796097,45],[999288411,47],[102627211,47],[397640342,47],[69197318,45],[992029521,43],[855777057,46],[643750204,46],[998048154,46],[505434713,45],[419393695,48]],"-1487240467":[[517311749,30],[779553368,31],[359901870,30],[905403767,30],[899221566,31],[502420528,30],[865228431,27],[632158866,30],[862301166,29],[353377832,27],[555354324,29],[518441849,29],[807000468,31],[943221661,29],[394416111,29],[148813978,30],[930002483,32],[957535220,29],[220584265,26],[132060070,33],[806602223,30],[217421780,27],[164924693,26],[555140744,29],[614435427,27],[432897287,25],[167467368,28],[704725277,30],[812651449,29],[413999641,32],[986801716,30],[383963550,31],[169253812,33],[489809536,30],[919745627,31],[874527827,29],[420587416,29],[269540869,27],[720733022,32],[367860448,30],[481485622,30],[864587310,29],[166244857,30],[431171752,28],[373866675,30],[644642173,28],[801569482,29],[17552661,25],[599427656,30],[556807266,27],[896510679,30],[789548689,31],[328732304,31],[751363202,32],[849600546,31],[320090596,31],[33882978,27],[530779519,31],[317073341,29],[183084739,29],[41191704,26],[785452373,31],[135483699,30],[702193759,30],[706620349,31],[777122961,28],[489561862,27],[183090953,27],[112790230,29],[240856267,33],[849489344,30],[456612460,33],[467510959,28],[427782219,29],[331145315,30],[303148032,29],[221906714,33],[527271606,31],[979683980,32],[747439287,31],[107044363,32],[53471920,31],[401057540,30],[501334050,29],[537036347,29],[232236497,28],[912866995,28],[961344131,29],[547590073,29],[764176667,29],[902070024,28],[581338021,29],[950146915,31],[802801644,31],[925301815,31],[226365099,30],[834885544,28],[620971585,31],[19742316,30],[213546098,26]],"-1383484848":[[517311749,17],[779553368,17],[359901870,15],[905403767,17],[899221565,15],[502420527,15],[865228431,14],[449626174,17],[632158866,16],[727952816,16],[60402820,14],[92967414,17],[474064842,13],[862301166,16],[353377831,17],[555354324,16],[518441848,14],[807000467,15],[943221661,20],[394416111,16],[148813978,15],[930002483,12],[957535220,17],[220584264,15],[132060070,18],[806602223,17],[217421779,16],[164924693,14],[555140744,17],[614435426,17],[432897287,15],[167467368,14],[704725277,17],[812651449,16],[413999641,15],[986801716,17],[888271664,17],[383963550,14],[169253812,15],[489809536,15],[919745627,16],[828988207,18],[874527827,17],[420587416,17],[269540869,16],[406991364,15],[55691246,14],[720733021,16],[481485622,17],[618809582,17],[864587310,19],[852005523,14],[166244857,17],[431171752,19],[373866675,14],[644642173,17],[801569480,14],[17552661,17],[599427656,16],[468175326,14],[556807265,15],[896510679,18],[66790086,18],[789548689,18],[328732304,15],[751363202,16],[849600546,17],[320090595,16],[33882977,17],[530779519,17],[317073340,19],[183084738,16],[41191704,14],[785452373,16],[135483697,17],[873347847,17],[702193759,17],[706620349,18],[777122960,18],[718392276,15],[489561862,16],[183090953,15],[112790229,17],[240856267,18],[849489344,15],[456612460,16],[467510959,15],[427782219,16],[331145315,16],[303148030,17],[221906714,15],[527271606,16],[979683980,17],[747439287,15],[306975095,19],[100686610,15],[107044363,16],[943165002,18],[53471920,16],[724317139,19]],"-1367843655":[[517311750,34],[779553368,30],[359901871,27],[905403767,32],[899221565,26],[502420527,27],[865228431,29],[449626174,29],[632158866,26],[727952816,25],[60402822,28],[92967414,26],[474064842,31],[862301166,28],[353377831,27],[555354324,24],[518441848,26],[807000467,30],[943221661,30],[394416111,30],[148813978,27],[930002483,31],[957535220,28],[220584264,27],[132060070,26],[806602223,28],[217421779,24],[164924693,22],[555140744,32],[614435426,26],[432897287,26],[167467368,25],[704725277,30],[812651449,25],[413999641,26],[986801716,24],[888271664,28],[383963550,31],[169253812,30],[489809536,30],[919745627,26],[828988207,32],[874527827,29],[420587416,28],[269540869,28],[406991364,25],[55691246,31],[720733022,28],[367860448,27],[481485622,35],[618809582,26],[864587310,31],[852005523,26],[166244858,28],[431171752,29],[373866675,27],[644642173,31],[801569480,29],[17552662,25],[599427656,29],[468175327,28],[556807265,24],[896510679,22],[66790086,29],[789548690,26],[328732304,28],[751363202,25],[849600546,26],[320090595,24],[33882977,28],[530779519,31],[317073341,30],[183084738,28],[41191705,29],[785452373,28],[135483697,29],[873347847,28],[702193759,28],[706620349,27],[777122961,27],[718392277,28],[489561862,28],[183090954,24],[112790229,29],[240856267,26],[849489344,27],[456612460,29],[467510959,27],[427782219,27],[331145316,27],[303148030,29],[221906714,26],[527271606,27],[979683980,30],[747439289,31],[306975095,28],[100686610,27],[107044363,29],[943165003,28],[53471922,23]],"-1278686646":[[517311749,16],[779553368,17],[359901870,13],[905403767,16],[899221565,15],[502420527,14],[865228431,14],[449626174,15],[632158866,15],[727952816,16],[60402820,14],[92967414,17],[474064842,13],[862301166,16],[353377831,13],[555354324,13],[518441848,12],[807000467,13],[943221661,20],[394416111,15],[148813978,15],[930002483,12],[957535220,17],[220584264,15],[132060070,18],[806602223,17],[217421779,14],[164924693,14],[555140744,15],[614435426,16],[432897287,13],[167467368,13],[704725277,15],[812651449,15],[413999641,15],[986801716,17],[888271664,17],[383963550,14],[169253812,14],[489809536,15],[919745627,16],[828988207,17],[874527827,14],[420587416,15],[269540869,16],[406991364,15],[55691246,14],[720733021,17],[367860448,16],[481485622,15],[618809582,17],[864587310,16],[852005523,13],[166244857,16],[431171752,19],[373866675,13],[644642173,14],[801569480,14],[17552661,17],[599427656,16],[468175326,12],[556807265,12],[896510679,18],[66790086,18],[789548689,18],[328732304,15],[751363202,14],[849600546,17],[320090595,18],[33882977,16],[530779519,17],[317073340,19],[183084738,15],[41191704,13],[785452373,15],[135483697,17],[873347847,18],[702193759,16],[706620349,18],[777122960,15],[718392276,15],[489561862,16],[183090953,15],[112790229,14],[240856267,18],[849489344,15],[456612460,16],[467510959,15],[427782219,16],[331145315,16],[303148030,16],[221906714,15],[527271606,15],[979683980,17],[747439287,14],[306975095,19],[100686610,15],[107044363,16],[943165002,18],[53471920,15]],"-1208358067":[[517311749,18],[779553368,21],[359901873,21],[905403767,22],[899221565,20],[502420527,21],[865228433,18],[449626174,22],[632158867,21],[727952818,20],[60402821,22],[92967414,24],[474064842,22],[862301166,23],[353377832,20],[518441848,23],[807000467,22],[943221661,20],[394416111,20],[148813978,22],[930002483,28],[957535220,21],[220584264,17],[132060070,20],[806602225,23],[217421780,21],[164924693,28],[555140744,20],[614435426,22],[432897287,18],[167467370,22],[704725277,24],[812651449,22],[413999641,18],[986801716,19],[888271664,20],[383963550,20],[169253812,23],[489809536,21],[919745627,21],[828988207,23],[874527827,17],[420587416,20],[269540869,24],[406991364,19],[55691246,21],[720733021,19],[367860450,20],[481485622,21],[618809582,21],[864587310,23],[852005523,25],[166244858,16],[431171754,22],[373866676,21],[644642173,22],[801569481,21],[17552661,20],[599427656,22],[468175326,19],[556807265,18],[896510679,21],[66790086,19],[789548689,22],[328732307,20],[751363202,20],[849600547,21],[320090596,19],[33882977,17],[530779519,22],[317073340,21],[183084738,24],[41191704,22],[785452378,18],[135483697,22],[873347847,22],[702193759,19],[706620350,21],[777122960,18],[489561863,21],[183090953,22],[112790229,21],[240856267,23],[849489344,21],[456612460,22],[467510959,24],[427782219,20],[331145315,20],[303148030,22],[221906714,20],[527271606,21],[979683980,24],[747439287,19],[306975095,23],[100686610,20],[107044363,19],[943165002,21],[53471920,24],[213092212,19],[776990685,19]],"-1176092191":[[779553368,61],[359901872,63],[905403767,59],[899221567,61],[502420531,61],[865228431,59],[632158866,61],[727952816,64],[92967415,53],[862301167,67],[807000470,64],[943221661,64],[394416112,70],[930002484,67],[957535222,63],[806602223,61],[217421780,70],[164924693,64],[432897289,61],[167467368,58],[704725277,79],[812651449,72],[888271664,75],[383963552,64],[489809536,61],[919745631,56],[874527827,55],[406991365,71],[55691246,63],[720733021,57],[367860449,68],[481485622,75],[864587310,57],[852005523,59],[166244857,58],[431171752,61],[644642174,59],[801569481,81],[17552662,64],[468175326,74],[556807265,57],[896510679,63],[66790086,57],[789548690,72],[849600546,61],[530779523,67],[317073341,60],[183084742,62],[41191704,64],[785452373,69],[135483697,57],[873347849,60],[706620350,63],[718392279,69],[489561862,65],[183090954,81],[240856273,69],[456612460,68],[427782220,77],[303148030,177],[221906714,56],[747439288,55],[306975096,60],[943165003,329],[53471923,72],[214329862,68],[522495705,63],[297649463,71],[522437065,64],[2843997,65],[193613731,78],[312635551,59],[868627365,62],[873547615,63],[688864707,74],[913840686,74],[490162379,67],[408129798,57],[882592116,67],[360145594,72],[742039743,68],[240856267,67],[695507488,59],[898650445,65],[759273922,66],[619622849,80],[701037259,57],[385106390,55],[131359383,56],[698086699,74],[456583092,63],[718545573,69],[47957928,66],[412314108,71],[457545402,74],[365451187,60],[274484782,77],[722483587,68],[615401133,52],[681751763,65]],"-1121316185":[[458128238,38],[865228432,38],[360468040,34],[92967430,32],[353377840,35],[673191325,37],[957535223,32],[220584264,34],[590049539,37],[164924700,36],[555140745,34],[614435436,36],[564310248,32],[167467376,36],[413999641,37],[383963550,38],[169253815,35],[919745627,33],[896474853,35],[55691263,38],[801569491,36],[17552662,38],[468175333,35],[896510681,37],[66790092,35],[33882985,38],[317073353,36],[183084742,38],[135483700,37],[873347850,36],[641892525,39],[364909525,35],[183090967,37],[456612469,36],[303148041,35],[527271613,34],[306975096,34],[107044363,31],[943165002,37],[53471929,35],[952101986,35],[148091173,37],[683605042,36],[244596027,38],[332846706,38],[297718260,38],[38237849,40],[980781946,36],[168437470,37],[286900055,37],[972565268,36],[877910720,33],[714907653,36],[43025626,33],[395016578,34],[802859797,37],[8062263,39],[447559488,38],[213207466,36],[45061290,39],[717986939,35],[586694077,38],[313410679,37],[807399935,35],[810620479,38],[518970568,31],[522160970,35],[928793264,34],[676944969,40],[379942012,36],[790768761,37],[959664214,38],[57572388,36],[473861469,34],[477704828,35],[958273256,36],[189378553,36],[978828586,37],[445093203,36],[921188771,34],[119290442,39],[362170643,37],[122936103,35],[120235089,36],[790355487,38],[751105777,36],[431628234,38],[414039224,37],[526785783,37],[24992897,37],[800378096,35],[433111230,33],[631767400,35],[882552231,35],[329169980,36],[325086300,39],[135618133,36],[554245784,37],[555103910,34],[213971428,36]],"-1112664773":[[517311751,14],[779553370,17],[359901870,12],[905403768,14],[899221565,13],[502420527,13],[865228431,16],[449626174,15],[632158867,12],[727952816,16],[60402822,19],[92967414,15],[474064842,13],[862301167,16],[353377832,12],[555354324,12],[518441850,15],[807000469,16],[943221661,12],[394416111,14],[148813979,16],[930002485,11],[957535220,13],[220584265,12],[132060070,17],[806602223,12],[217421779,13],[164924694,14],[555140745,10],[614435426,11],[432897289,10],[167467368,15],[704725277,17],[812651453,17],[413999644,14],[986801716,16],[888271664,14],[383963550,16],[169253815,13],[489809536,13],[919745627,13],[828988207,16],[874527828,19],[420587420,11],[269540869,15],[406991364,13],[55691247,13],[720733022,11],[367860448,15],[481485622,14],[618809585,13],[864587310,14],[852005523,18],[166244857,13],[431171752,15],[373866676,13],[644642176,14],[801569481,15],[17552664,14],[599427657,14],[468175327,17],[556807265,12],[896510679,17],[66790086,14],[789548690,15],[328732305,15],[751363203,12],[849600546,14],[320090597,17],[33882977,13],[530779519,14],[317073340,16],[183084738,17],[41191705,17],[785452373,16],[135483699,13],[873347847,12],[702193760,14],[706620349,16],[777122960,14],[718392276,10],[489561862,16],[183090954,14],[112790229,13],[240856267,15],[849489344,16],[456612460,15],[467510959,13],[427782219,12],[331145315,13],[303148030,17],[221906716,15],[527271608,16],[979683980,15],[747439288,13],[306975095,16],[100686610,15],[107044364,14],[943165002,16],[53471920,14]],"-1013080196":[[517311750,32],[779553368,30],[359901870,30],[905403767,30],[899221566,31],[502420528,30],[865228432,29],[632158866,30],[92967414,28],[474064843,30],[862301166,30],[353377832,27],[555354324,29],[518441849,29],[943221661,28],[394416111,29],[148813979,30],[930002483,32],[957535220,26],[220584265,23],[132060070,32],[806602223,31],[217421780,27],[164924693,24],[555140744,29],[432897287,25],[167467370,29],[704725277,30],[986801716,30],[383963550,26],[169253812,33],[489809536,30],[919745628,28],[874527827,29],[269540869,26],[720733021,27],[367860448,30],[481485622,26],[864587310,29],[852005527,28],[166244857,30],[431171752,27],[373866675,29],[644642173,28],[801569482,29],[17552661,25],[599427656,28],[556807265,28],[896510679,31],[66790087,29],[789548689,31],[751363202,31],[849600546,32],[33882978,27],[530779519,31],[317073341,29],[183084739,29],[41191704,27],[785452373,31],[702193759,29],[706620349,30],[777122961,28],[489561862,25],[183090953,27],[112790229,29],[240856267,30],[849489346,29],[456612461,28],[467510959,26],[427782219,29],[331145315,27],[303148030,26],[221906714,33],[979683980,32],[747439287,31],[100686611,31],[107044363,34],[53471920,31],[387546106,31],[74123423,27],[373001036,29],[156563090,32],[862833726,28],[919009785,30],[130157051,26],[71622838,28],[21506322,33],[577380490,29],[48597122,24],[380341319,22],[197991936,30],[467556033,28],[760926187,29],[244845377,28],[800644473,27],[717290121,29],[225703192,31],[68293990,24],[866983966,26],[288711922,29]],"-960771562":[[359901870,78],[899221565,80],[632158867,65],[727952817,65],[92967416,72],[474064842,67],[353377836,55],[518441849,76],[807000468,56],[220584265,67],[217421786,64],[555140745,76],[432897287,67],[704725289,82],[413999642,59],[986801725,53],[919745628,67],[874527828,78],[269540870,68],[720733022,63],[864587312,79],[852005526,52],[431171756,67],[373866682,53],[801569483,55],[556807268,59],[896510680,71],[66790087,62],[328732304,63],[751363209,72],[317073343,52],[183084741,59],[135483700,57],[873347848,73],[489561870,61],[849489347,70],[427782234,56],[303148031,74],[747439287,61],[100686610,60],[943165005,67],[53471921,53],[728606499,57],[858946078,91],[617071734,75],[510539916,77],[948880622,67],[283169047,68],[977497457,70],[25600217,58],[350438071,62],[95941797,78],[305269965,59],[2267590,63],[599904280,75],[835212292,52],[652800176,61],[518932701,67],[547726861,72],[840783155,73],[490562531,79],[138108209,74],[781931660,56],[30991808,89],[855131061,64],[182080071,109],[69074080,64],[636531838,55],[112948844,79],[389106445,87],[387803255,58],[251677693,70],[501746863,87],[835589882,79],[197879703,87],[311002989,69],[395527262,52],[669561880,59],[650828297,71],[552306137,67],[651317552,87],[719812818,84],[488302885,73],[209997934,82],[905803282,68],[384299638,64],[76908432,66],[752112481,95],[648664656,70],[351948887,63],[742661439,78],[613091597,61],[347935592,74],[797972596,82],[282942580,62],[283860031,56],[256558676,48],[820397259,68],[589508991,67],[408477780,86]],"-946946414":[[517311752,21],[779553371,21],[905403770,22],[899221565,22],[865228433,18],[449626177,24],[632158866,21],[92967417,23],[474064843,24],[555354324,21],[943221661,21],[394416115,18],[930002486,23],[220584264,19],[217421780,21],[812651450,19],[986801720,22],[888271664,23],[383963552,22],[169253816,24],[828988214,21],[269540870,23],[406991364,20],[481485623,26],[618809587,22],[864587312,19],[852005527,19],[373866677,22],[644642174,24],[599427656,20],[556807265,21],[896510686,24],[66790086,21],[789548694,25],[328732307,24],[751363202,22],[849600549,20],[33882981,21],[530779525,23],[317073344,23],[41191705,22],[785452373,25],[873347853,23],[706620350,24],[489561864,22],[112790229,20],[240856274,17],[331145322,21],[303148032,22],[221906715,18],[527271606,23],[100686612,21],[943165002,23],[53471920,27],[102084976,25],[638956012,21],[53791654,23],[617508216,21],[916371431,22],[534119098,21],[807687785,19],[524972883,27],[154224897,27],[497673485,18],[580772320,25],[430836949,23],[487282335,16],[208071024,24],[510368407,19],[558731826,27],[759656490,26],[551173720,22],[247577356,25],[862062812,20],[653210562,18],[670986167,22],[581162351,21],[510172254,24],[188102575,19],[818006594,22],[160344429,23],[965182451,24],[529563524,27],[74192850,21],[852011098,22],[149287815,25],[40388534,23],[670442454,22],[675402583,24],[468406901,22],[981572413,21],[299800222,23],[872296322,27],[854498839,19],[25533863,23],[723761268,21],[414590431,25],[490505348,27],[501322115,22],[38484307,23]],"-906730025":[[517311749,4],[779553368,5],[359901870,4],[905403767,3],[899221565,5],[502420527,3],[865228432,4],[449626174,3],[632158866,4],[727952816,3],[60402820,5],[92967414,4],[474064842,7],[862301166,6],[353377831,4],[555354324,4],[518441848,4],[807000467,7],[943221661,5],[394416112,4],[148813978,4],[930002484,3],[957535220,3],[220584264,5],[132060071,4],[806602223,4],[217421779,6],[164924693,6],[555140744,5],[614435426,3],[432897287,4],[167467368,3],[704725277,4],[812651449,4],[413999641,4],[986801716,5],[888271664,5],[383963551,4],[169253812,3],[489809536,4],[919745627,5],[828988207,4],[874527827,3],[420587416,4],[269540871,5],[406991364,3],[55691246,4],[720733021,6],[367860448,3],[481485622,3],[618809582,4],[864587310,5],[852005523,7],[166244857,5],[431171752,4],[373866675,3],[644642173,4],[801569480,3],[17552663,4],[599427657,6],[468175327,5],[556807266,6],[896510679,7],[66790086,3],[789548689,3],[328732304,3],[751363202,4],[849600546,6],[320090595,5],[33882977,5],[530779519,3],[317073340,6],[183084739,4],[41191704,4],[785452373,7],[135483697,3],[873347847,5],[702193761,4],[706620349,5],[777122960,5],[718392276,3],[489561862,6],[183090953,4],[112790229,4],[240856267,5],[849489344,4],[456612460,3],[467510959,5],[427782219,4],[331145315,4],[303148030,7],[221906714,4],[527271606,5],[979683980,4],[747439287,4],[306975095,4],[100686610,3],[107044363,5],[943165002,4],[53471920,3]],"-862639388":[[517311749,16],[779553368,17],[359901870,16],[905403767,16],[899221565,18],[502420527,18],[865228431,14],[449626174,15],[632158866,17],[727952816,16],[60402820,14],[92967414,17],[474064842,14],[862301166,16],[353377831,14],[555354324,15],[518441848,13],[807000467,15],[943221662,16],[394416111,16],[148813978,16],[930002483,12],[957535220,17],[220584264,15],[132060070,18],[806602223,19],[217421779,16],[164924693,14],[555140744,17],[614435426,19],[432897287,15],[167467368,15],[704725277,16],[812651449,15],[413999641,19],[986801716,17],[888271664,17],[383963550,17],[169253812,14],[489809536,15],[919745627,17],[828988207,19],[874527827,15],[420587416,17],[269540869,16],[406991364,17],[55691246,14],[720733021,16],[367860448,17],[481485622,15],[618809582,17],[864587310,16],[852005523,13],[166244857,16],[431171752,19],[373866675,16],[644642173,15],[801569480,15],[17552661,17],[599427656,16],[468175326,15],[556807265,14],[896510679,18],[66790086,18],[789548689,19],[328732304,15],[751363202,16],[849600546,17],[320090595,17],[33882977,19],[530779519,18],[183084738,16],[41191704,15],[785452373,15],[135483697,17],[873347847,18],[702193759,19],[706620349,18],[777122960,18],[718392276,15],[489561862,18],[183090953,16],[112790229,17],[240856267,18],[849489344,16],[456612460,16],[467510959,15],[427782219,16],[331145315,16],[303148030,16],[221906714,19],[527271606,15],[979683980,19],[747439287,17],[306975095,19],[100686610,18],[107044363,18],[943165002,18],[53471920,17],[777773891,16]],"-848879898":[[449626175,35],[862301171,33],[555354331,34],[518441849,37],[807000467,30],[394416116,31],[148813986,32],[220584265,34],[132060074,35],[806602225,32],[217421783,30],[432897298,31],[167467379,36],[812651450,33],[888271676,34],[919745634,35],[874527827,33],[153460468,32],[406991364,34],[720733022,32],[644642180,35],[17552670,35],[66790095,33],[789548692,33],[751363209,34],[320090601,33],[33882977,34],[530779521,32],[317073344,35],[873347849,35],[702193760,32],[706620353,34],[718392283,35],[489561863,33],[112790231,32],[467510964,34],[303148031,34],[221906720,35],[527271617,33],[979683980,37],[306975095,31],[100686617,37],[943165012,36],[626147845,35],[104109950,36],[917005602,33],[567468250,34],[910418400,33],[816386649,37],[903351711,33],[948880610,34],[703371393,37],[453718224,35],[822417361,33],[398431603,34],[918590137,33],[115020499,34],[480260373,34],[324591399,32],[391252634,30],[170685924,36],[47557949,35],[220484295,36],[936959693,36],[701469929,37],[862293931,35],[70770586,35],[622920253,31],[8052726,33],[909719398,33],[602433822,34],[168306041,34],[19929178,32],[252190293,36],[774460112,34],[595646018,32],[681059164,35],[703820688,34],[219806995,33],[767613365,34],[435327448,34],[760673598,33],[354580096,33],[453885338,35],[660184470,31],[307000527,32],[129841269,36],[449849670,32],[299018236,32],[160229167,35],[91378588,35],[555257759,36],[946814762,31],[353088465,36],[845201,36],[396796255,33],[347152806,34],[695852641,34],[667585750,32],[127921617,35]],"-774542587":[[865228442,76],[449626175,72],[92967425,71],[518441866,88],[943221669,80],[957535222,86],[555140755,97],[369922197,88],[888271676,131],[683783301,106],[431171767,69],[151519474,95],[749051104,102],[104109965,108],[718392293,74],[518097073,108],[953023078,95],[465447944,64],[101989990,132],[432959291,79],[368674597,94],[73130075,78],[775719695,86],[669679088,81],[134062100,86],[664449335,92],[387417120,129],[847785584,89],[162189257,83],[295540804,80],[452778338,76],[220484294,94],[260611107,63],[659981192,136],[42880734,79],[157046645,67],[85457894,99],[686226403,89],[981649426,91],[895109067,104],[962724775,123],[328919199,75],[564907499,132],[926663139,113],[416427424,93],[493568286,93],[337534347,81],[518970578,112],[843138033,73],[50396843,116],[261167478,85],[432611777,69],[864016081,75],[332852082,117],[916261007,94],[764510545,113],[497439651,78],[471533031,103],[761034853,116],[885296650,100],[879980147,89],[824390200,95],[297980167,96],[857820480,99],[346578104,97],[240098418,76],[496253343,101],[315607218,89],[359775116,96],[400685836,98],[66920154,90],[625088407,66],[247719809,92],[422316086,98],[379144913,93],[136761615,85],[832121092,90],[519436675,97],[629822835,83],[492898570,91],[861293420,92],[431824044,123],[590480531,95],[722046252,98],[541277345,121],[104532420,101],[296549854,80],[532220556,98],[507828881,90],[130896811,88],[455952044,109],[671798202,103],[914988538,78],[226830737,115],[481466556,105],[268528397,73],[126070543,101],[727135610,111],[218185505,86],[316467301,85]],"-740687052":[[965816938,68],[327077998,78],[126378691,70],[557884159,66],[53439153,69],[882377273,76],[866136349,70],[278657233,68],[352956391,74],[78095337,73],[692319744,74],[540877125,79],[167447682,79],[239281939,78],[96477139,81],[572499114,63],[653766508,73],[195211877,77],[490505426,81],[272394249,78],[369307756,75],[156108036,73],[723470018,71],[970883420,71],[322206457,69],[270933750,68],[26319511,66],[453284718,74],[679518113,70],[341566140,77],[540705363,63],[452357417,68],[438378475,72],[93590740,69],[591487607,79],[614085914,77],[473230111,84],[606530898,66],[362834579,80],[579422968,73],[331940866,64],[479954827,77],[353912869,78],[551841075,74],[736770728,76],[919232160,72],[876495672,63],[168306073,120],[439921511,79],[208347321,64],[274894019,78],[895165219,71],[710683592,69],[156712275,76],[93324179,79],[245842184,72],[8348050,74],[96522078,79],[266810,61],[518374930,72],[928117774,76],[946918817,81],[778148679,71],[204377608,82],[283277509,74],[794768784,91],[749987820,68],[123690510,67],[437088982,82],[695987156,83],[831440171,70],[285844971,72],[919502053,99],[962664936,63],[503055575,70],[121886794,62],[858063371,72],[63425496,67],[71313194,78],[263112355,76],[659594610,78],[314076756,74],[205732394,63],[661812973,69],[289150658,83],[747978517,62],[47321232,65],[123226331,71],[641645047,80],[769788026,83],[788666610,82],[431487576,80],[854760024,69],[471251637,75],[16049584,76],[166561137,73],[906168716,84],[433583172,67],[212972167,77],[463930445,80]],"-703861492":[[779553369,34],[905403767,30],[899221565,33],[502420528,36],[865228432,34],[449626174,34],[727952816,34],[60402820,35],[92967414,30],[862301166,36],[353377831,34],[555354324,34],[807000467,35],[394416111,36],[930002483,35],[957535222,34],[220584265,35],[132060070,37],[217421779,31],[164924693,35],[614435427,31],[432897290,34],[704725278,35],[812651449,33],[888271664,34],[383963555,34],[169253813,35],[919745629,34],[874527827,34],[420587420,35],[269540869,34],[406991365,35],[55691247,35],[720733021,35],[367860448,35],[481485624,36],[618809583,36],[864587311,35],[852005524,37],[373866676,33],[644642175,32],[801569480,31],[17552661,27],[599427656,35],[468175326,36],[789548689,34],[328732305,33],[751363202,33],[849600546,39],[320090596,31],[33882978,30],[530779519,33],[317073340,33],[183084738,38],[41191705,34],[785452373,33],[135483697,32],[702193760,35],[706620350,36],[777122960,36],[489561862,34],[240856267,34],[849489350,35],[456612461,34],[467510960,33],[221906718,30],[747439288,35],[100686615,33],[107044364,33],[53471920,36],[747782150,34],[877249320,36],[67514104,32],[33070402,33],[36840176,34],[707719092,36],[899330573,32],[587011656,30],[306565801,33],[555917120,34],[999128127,34],[913190620,36],[259163285,33],[690145023,31],[74196754,35],[742144792,34],[922892672,35],[311533573,34],[470994771,36],[25600215,37],[82206875,34],[36424315,34],[872353649,35],[483329965,32],[237796802,36],[129709033,35],[525095994,32],[275535093,34],[852912749,37],[800796209,35]],"-671914411":[[517311750,115],[779553368,61],[359901870,74],[899221565,88],[502420527,62],[865228432,76],[727952817,72],[60402825,78],[92967414,83],[474064842,74],[862301168,70],[353377831,52],[555354325,93],[518441848,118],[807000467,60],[943221662,71],[394416112,93],[148813978,64],[930002483,70],[132060071,121],[164924697,94],[614435426,159],[167467368,111],[812651449,236],[888271664,118],[383963550,121],[169253815,161],[489809536,140],[828988207,70],[420587417,105],[406991365,84],[720733022,112],[367860448,111],[618809583,121],[864587311,197],[852005524,125],[166244857,161],[431171753,304],[373866675,63],[644642173,134],[801569480,108],[468175327,74],[66790086,77],[789548689,94],[751363204,458],[849600547,120],[33882978,74],[530779520,201],[317073343,494],[183084739,148],[41191704,62],[785452373,321],[135483699,177],[702193759,78],[183090953,132],[112790231,148],[240856270,68],[456612461,100],[467510959,122],[303148030,139],[221906714,64],[527271610,86],[979683982,156],[747439287,149],[306975097,80],[100686611,96],[107044363,202],[943165005,129],[53471920,94],[81432310,67],[796581526,128],[177192686,125],[347935589,121],[220954068,103],[934688212,90],[267930979,60],[177158280,55],[292674913,76],[903784573,87],[217686313,124],[290396307,90],[430006424,125],[329494963,81],[312682995,76],[482031334,114],[16691991,85],[969121358,93],[535890799,98],[297718249,85],[245992207,87],[60900260,103],[929658630,87],[185665340,109],[628366297,73],[664126927,57],[473861471,80],[378820373,89],[664039,84],[902622972,91],[321824716,83]],"-651621524":[[449626182,84],[474064845,82],[253945714,111],[406991366,91],[19192784,100],[815709796,91],[789548696,96],[41191705,97],[135483711,100],[873347858,101],[768394370,99],[849489362,82],[456612467,85],[534197191,84],[303148045,86],[538279440,96],[979683981,88],[107044364,101],[531250188,97],[759818951,119],[82412683,74],[442714477,96],[20226445,132],[830556131,115],[842074150,122],[858876580,97],[744845719,96],[224565826,78],[699279748,125],[215980402,115],[328043794,82],[665625814,99],[86422799,123],[764431654,96],[913145572,125],[502824440,89],[962651556,112],[992029523,93],[251500790,138],[50476226,132],[890117837,101],[395034473,153],[449849685,104],[283277462,102],[715652977,106],[364590020,95],[171688406,115],[157593702,83],[690234350,100],[641399688,105],[577225277,125],[333518842,123],[234982527,89],[571494467,130],[235739702,106],[28468883,101],[471098535,107],[343161844,94],[141753493,132],[613592279,103],[297699279,115],[78770721,112],[295877771,103],[868379176,101],[837177591,78],[983530600,72],[798745818,73],[66554224,76],[661786809,87],[936260586,90],[504655627,101],[749834278,112],[143217031,142],[325037159,105],[396817769,136],[319165316,90],[570768024,97],[807867368,102],[924660835,89],[127733341,118],[996696941,86],[263808537,92],[242030195,81],[895521920,77],[269529671,98],[123558684,129],[775429750,91],[267482550,105],[78022030,110],[756431749,93],[903784585,110],[905803293,105],[880399980,96],[482785620,97],[969259033,89],[130239731,106],[515230766,92],[802082620,98],[450617657,95],[692375532,87]],"-635280127":[[517311750,34],[779553368,30],[359901870,30],[905403768,34],[899221566,31],[502420528,31],[632158866,31],[862301166,29],[353377832,28],[555354324,29],[518441849,29],[394416111,30],[148813979,32],[930002483,32],[220584265,28],[132060070,33],[806602223,31],[217421780,28],[164924693,24],[555140744,30],[432897287,27],[167467368,30],[704725277,31],[986801719,30],[383963550,32],[169253812,33],[919745628,29],[828988207,31],[874527827,29],[420587416,29],[269540871,31],[720733021,28],[367860448,31],[481485622,30],[618809583,28],[864587311,30],[852005527,28],[166244857,31],[431171752,27],[373866675,32],[644642173,32],[801569482,29],[17552661,25],[599427656,29],[896510679,31],[66790087,33],[789548689,31],[328732304,31],[751363202,31],[849600546,33],[320090595,29],[530779519,34],[317073340,32],[183084738,32],[41191704,28],[785452373,30],[702193759,30],[706620349,31],[489561863,30],[183090953,27],[112790230,30],[240856267,33],[849489346,31],[456612461,29],[467510959,29],[427782219,29],[331145315,25],[303148031,33],[221906714,33],[527271606,31],[979683980,32],[747439287,30],[107044363,34],[53471920,31],[374194555,30],[344228150,28],[30481107,29],[570354194,31],[287007624,31],[228805153,29],[993725837,33],[728903636,31],[421294475,29],[700228074,28],[754150179,27],[936489487,30],[381028156,29],[866045468,28],[977497457,29],[531130564,32],[809915733,30],[596096084,32],[524888190,30],[398490593,27],[964733867,30],[178417472,29],[281402301,32],[616274952,33],[863777159,30],[307644613,32]],"-606567752":[[517311749,13],[779553368,15],[359901870,12],[905403767,14],[899221565,14],[502420527,12],[865228431,11],[449626174,13],[632158866,12],[727952816,11],[60402820,12],[92967414,12],[474064842,8],[862301166,14],[353377831,10],[555354324,13],[518441848,11],[807000467,13],[943221661,13],[394416111,11],[148813978,13],[930002483,12],[957535220,12],[220584264,11],[132060070,13],[806602223,11],[217421779,12],[164924693,15],[555140744,14],[614435426,13],[432897287,12],[167467368,9],[704725277,13],[812651449,13],[413999641,14],[986801716,14],[888271664,13],[383963550,10],[169253812,12],[489809536,14],[919745627,12],[828988207,13],[874527828,12],[420587416,15],[269540869,12],[406991364,12],[55691246,9],[720733021,12],[367860448,13],[481485622,12],[618809582,15],[864587310,13],[852005523,12],[166244857,16],[431171752,13],[373866675,11],[644642173,12],[801569480,12],[17552661,14],[599427656,13],[468175326,13],[556807265,12],[896510679,16],[66790086,14],[789548689,12],[328732304,11],[751363202,15],[849600546,13],[320090595,11],[33882977,13],[530779519,9],[317073340,14],[183084738,11],[41191704,13],[785452373,12],[135483697,10],[873347847,14],[702193759,10],[706620349,10],[777122960,12],[718392276,10],[489561862,12],[183090953,11],[112790229,12],[240856267,12],[849489344,10],[456612460,12],[467510959,9],[427782219,12],[331145315,10],[303148030,9],[221906714,12],[527271606,14],[979683980,14],[747439287,11],[306975095,11],[100686610,13],[107044363,12],[943165002,12],[53471920,11]],"-600281204":[[517311749,18],[779553368,21],[359901870,13],[905403767,23],[899221565,21],[865228431,19],[449626174,24],[632158866,20],[727952817,21],[60402821,22],[92967414,23],[474064842,21],[862301166,23],[353377832,22],[555354324,21],[518441848,21],[807000467,23],[943221661,21],[394416111,19],[148813978,20],[957535220,20],[220584264,17],[132060070,19],[806602223,21],[217421779,18],[164924693,22],[555140744,22],[614435426,25],[432897287,19],[167467369,17],[704725277,23],[812651449,21],[413999642,15],[986801716,19],[888271664,20],[383963550,20],[169253812,23],[489809536,23],[919745628,21],[828988207,23],[874527827,17],[420587416,22],[269540869,23],[406991364,19],[55691246,21],[720733021,21],[367860448,22],[481485622,21],[618809582,21],[864587310,23],[852005523,24],[166244857,19],[431171752,21],[373866675,16],[644642173,22],[17552661,21],[599427656,19],[468175326,19],[556807265,18],[896510679,23],[789548689,23],[328732304,22],[751363202,20],[849600546,23],[320090595,18],[33882977,16],[530779519,22],[317073340,20],[183084738,25],[41191704,23],[785452373,21],[135483697,21],[873347847,23],[702193759,20],[706620349,20],[777122960,19],[718392277,23],[489561862,22],[183090953,20],[112790229,19],[240856267,24],[849489344,20],[456612460,22],[467510959,24],[427782219,21],[331145315,20],[303148031,24],[221906714,20],[527271606,22],[979683980,23],[747439287,19],[306975095,21],[100686611,20],[107044363,19],[943165002,22],[53471920,23],[279497004,22],[264350792,19],[446427122,22],[260577741,19]],"-600105873":[[517311749,19],[779553368,19],[359901870,20],[905403767,21],[899221565,19],[502420527,21],[865228431,18],[632158866,21],[727952816,18],[60402820,19],[92967414,20],[474064842,19],[862301166,20],[353377831,19],[555354324,21],[518441848,18],[807000467,20],[943221662,20],[394416111,20],[148813978,18],[930002483,20],[957535220,21],[220584264,21],[132060071,22],[806602223,17],[217421779,21],[164924693,19],[555140744,21],[614435426,18],[432897287,16],[167467368,21],[704725277,22],[812651449,20],[413999641,19],[986801716,22],[888271664,16],[383963550,18],[169253812,20],[489809536,18],[919745627,17],[828988209,16],[874527827,18],[420587416,20],[269540869,19],[406991364,19],[55691246,16],[720733021,19],[367860448,22],[481485622,17],[618809582,20],[864587311,22],[852005523,17],[166244857,19],[431171752,20],[373866675,18],[644642173,19],[801569480,21],[17552661,15],[599427656,21],[468175326,18],[556807265,20],[896510680,21],[66790086,22],[789548689,20],[328732304,21],[751363203,20],[849600546,21],[320090595,17],[33882977,17],[317073340,24],[41191704,19],[785452373,20],[135483697,19],[873347848,20],[702193759,21],[706620349,22],[777122960,21],[718392276,17],[489561862,21],[183090953,21],[112790229,20],[240856267,20],[849489344,23],[456612460,22],[467510959,21],[427782219,21],[331145315,20],[303148030,21],[221906714,20],[527271606,17],[979683980,20],[747439288,15],[306975095,18],[100686610,18],[107044363,16],[943165002,19],[53471920,19],[534231595,20],[406613771,19],[228762780,19]],"-580278930":[[517311749,70],[905403767,88],[899221567,71],[502420528,94],[632158867,75],[60402821,98],[92967414,81],[807000467,78],[943221669,70],[394416114,73],[217421780,81],[432897288,85],[167467368,83],[704725277,75],[812651451,79],[413999642,89],[986801718,78],[383963560,81],[828988208,79],[420587428,72],[269540869,79],[406991364,94],[720733021,78],[367860449,75],[166244857,87],[556807266,78],[896510680,99],[66790094,85],[849600548,85],[183084738,106],[41191705,76],[785452382,89],[135483700,97],[777122967,92],[183090954,84],[112790231,81],[240856268,86],[849489347,87],[427782221,101],[303148031,96],[979683985,71],[107044364,78],[943165003,84],[453718225,89],[286433979,93],[302814727,73],[354620069,80],[124267741,85],[586516324,86],[137352727,84],[147602271,75],[359950547,98],[87393951,93],[394159644,82],[916015453,82],[444903700,80],[701469941,86],[702811047,94],[592303853,85],[699929495,90],[607677663,80],[305632650,96],[989897535,85],[69942325,78],[363278510,76],[447821650,102],[806536939,88],[802762954,80],[66732405,86],[377692182,110],[195211804,76],[890610708,81],[40388534,91],[749597950,90],[781931663,87],[572815717,75],[227631813,85],[134965501,89],[57077703,81],[887613938,69],[813137172,90],[72305095,84],[566730821,71],[913400975,81],[848977500,93],[86585343,93],[744827989,99],[410296434,105],[742015317,78],[209291404,83],[524089351,74],[114805466,81],[59987059,79],[323222387,79],[355993083,93],[582582267,81],[376339386,78],[605357194,86],[958682204,77],[281178406,79]],"-577494253":[[392835183,18],[457543909,17],[925320773,18],[143128840,18],[656400547,17],[587416004,13],[165442984,19],[89636288,18],[300397497,16],[811248171,19],[391176713,18],[459302332,18],[42035479,16],[345065545,17],[623830278,17],[657520556,19],[648538419,18],[198372721,18],[480243772,13],[880878096,17],[642170784,15],[707894701,17],[569811292,18],[599830253,14],[173351994,16],[844544370,17],[878295489,18],[212077777,18],[740422798,18],[221390823,18],[886842141,18],[582843897,19],[9051630,19],[473953836,19],[802712618,15],[273883066,15],[365772992,19],[754254286,17],[778749561,14],[465843467,19],[974174162,14],[139598387,17],[317095158,19],[852017990,17],[551535310,21],[359951871,14],[338441589,17],[711648129,14],[940022951,16],[402791710,18],[835232915,14],[565881759,18],[443872639,16],[240188477,18],[815549207,15],[863008841,19],[404660566,20],[361812179,17],[401799942,16],[291816840,14],[442175636,18],[759905495,18],[95029266,14],[113199281,19],[536797890,19],[259743879,18],[255990098,16],[44706801,18],[49769267,14],[695507489,16],[534197201,17],[306565816,21],[879207655,16],[538279432,18],[509911281,18],[912752757,18],[226033149,16],[107044369,17],[331241237,17],[696348220,17],[473213763,20],[961804658,16],[857449024,17],[28558383,18],[291634886,17],[222774826,14],[464082631,21],[730327092,16],[952867642,19],[994508582,17],[889788747,19],[225443293,20],[256278020,20],[894502220,18],[600978,17],[929794436,18],[861199292,17],[713804243,22],[844115390,17],[770044344,16]],"-576443611":[[359901871,30],[502420530,29],[449626176,35],[632158868,33],[727952817,30],[92967433,30],[474064842,30],[862301171,32],[353377835,32],[518441848,33],[807000468,31],[394416124,36],[957535221,30],[217421788,32],[164924695,32],[614435430,30],[167467370,35],[812651452,32],[413999650,34],[888271674,29],[383963551,34],[169253812,30],[489809538,33],[420587419,32],[406991364,33],[367860451,34],[481485625,31],[618809595,29],[864587316,31],[852005528,32],[808107526,33],[431171756,31],[373866677,33],[468175328,32],[66790086,29],[328732309,32],[751363209,33],[183084747,31],[785452373,31],[702193759,34],[706620349,31],[777122962,31],[489561863,31],[112790230,33],[240856271,30],[849489352,33],[331145331,31],[303148032,33],[107044364,35],[943165004,30],[854642060,31],[907206708,32],[185352202,31],[891036564,31],[621166281,33],[517336138,32],[627794393,31],[544326376,27],[228212650,33],[991204665,35],[95513852,30],[676816793,29],[566666123,33],[743781993,28],[836965830,33],[473861474,30],[904400873,32],[912504855,34],[916161514,34],[969612068,31],[245992208,31],[432647893,33],[432127017,33],[230530576,32],[918018807,31],[16246502,31],[976036163,29],[715758088,33],[887657226,29],[18337641,33],[784435643,33],[664420315,30],[288530060,29],[421054985,31],[617704179,29],[267930979,31],[799764156,29],[386129553,32],[860353278,31],[211679073,30],[521388602,30],[2748082,28],[740188398,32],[965840288,31],[509265514,32],[141320235,32],[99781915,31],[956268403,28],[979825400,33],[157046638,33]],"-492540558":[[482302217,73],[733973323,71],[469827073,69],[151875369,69],[264046886,68],[501018920,75],[736996154,71],[648603882,70],[125643728,66],[282870800,64],[145780969,67],[986862083,74],[651701474,68],[964591227,69],[646998734,88],[82412667,66],[318375511,61],[340577080,75],[438965723,67],[954008882,77],[811433800,82],[910872697,66],[648437950,72],[100881712,76],[866323992,76],[795243583,72],[86658901,75],[273101542,74],[652746757,72],[372694741,67],[143830927,69],[895390801,75],[829949132,63],[529680405,66],[164130216,64],[626838921,70],[483767096,64],[13966671,71],[38421027,65],[716315414,85],[114949153,72],[609741739,82],[147450710,87],[354444028,76],[224542338,85],[85525580,78],[755558311,78],[546577358,64],[809665722,81],[300354937,77],[496717395,77],[463055381,75],[663663567,65],[620346133,74],[34890054,75],[772973238,57],[330832950,67],[827128537,68],[255301414,74],[844977067,69],[737320913,75],[255349370,77],[734574226,67],[417677399,67],[260322713,61],[218468994,78],[613072381,71],[596916069,65],[864681711,71],[951131502,76],[244829658,76],[761546027,80],[521880248,78],[527042517,76],[498799745,81],[22230503,75],[825549565,69],[120381649,75],[442563328,80],[879887445,83],[648793610,71],[890958172,77],[389790616,70],[235175072,81],[441074138,69],[487122350,64],[783768184,74],[505434723,70],[509767741,73],[351220913,64],[8062256,69],[366264984,75],[619738008,70],[696170976,71],[979905948,70],[820793969,66],[709504354,71],[714052679,72],[89908374,78],[432940670,78]],"-413861179":[[779553368,76],[353377833,63],[807000467,67],[943221678,77],[806602223,71],[164924693,64],[413999653,67],[169253813,65],[489809537,72],[420587427,70],[166244865,66],[644642187,64],[801569481,71],[328732306,68],[33882979,62],[785452377,70],[706620353,63],[832015652,68],[112790239,68],[575485113,67],[768000139,67],[107044364,59],[945096814,71],[517666965,75],[217008251,62],[405639139,71],[177776951,68],[912546418,74],[873333592,70],[944466490,71],[269625332,64],[369857999,65],[667961481,66],[391917630,68],[772056989,65],[541360865,71],[158169557,68],[330832959,67],[898587525,58],[223036995,76],[272424179,78],[881446496,63],[881036789,65],[147796091,67],[174846167,69],[889662352,71],[127058391,71],[59853102,69],[647277882,70],[58732437,66],[928344204,66],[297117046,67],[296100002,68],[173793351,68],[54436676,70],[617928238,69],[850819032,66],[30051300,64],[564379291,64],[417977598,63],[128261012,67],[580623554,70],[847726681,68],[632420968,69],[430821127,67],[305435361,70],[259395645,70],[765153128,70],[122188237,72],[82009811,66],[610470739,77],[291729271,64],[123225447,70],[803115751,72],[572333090,68],[147140308,77],[629545008,77],[490544655,65],[515324053,66],[936300630,70],[470365339,73],[127921613,80],[615969365,72],[592309990,68],[583316429,66],[972526998,68],[775391706,68],[807883653,67],[766009320,70],[656400767,65],[303155227,66],[690902129,76],[185013284,70],[610586946,77],[413544571,67],[179197664,68],[157972321,64],[535580900,72],[329151991,68],[48745485,66]],"-407920866":[[517311752,47],[905403774,48],[632158868,47],[92967414,46],[862301166,49],[353377833,46],[555354327,46],[807000468,47],[394416114,46],[148813979,47],[930002487,45],[220584269,49],[167467372,44],[704725281,46],[888271668,45],[383963550,49],[489809540,47],[874527827,50],[420587419,47],[720733021,48],[367860448,50],[618809587,47],[852005523,44],[431171756,49],[801569486,42],[599427657,46],[468175333,50],[556807276,49],[328732311,46],[751363202,49],[320090597,52],[530779520,51],[183084756,47],[41191717,45],[706620353,49],[489561865,49],[112790234,51],[331145315,48],[303148032,47],[221906718,46],[747439288,47],[306975099,50],[561521792,47],[842683491,45],[59508969,45],[317129019,42],[699164131,52],[807538516,48],[395542937,49],[156563090,45],[638724363,47],[868718391,47],[573698161,45],[282303430,45],[103221486,49],[197879701,48],[718989200,45],[805973614,48],[253945712,49],[213546099,44],[476755412,49],[895071988,42],[600538385,49],[709090745,46],[582254289,49],[250192741,48],[247559904,49],[507187854,50],[891696480,46],[367939804,46],[495254011,49],[849675978,46],[878009465,48],[988375256,47],[452225655,46],[56701646,48],[931770554,46],[304854023,48],[929518687,45],[745242371,47],[143919690,48],[441341226,46],[994633062,51],[990693752,46],[123558668,49],[667810182,43],[694782292,50],[901375987,49],[344228166,50],[59632329,45],[506184,50],[699751698,45],[897151357,48],[68187506,48],[738313535,48],[797972602,50],[968595162,47],[886825782,45],[576496949,48],[664420310,48]],"-384850278":[[517311749,25],[779553368,25],[359901872,22],[905403767,23],[899221565,23],[502420527,22],[865228432,24],[449626174,21],[632158866,23],[727952816,21],[60402820,21],[92967415,18],[474064842,27],[862301166,23],[353377831,18],[555354324,22],[518441848,23],[807000467,25],[943221661,23],[394416112,23],[148813979,24],[930002483,24],[957535222,22],[220584265,22],[132060070,19],[806602223,25],[217421779,25],[164924693,21],[555140744,22],[614435426,24],[432897288,23],[167467368,21],[704725277,24],[812651449,29],[413999641,22],[986801716,26],[888271664,23],[383963550,23],[169253814,21],[919745627,21],[828988207,24],[874527827,22],[420587418,24],[269540869,22],[406991367,20],[55691246,18],[720733022,25],[367860448,22],[481485622,21],[618809582,27],[864587310,23],[852005523,24],[166244857,29],[431171752,20],[373866675,20],[644642173,20],[801569480,24],[17552662,25],[599427658,20],[468175328,21],[556807265,16],[896510679,23],[66790086,22],[789548689,23],[328732305,22],[751363202,20],[849600546,20],[320090595,23],[33882977,23],[530779519,25],[317073340,21],[183084738,26],[41191704,22],[785452373,22],[135483697,25],[873347847,21],[702193759,23],[706620349,23],[777122960,27],[718392276,22],[489561862,24],[183090954,20],[112790229,21],[240856267,24],[849489344,21],[456612460,22],[467510959,25],[427782219,24],[331145316,23],[303148032,21],[221906714,20],[527271608,28],[979683980,25],[747439288,23],[306975095,20],[100686611,27],[107044363,25],[943165002,21],[53471920,21],[308689473,21]],"-368467311":[[359901870,65],[905403767,76],[502420536,68],[865228435,77],[632158872,80],[862301171,74],[518441852,69],[220584265,74],[132060072,80],[299800223,74],[888271666,74],[720733022,75],[468175329,74],[751363209,81],[317073342,72],[135483698,70],[489561862,70],[427782220,57],[303148031,88],[82206875,73],[638510972,80],[852011097,66],[678918016,71],[705001359,85],[300817061,73],[905166772,66],[655893407,71],[437225885,77],[200655,78],[398431610,74],[875445001,77],[810162171,75],[540447592,76],[925856065,72],[501322122,72],[77741400,77],[636531850,89],[185857779,67],[747978518,68],[332842197,73],[463068652,81],[795242014,77],[273144880,73],[654856205,77],[151533841,72],[238290185,73],[70085781,77],[110716833,74],[147734992,71],[889054306,80],[256996757,80],[371940876,71],[296549854,70],[629792455,68],[63167091,64],[602508982,79],[35097670,77],[182095059,71],[178321109,78],[53471929,83],[220521220,67],[358383371,80],[733611714,68],[649802605,73],[547557756,69],[78922923,71],[600407088,71],[745672598,72],[417715598,72],[994869408,64],[606055969,74],[212034860,81],[457545404,67],[864847383,84],[705621743,69],[439813542,69],[882222594,83],[905795572,75],[110440100,71],[506919024,65],[100382286,66],[480485746,72],[466300201,80],[19396790,69],[114477018,78],[559416718,64],[850341312,90],[346578094,84],[628398000,77],[14762896,65],[418246970,66],[837389446,77],[944832603,76],[418315672,79],[123946461,68],[245896452,74],[239523789,78],[759204305,76],[476558764,80],[233978704,69]],"-305057966":[[453224256,15],[955087335,14],[724765126,15],[365728647,15],[648432742,16],[518833950,41],[109397776,12],[271009115,13],[450828192,14],[325990380,39],[672259042,17],[149758807,15],[25107910,16],[242938045,15],[770130882,16],[392683637,13],[661667266,13],[943485692,14],[781910226,13],[93204457,12],[650929702,13],[394044940,14],[150341193,12],[531424253,12],[271359080,13],[974199904,15],[10696690,14],[719965461,11],[933714229,13],[285845329,13],[279225135,14],[619888190,17],[170861474,15],[375931783,12],[645371496,17],[84433533,16],[460957699,16],[23641296,13],[471725112,12],[120132822,14],[713759926,16],[282266967,14],[301338754,13],[802572892,14],[646633352,17],[112818580,16],[258733138,15],[204733234,31],[575228254,14],[808500641,15],[23564583,16],[338614635,14],[73102279,14],[932733523,17],[23480996,22],[292528353,13],[926322911,12],[521866946,12],[722714402,14],[43862269,15],[812207782,15],[189205207,12],[768394385,13],[255990106,16],[718392290,16],[795347056,21],[240856274,14],[849489360,15],[286900058,16],[314627884,16],[919670808,16],[769028148,13],[747439303,10],[904642802,10],[490553494,17],[107044370,12],[330592261,15],[182031076,15],[4120296,13],[533393009,15],[753021471,11],[727678801,13],[505140993,13],[876662048,16],[945650279,15],[765544668,15],[273748615,18],[826023737,14],[990955868,19],[935081767,12],[50396853,15],[202802572,16],[307896264,15],[889788744,15],[2094719,18],[101994218,15],[793542146,15],[380453845,11],[921460747,15],[578687632,11]],"-304007324":[[359901870,26],[502420530,29],[449626174,28],[60402820,31],[92967415,27],[353377831,29],[943221661,29],[930002488,30],[132060072,30],[555140747,28],[614435430,31],[432897287,27],[704725280,28],[703212758,31],[413999641,27],[986801717,28],[888271664,29],[489809540,30],[919745632,31],[874527832,33],[420587417,28],[269540871,29],[406991365,28],[720733021,30],[481485626,29],[864587313,30],[166244857,30],[431171753,29],[373866675,30],[801569480,31],[896510688,30],[66790091,30],[751363202,28],[849600547,28],[33882982,29],[530779519,28],[183084738,29],[135483697,28],[873347848,30],[706620351,29],[777122960,28],[489561869,27],[112790231,27],[456612469,28],[427782224,30],[121801048,30],[303148032,31],[979683980,29],[747439289,32],[100686613,29],[107044372,32],[889662355,31],[273090962,30],[893204360,29],[134218459,30],[212243106,31],[451329369,29],[312929054,31],[335318897,29],[188102580,27],[854642062,28],[954187309,31],[792047015,31],[560916602,29],[425320183,33],[33388730,30],[898605278,27],[562363758,31],[871361912,29],[178138191,31],[741606348,29],[552169014,29],[984978730,29],[823413288,29],[465448041,27],[196278088,28],[597423024,29],[765153127,30],[920395599,28],[702730281,29],[406926784,28],[922198415,31],[295485064,29],[883075817,31],[366549821,30],[843816921,31],[813216004,30],[38566165,27],[601740214,28],[677246682,29],[942614005,28],[595750012,30],[233830753,30],[511965085,30],[605715268,30],[561339114,29],[225593954,31],[342864795,31],[215478718,28],[968063498,30]],"-290327439":[[517311750,16],[779553368,19],[359901870,17],[905403769,20],[899221565,17],[502420528,18],[865228432,20],[449626174,15],[632158868,18],[727952816,20],[60402820,16],[92967415,19],[474064842,15],[353377836,16],[555354326,18],[518441848,16],[807000467,16],[943221662,15],[394416111,19],[148813978,15],[930002483,13],[957535220,20],[220584264,16],[132060070,24],[806602223,18],[217421780,20],[164924693,15],[555140745,22],[432897287,14],[167467368,17],[704725278,15],[812651449,20],[986801717,14],[888271664,20],[383963551,17],[489809536,15],[919745629,16],[874527827,16],[406991364,18],[55691246,14],[720733022,20],[367860448,17],[481485624,18],[618809583,17],[864587311,15],[852005524,20],[166244858,20],[431171753,15],[373866675,18],[644642173,17],[801569481,24],[17552661,21],[599427659,17],[556807265,14],[896510679,20],[328732304,15],[751363202,18],[320090596,13],[530779520,23],[317073340,20],[183084738,17],[785452373,17],[135483697,20],[873347847,19],[702193760,20],[706620350,16],[777122960,20],[718392276,16],[183090954,17],[112790229,15],[240856267,21],[849489344,18],[467510959,19],[427782222,19],[527271607,17],[747439288,17],[306975095,21],[100686610,16],[107044364,20],[943165003,16],[53471920,16],[384097525,16],[55183806,19],[3106686,18],[729874061,17],[752531116,21],[455768173,17],[59628687,21],[261889364,17],[913840687,16],[85425644,16],[449172184,20],[495173438,22],[687978810,18],[61006410,18],[880720950,17],[220935635,21],[909252369,18],[308305664,19],[250580174,20]],"-283360499":[[779553369,82],[60402821,59],[862301167,72],[943221661,113],[148813978,71],[930002484,79],[217421779,81],[555140746,81],[432897287,119],[167467370,99],[704725277,146],[812651450,82],[413999643,82],[888271666,79],[919745630,66],[269540869,73],[406991368,78],[55691247,99],[720733021,129],[367860448,216],[618809585,67],[864587310,75],[852005523,77],[431171755,82],[801569480,107],[17552661,98],[599427656,83],[789548691,100],[849600547,79],[320090600,66],[530779519,73],[135483699,78],[873347849,72],[706620350,82],[777122961,62],[718392277,77],[112790233,62],[467510964,94],[331145315,66],[303148030,76],[221906720,77],[527271606,107],[306975095,68],[100686610,71],[107044363,116],[943165010,71],[79529666,77],[654606296,64],[891772306,65],[827675735,61],[137336584,68],[580395755,75],[781643285,75],[989636855,88],[190514162,80],[605025171,69],[794758317,79],[982812233,76],[216156289,70],[666781851,80],[831707963,75],[647935775,73],[245026874,71],[668402829,65],[142468983,70],[990235523,82],[601740212,73],[497127724,77],[182368008,75],[690145026,81],[120404371,69],[727236248,69],[469774520,74],[997941875,74],[943546922,75],[558476733,75],[517295414,79],[395865046,67],[87137650,78],[251496439,73],[609298281,82],[356035220,89],[948226112,69],[298016968,71],[82175825,80],[266554252,76],[104287435,71],[850004175,56],[819644570,75],[324643316,66],[326653501,68],[378028448,69],[15555479,71],[696060044,73],[710454047,66],[300583733,62],[197712968,76],[156758067,74],[503006523,59],[380084399,72]],"-205274512":[[517311749,23],[779553368,23],[359901870,19],[905403767,21],[502420527,22],[727952816,21],[92967416,24],[474064842,24],[555354324,22],[518441848,19],[943221663,20],[148813978,20],[930002485,17],[220584266,21],[132060072,23],[806602223,23],[164924696,22],[432897287,20],[704725278,19],[812651449,19],[413999642,20],[986801717,20],[888271664,20],[383963550,24],[169253813,20],[489809536,21],[919745627,19],[420587417,25],[269540869,27],[55691247,22],[367860448,27],[864587310,28],[373866677,24],[644642173,22],[17552661,22],[599427656,28],[468175326,19],[556807266,20],[328732307,25],[751363202,24],[320090595,15],[33882977,19],[530779519,27],[183084738,23],[41191705,25],[135483698,26],[702193759,29],[706620350,23],[777122960,26],[489561863,21],[183090953,24],[112790230,24],[467510961,24],[331145316,24],[303148030,24],[221906714,20],[527271606,17],[979683981,24],[747439287,19],[306975095,21],[107044363,21],[943165003,22],[53471922,22],[974242447,30],[850341309,20],[434025512,21],[84861351,24],[36237590,22],[441295468,21],[774587610,18],[564255500,24],[400034001,21],[706569303,20],[825684135,26],[792191690,28],[585770587,23],[219504509,19],[688199414,22],[4120294,21],[785089222,20],[190908286,21],[358791853,25],[896480501,22],[66554216,19],[143217030,21],[969194251,26],[956446592,26],[973442670,23],[189205179,23],[895165214,21],[775040399,19],[130112665,20],[493949705,18],[134111762,22],[74123423,21],[358383360,27],[656672355,20],[127921611,22],[495883682,19],[901271210,20]],"-180984719":[[779553373,14],[359901872,15],[865228431,15],[632158870,13],[92967426,12],[862301166,18],[555354328,17],[518441850,19],[807000468,17],[943221663,18],[957535223,17],[220584267,19],[132060087,18],[806602224,19],[164924695,13],[614435433,13],[432897288,21],[209086736,16],[812651455,18],[413999641,19],[986801717,17],[169253827,17],[489809537,16],[919745638,14],[828988213,13],[420587426,14],[55691259,15],[720733027,18],[367860454,14],[481485623,12],[618809585,18],[852005523,15],[166244859,13],[373866675,14],[644642174,20],[468175326,13],[66790087,18],[789548697,15],[328732316,18],[751363202,15],[320090597,15],[317073345,14],[41191704,13],[785452373,15],[873347853,15],[706620356,13],[183090953,18],[112790230,16],[240856268,19],[467510961,15],[427782226,17],[303148031,21],[221906720,17],[527271622,14],[979683981,20],[747439287,14],[100686614,19],[107044365,14],[53471922,15],[639266369,23],[988375255,16],[563826826,15],[657292531,18],[617970171,19],[190135184,21],[811846150,16],[789551541,14],[968543135,15],[444307129,16],[28561535,17],[272342069,18],[582633681,21],[608271396,16],[169205811,15],[673208557,11],[148638837,16],[414546222,15],[114957032,17],[775391708,14],[593922932,16],[646998711,15],[571896601,15],[423512818,18],[330396854,18],[667558086,14],[790821375,17],[600131790,18],[119992686,20],[821871310,17],[793542146,15],[540081046,25],[130283843,18],[867465286,12],[532220527,18],[175601169,14],[113832908,15],[703676500,18],[453739275,15],[661183080,20],[38566170,17]],"-154035910":[[517311749,42],[779553368,45],[359901870,36],[905403767,44],[899221565,53],[502420527,36],[865228431,35],[449626174,50],[632158866,42],[727952816,35],[60402820,41],[92967414,41],[474064842,38],[862301166,39],[353377831,38],[555354324,34],[518441848,39],[807000467,38],[943221661,51],[394416111,48],[148813978,38],[930002484,52],[957535220,40],[220584264,38],[132060070,48],[806602223,48],[217421779,47],[164924693,26],[555140744,53],[614435426,51],[432897287,39],[167467368,47],[704725277,54],[812651449,37],[413999641,35],[986801716,46],[888271664,51],[383963550,46],[169253812,41],[489809536,46],[919745627,45],[828988207,44],[874527827,35],[420587416,49],[269540869,60],[406991364,37],[55691246,37],[720733021,31],[367860448,34],[481485622,47],[618809582,44],[864587310,43],[852005523,46],[166244857,38],[431171752,31],[373866675,48],[644642173,39],[801569480,54],[17552661,33],[599427656,49],[468175326,48],[556807265,41],[896510679,48],[66790086,39],[789548689,40],[328732304,44],[751363202,39],[849600546,49],[320090595,47],[33882977,37],[530779519,51],[317073340,44],[183084738,46],[41191704,36],[785452374,49],[135483697,49],[873347847,54],[702193759,41],[706620349,49],[777122960,52],[718392276,35],[489561862,50],[183090953,42],[112790229,40],[240856267,43],[849489344,45],[456612460,32],[467510959,38],[427782219,38],[331145315,42],[303148030,38],[221906714,33],[527271606,59],[979683980,48],[747439287,42],[306975095,41],[100686610,44],[107044363,42],[943165002,40],[53471920,41]],"-45478100":[[517311749,15],[779553368,15],[359901871,18],[905403767,16],[899221565,15],[502420527,17],[865228431,18],[449626174,17],[632158866,16],[727952816,17],[60402820,19],[92967414,18],[474064842,15],[862301166,16],[353377832,17],[555354324,15],[518441848,15],[807000467,15],[943221662,20],[394416111,14],[148813978,18],[930002484,18],[957535220,16],[220584264,16],[132060070,16],[806602223,16],[217421779,16],[164924693,15],[555140744,20],[614435426,17],[432897287,18],[167467368,17],[704725277,18],[812651450,18],[413999641,17],[986801716,17],[888271664,15],[383963550,16],[169253812,18],[489809536,16],[919745627,14],[828988207,19],[874527827,16],[420587416,17],[269540869,18],[406991364,18],[55691246,14],[720733021,15],[367860448,20],[481485622,17],[618809583,19],[864587310,16],[852005523,18],[166244857,16],[431171752,16],[373866676,15],[644642173,18],[17552662,15],[599427656,18],[468175327,13],[556807265,14],[896510679,18],[66790086,16],[789548689,16],[328732304,17],[751363202,18],[849600546,14],[320090595,16],[33882977,17],[530779519,17],[317073340,16],[183084738,18],[41191704,16],[785452373,16],[135483697,17],[873347847,17],[702193759,15],[706620349,17],[777122960,17],[718392276,16],[489561862,19],[183090955,17],[112790229,21],[240856267,19],[849489344,18],[456612460,17],[467510959,18],[427782219,18],[331145316,15],[303148031,17],[221906714,18],[527271606,19],[979683981,17],[747439287,18],[306975095,14],[100686610,15],[107044363,17],[943165002,18],[53471920,17],[576217591,18]],"-32621679":[[494195628,8],[246460598,10],[359901872,12],[905403772,9],[899221568,9],[502420554,13],[865228442,9],[449626175,11],[727952819,14],[60402830,9],[862301185,10],[392912493,8],[555354327,9],[242487882,9],[351041154,11],[394416112,14],[930002487,8],[957535228,10],[220584268,10],[557969583,11],[806602234,10],[217421789,12],[164924703,12],[62375277,12],[614435440,18],[432897306,10],[167467379,10],[704725284,10],[810302261,10],[369922194,10],[888271672,11],[383963564,11],[741614006,11],[489809540,10],[919745630,9],[789313998,13],[874527841,10],[420587420,9],[269540871,14],[896474844,12],[683783293,12],[864587323,11],[852005526,11],[451942199,13],[431171755,12],[28221385,10],[644642174,12],[822576243,12],[151999867,15],[468175334,13],[556807265,12],[815709802,8],[595167594,11],[936938793,8],[427611537,10],[812357752,13],[811011340,11],[386672421,9],[1886401,12],[885809648,12],[128917440,10],[66993530,12],[501117463,11],[328744492,10],[861903938,10],[853992505,11],[850660488,13],[823034153,12],[57405412,12],[528079300,10],[778789331,11],[613309715,10],[619149958,12],[878916086,12],[635582690,13],[793267432,8],[311708490,12],[712735092,13],[11570438,11],[366922888,11],[139353059,12],[93770000,9],[225355950,14],[49498411,12],[553126451,13],[168457157,10],[502373365,11],[498654590,12],[387842129,12],[977983445,11],[693164859,10],[339865003,12],[86851021,12],[103481205,11],[475697415,12],[534638718,12],[306588526,11],[29574660,13],[95941797,14],[934887563,8]],"-31571037":[[517311755,26],[779553374,28],[359901871,27],[899221570,26],[502420527,27],[449626176,30],[632158869,24],[727952817,29],[60402821,25],[92967415,29],[474064842,28],[862301167,28],[353377831,25],[807000469,29],[394416113,26],[930002483,28],[957535225,28],[220584265,26],[806602223,27],[555140753,30],[614435428,27],[432897287,26],[167467373,27],[413999641,28],[888271667,26],[169253812,28],[919745629,29],[269540869,28],[406991364,27],[55691246,27],[367860448,25],[618809584,26],[852005525,26],[166244857,26],[431171753,27],[644642176,26],[599427658,30],[468175326,28],[896510679,26],[789548691,28],[328732305,26],[751363203,27],[849600546,29],[320090595,27],[317073341,30],[183084738,28],[785452374,26],[135483698,27],[873347847,27],[702193761,28],[706620349,29],[777122961,28],[718392278,29],[240856272,26],[456612465,25],[427782220,25],[303148030,28],[221906715,25],[747439289,28],[306975100,26],[53471921,30],[643388670,27],[449172186,28],[95432311,29],[743781982,29],[214212076,29],[917077155,23],[921731777,26],[807396776,26],[832044790,27],[959922774,26],[195434137,26],[817774877,30],[465215830,25],[724184585,29],[987252040,28],[748395571,28],[128051304,27],[713560156,28],[995418691,28],[298206261,26],[500104787,29],[327505702,28],[804050144,27],[210489351,27],[249622423,25],[605242028,29],[916847144,27],[489615150,27],[389519834,26],[25910993,26],[936327594,27],[663663556,27],[948880614,27],[124585124,26],[735132057,27],[267482548,29],[497287981,28],[260611082,26],[300224539,29]],"-23063370":[[905403769,53],[862301169,49],[353377832,51],[555354325,49],[394416111,58],[930002502,47],[704725281,52],[812651450,52],[383963553,50],[153460486,53],[618809593,52],[17552663,52],[507734877,49],[66790086,55],[328732306,55],[849600553,55],[530779519,52],[890488300,47],[785452373,55],[873347860,52],[706620353,51],[718392291,51],[36424329,47],[427782225,50],[527271623,49],[100686614,50],[53471925,53],[860295484,55],[903894887,52],[997751146,52],[835143260,52],[326322248,55],[918809263,55],[58946752,50],[829444108,51],[223377174,55],[703993395,52],[144871377,51],[112948847,53],[82193471,57],[924867734,45],[134304507,52],[394821700,53],[582914996,53],[150677643,48],[451190366,50],[145840982,55],[958007742,56],[59628712,51],[309601169,58],[443146433,53],[28904102,48],[877341465,50],[959791984,51],[131166824,49],[617070360,52],[668932993,54],[632932307,53],[787471914,53],[946160961,53],[567077424,53],[248818411,49],[352213066,53],[169308269,54],[943084913,52],[633833584,52],[365477293,54],[803515006,54],[457468470,55],[607805870,53],[691841471,54],[685846666,51],[675638116,53],[709090769,49],[632033195,52],[928740691,53],[652360139,51],[178138198,50],[165535619,54],[621059500,53],[410262497,50],[64014298,52],[101821984,51],[5659703,55],[608620694,51],[773105671,49],[267401874,51],[909148600,53],[630315743,56],[683593520,52],[825707296,50],[815371929,51],[43299718,54],[862078660,52],[659945430,51],[89629071,52],[768070323,53],[992029529,53],[95011306,55],[924718919,52]],"-5036954":[[102552640,72],[449626180,65],[474064859,76],[587896349,65],[627853245,67],[164924698,75],[895395975,81],[704725277,59],[812651453,80],[413999653,88],[369922207,72],[294736199,80],[618809582,81],[254652653,69],[166244866,79],[101569644,74],[644642177,72],[708835267,84],[556807278,66],[789548692,66],[849600559,83],[702193765,69],[777122970,81],[795347063,68],[471116735,75],[979683991,67],[8254389,74],[272913912,74],[362525292,66],[162393345,67],[172807403,81],[558476736,64],[571494482,65],[872674388,76],[384863439,88],[413509219,72],[362523474,74],[372434763,71],[627821545,72],[316188689,73],[7393117,76],[4120301,72],[731602805,72],[365768906,75],[907537316,67],[950423984,61],[930543114,85],[507628918,75],[864579496,72],[274484786,72],[710346282,69],[864116160,69],[157864893,86],[497026938,80],[953038604,71],[631524038,74],[743154892,67],[984207334,75],[274856151,64],[394085636,67],[543509435,78],[413544571,66],[518970594,69],[27015747,73],[5233153,80],[502758809,73],[789290589,72],[51557417,89],[253125449,77],[287892499,71],[340577086,69],[412868573,69],[599839530,69],[776990692,71],[459046024,72],[841499588,77],[40388534,78],[822649951,67],[850947593,74],[949176014,68],[671864448,70],[225703205,66],[201424649,75],[747569614,71],[43862275,60],[100881728,81],[72860646,66],[630118733,83],[445803719,76],[844759486,73],[306932945,80],[744347768,72],[105820459,77],[203620343,90],[21745885,80],[632991155,78],[464988566,73],[274445132,77],[502143291,72],[359240758,81]]}'
 );
